### 2022-04-20 #4329 @editorialbot
[REVIEW]: PySD: System Dynamics Modeling in Python
**Submitting author:** @rogersamso (<a href="http://orcid.org/0000-0003-0348-3047">Roger Samsó</a>)
**Repository:** https://github.com/SDXorg/pysd
**Branch with paper.md** (empty if default branch): paper_pysd
**Version:** v3.7.1
