+++ pysd/building/python/namespace.py
@@ -1,6 +1,7 @@
 import re
 
 from unicodedata import normalize
+from typing import List
 
 # used to create python safe names with the variable reserved_words
 from keyword import kwlist
@@ -15,21 +16,50 @@ from pysd.py_backend.utils import __dir__ as udir
 
 
 class NamespaceManager:
+    """
+    NamespaceManager object allows includying new elements to the namespace
+    and searching for elements in the namespace. When includying new
+    elements a python safe name is used to be able to write the equations.
+
+    Parameters
+    ----------
+    parameters: list (optional)
+        List of the parameters that are used as argument in the Macro.
+        By defaukt it is an empty list.
+
+    """
     reserved_words = set(
         dir() + bidir() + cdir() + ddir() + cadir() + edir() + fdir()
         + sdir() + udir()).union(kwlist)
 
-    def __init__(self, parameters=[]):
+    def __init__(self, parameters: List[str] = []):
         self.used_words = self.reserved_words.copy()
+        # inlcude time to the namespace
         self.namespace = {"Time": "time"}
+        # include time to the cleanspace (case and whitespace/underscore
+        # insensitive namespace)
         self.cleanspace = {"time": "time"}
         for parameter in parameters:
             self.add_to_namespace(parameter)
 
-    def add_to_namespace(self, string):
+    def add_to_namespace(self, string: str) -> None:
+        """
+        Add a new string to the namespace.
+
+        Parameters
+        ----------
+        string: str
+            String to add to the namespace.
+
+        Returns
+        -------
+        None
+
+        """
         self.make_python_identifier(string, add_to_namespace=True)
 
-    def make_python_identifier(self, string, prefix=None, add_to_namespace=False):
+    def make_python_identifier(self, string: str, prefix: str = None,
+                               add_to_namespace: bool = False) -> str:
         """
         Takes an arbitrary string and creates a valid Python identifier.
 
@@ -48,12 +78,13 @@ class NamespaceManager:
         string: str
             The text to be converted into a valid python identifier.
 
-        namespace: dict
-            Map of existing translations into python safe identifiers.
-            This is to ensure that two strings are not translated into
-            the same python identifier. If string is already in the namespace
-            its value will be returned. Otherwise, namespace will be mutated
-            adding string as a new key and its value.
+        prefix: str or None (optional)
+            If given it will be used as a prefix for the output string.
+            Default is None.
+
+        add_to_namespace: bool (optional)
+            If True it will add the passed string to the namespace and
+            to the cleanspace. Default is False.
 
         Returns
         -------
@@ -88,15 +119,15 @@ class NamespaceManager:
         'nvs_123abc'
 
         already in namespace
-        >>> make_python_identifier('Var$', namespace={'Var$': 'var'})
+        >>> make_python_identifier('Var$')  # namespace={'Var$': 'var'}
         'var'
 
         namespace conflicts
-        >>> make_python_identifier('Var@', namespace={'Var$': 'var'})
+        >>> make_python_identifier('Var@')  # namespace={'Var$': 'var'}
         'var_1'
 
-        >>> make_python_identifier('Var$', namespace={'Var@': 'var',
-        ...                                           'Var%':'var_1'})
+        >>> make_python_identifier('Var$')  # namespace={'Var@': 'var',
+        ...                                              'Var%':'var_1'}
         'var_2'
 
         References
@@ -139,9 +170,11 @@ class NamespaceManager:
             identifier = s + '_' + str(i)
             i += 1
 
+        # include the word in used words to avoid using it againg
         self.used_words.add(identifier)
 
         if add_to_namespace:
+            # include word to the namespace
             self.namespace[string] = identifier
             self.cleanspace[clean_s] = identifier
 
