### 2022-04-22 7011710
Improve subscripts simplification
