### 2022-04-22 521640c
Improve more the loc in multiline definitions
