### 2022-04-24 SDXorg/pysd#316 @blsqr
[JOSS Review] Questions and suggestions for the Software Paper
Hi @rogersamso, @enekomartinmartinez, and co-authors 👋

I'm one of the reviewers of your [submission to JOSS](https://github.com/openjournals/joss-reviews/issues/4329) and have a few questions and suggestions regarding the [software paper]((https://github.com/JamesPHoughton/pysd/blob/paper_pysd/paper/paper.md)) (currently on branch [`paper_pysd`](https://github.com/JamesPHoughton/pysd/tree/paper_pysd), PDF version [here](https://github.com/openjournals/joss-papers/blob/7d8b7d8dff5ad9053b8e6b27f23ea6b33b3f62ac/joss.04329/10.21105.joss.04329.pdf)).

First of all: The paper is well written and structured and I find it easy to understand. While I would probably not be considered a “non-specialist audience”, I think that readers from other backgrounds will be able to understand the key motivation and ideas of PySD and SD modelling. 👍

I have some questions and suggestions regarding the text and the figures. I thought about opening a PR for these but I think we’ll have a quicker turnaround if I just list them here and you decide how to proceed:

1. **Typos:**
   - Line 27: *Stella’s or Vensim’s domain specific languages* (*both* need to end with an `'s` and language should then be plural)
   - Line 60: *causal* (not casual) loop diagrams
   - Line 62: *thoroughly*
2. **Superfluous abbreviations** (not used again, would suggest to drop them):
   - Line 10: MIT
   - Line 24: GUI
3. **Figure 1:** Is there a higher-quality figure available, ideally in a vector-based format? I would deem this one **not** fit for publication.
4. **Line 24:** Would it be worth highlighting that Stella and Vensim are propriertary and (afais) commercial tools? It’s done in the statement of need, but might be more helpful early on in the text.
5. **Line 29:** Perhaps rephrase to *“… of the extensive data-science capabilities of the Python ecosystem.”*?
6. **Line 31:** *“has 26 public releases, 2.2.4 being the most recent”*: I would drop that line, as it becomes outdated very quickly and provides little additional value in the context of the software paper – the linked repository is a much better indicator for development speed and contribution history.
7. **Figure 2:** Any chance to include the term *“Parsing Expression Grammar”* into the figure or caption to make the PEG abbreviation easier to understand? While you mention the abbreviation in the text, readers that first look at the figures might have an easier time understanding the figure that way. (Not a particularly important remark.)
8. **Figure 2 subcaption:** I would suggest to drop or rephrase the *“NOTE: …”* clause. To me, that information feels unnecessary at that point.

And finally, a question regarding the last paragraph before the Statement of Need, where you describe the future development roadmap of PySD:  Together with Figure 2 that paragraph represents a large chunk of the paper that does not refer to the submitted version of PySD but to a future one. Are these plans certain enough that you want to include them into the software paper? How far in the future is v3.0?
The roadmap is definitely interesting and gives valuable insight into the current and proposed future workings of the parsing and model building logic. To be clear: I am not proposing to drop or shorten that paragraph, but would primarly want to stress that whatever is written there should be reasonably fixed so that it does not become outdated too easily.

So far from my side. I hope that these comments are helpful – please let me know if there are further questions regarding the above or if there is anything else I can assist with :)
