### 2022-04-24 #4329 @blsqr
I have opened a first issue regarding the software paper [over in the PySD project](https://github.com/JamesPHoughton/pysd/issues/316).

FYI: The next two weeks I will not be available for review or questions, but will pick up the review some time in the middle of May.
