+++ paper/paper.md
@@ -32,13 +32,13 @@ bibliography: paper.bib
 
 ## Summary
 
-System Dynamics (SD) is a mathematical approach used to describe and simulate the dynamics of complex systems over time. The foundations of the methodology were laid in the 1950s by Professor Jay W. Forrester of the Massachusetts Institute of Technology (MIT) [@Forrester:1971]. The building blocks of SD models are stocks, flows, variables, parameters, and lookup tables. Stocks represent cumulative quantities which take a certain value at each moment in time (integral), and flows are the rates at which those quantities change per unit of time (derivative). Variables express intermediate calculations, parameters set external conditions for the simulation, and lookup tables are single-valued functions that accept another model component as an argument. These components can combine to create feedback loops in which the state of the stock variables feeds back to influence flows in the model. The relationships between these model components can be represented using causal loop diagrams (Figure \ref{fig:CLD}).
+System Dynamics (SD) is a mathematical approach used to describe and simulate the dynamics of complex systems over time. The foundations of the methodology were laid in the 1950s by Professor Jay W. Forrester of the Massachusetts Institute of Technology [@Forrester:1971]. The building blocks of SD models are stocks, flows, variables, parameters, and lookup tables. Stocks represent cumulative quantities which take a certain value at each moment in time (integral), and flows are the rates at which those quantities change per unit of time (derivative). Variables express intermediate calculations, parameters set external conditions for the simulation, and lookup tables are single-valued functions that accept another model component as an argument. These components can combine to create feedback loops in which the state of the stock variables feeds back to influence flows in the model. The relationships between these model components can be represented using causal loop diagrams (Figure \ref{fig:CLD}).
 
 ![An example Causal Loop Diagram showing how the various components of system dynamics models are visualized by domain-specific modeling environments. \label{fig:CLD}](CLD_example.png)
 
-Since its inception, the SD methodology has been applied in different areas, including manufacturing, energy, climate, population, ecology and economics [@croads:2012;@Moallemi:2021;@FETENEADANE2019212;@harvey:2021]. In the 1990s, the popularity of the methodology gave rise to the development of several visual programming systems for building SD models. These domain-specific modeling environments were widely adopted by the SD community due to their convenient graphical user interfaces (GUIs). Stella&#174; and Vensim&#174; are two of the most popular system dynamics modeling tools, but many others exist (see @wiki:Comparison_of_system_dynamics_software;@sdopentools).
+Since its inception, the SD methodology has been applied in different areas, including manufacturing, energy, climate, population, ecology and economics [@croads:2012;@Moallemi:2021;@FETENEADANE2019212;@harvey:2021]. In the 1990s, the popularity of the methodology gave rise to the development of several visual programming systems for building SD models. These domain-specific modeling environments were widely adopted by the SD community due to their convenient graphical user interfaces. Stella&#174; and Vensim&#174;, though not open-source, are two of the most popular commercial system dynamics modeling tools, but many others exist (see @wiki:Comparison_of_system_dynamics_software;@sdopentools).
 
-PySD [@pysd:2014] is a Python library that transpiles models built in Stella&#174; or Vensim&#174;'s domain-specific language into Python, allowing the user to load, parametrize and execute SD models, and to take advantage of Python's extensive data-science capabilities. PySD was first released in September 2014 by James Houghton [@houghton_siegel_2015], and has 26 public releases to date, *v2.2.4* the most recent.
[…]
