### 2022-04-25 374195e
Improve subscript updimensioning and transposition
