+++ pysd/building/python/python_expressions_builder.py
@@ -19,9 +19,7 @@ class BuildAST:
         # makes easier building
         return self.expression
 
-    def reshape(self, subscripts, final_subscripts):
-        subscripts_out = subscripts.simplify_subscript_input(
-            final_subscripts)[1]
+    def reshape(self, subscripts, final_subscripts, final_element=False):
         if not final_subscripts or (
           self.subscripts == final_subscripts
           and list(self.subscripts) == list(final_subscripts)):
@@ -30,22 +28,42 @@ class BuildAST:
         elif not self.subscripts:
             # original expression is not an array
             # NUMPY: object.expression = np.full(%s, %(shape)s)
+            subscripts_out = subscripts.simplify_subscript_input(
+                final_subscripts)[1]
             self.expression = "xr.DataArray(%s, %s, %s)" % (
                 self.expression, subscripts_out, list(final_subscripts)
             )
             self.order = 0
+            self.subscripts = final_subscripts
         else:
             # original expression is an array
-            # NUMPY: reorder dims if neccessary with np.moveaxis or similar
-            # NUMPY: add new axis with [:, None, :] or np.tile,
-            #        depending on an input argument
-            # NUMPY: if order is not 0 need to lower the order to 0
-            # using force!
-            self.expression = "(xr.DataArray(0, %s, %s) + %s)" % (
-                subscripts_out, list(final_subscripts), self.expression
-                )
-            self.order = 0
-        self.subscripts = final_subscripts
+            self.lower_order(0, force_0=True)
+
+            # reorder subscrips
+            final_order = {
+                sub: self.subscripts[sub]
+                for sub in final_subscripts
+                if sub in self.subscripts
+            }
+            if list(final_order) != list(self.subscripts):
+                # NUMPY: reorder dims if neccessary with np.moveaxis or similar
+                self.expression +=\
+                    f".transpose({', '.join(map(repr, final_order))})"
+                self.subscripts = final_order
+
+            # add new dimensions
+            if final_element and final_subscripts != self.subscripts:
+                # NUMPY: remove final_element condition from top
+                # NUMPY: add new axis with [:, None, :]
+                # NUMPY: move final_element condition here and use np.tile
+                for i, dim in enumerate(final_subscripts):
+                    if dim not in self.subscripts:
+                        subscripts_out = subscripts.simplify_subscript_input(
+                            {dim: final_subscripts[dim]})[1]
+                        self.expression +=\
+                            f".expand_dims({subscripts_out}, {i})"
+
+                self.subscripts = final_subscripts
 
     def lower_order(self, new_order, force_0=False):
         if self.order >= new_order and self.order != 0\
@@ -93,7 +111,8 @@ class StructureBuilder:
             final_subscripts = self.get_final_subscripts(
                 arguments, def_subs)
 
-        [arguments[key].reshape(self.section.subscripts, final_subscripts)
+        [arguments[key].reshape(
+            self.section.subscripts, final_subscripts, force == "equal")
          for key in arguments
          if arguments[key].subscripts or force == "equal"]
 
@@ -364,31 +383,37 @@ class CallBuilder(StructureBuilder):
                 def_subs=self.def_subs
             )
             if self.function == "xidz" and final_subscripts:
+                # xidz must always return the same shape object
                 if not arguments["1"].subscripts:
-                    new_args = {"0": arguments["0"], "2": arguments["2"]}
-                    self.reorder(
-                        new_args,
-                        def_subs=self.def_subs,
-                        force="equal"
-                    )
-                    arguments.update(new_args)
-            if self.function == "if_then_else" and final_subscripts:
[…]
