### 2022-04-25 JamesPHoughton/pysd#316 @rogersamso
Hi @blsqr,
Thank you for taking the time to review the paper, and for the valuable comments. We accepted all your suggestions and they have already been merged into the paper branch. We also made a new version of the first figure in svg format, and added it to the paper.

Regarding your last question, we are working on the v3.0 release in #312. The separation between parsing and building is already complete (all pre-existing test models work again) but we still need to work on the documentation and a few other minor things. We will soon start working on the migration from xarray to numpy.  So we are confident that v3.0 will be released before the summer break.

Thanks again for the useful feedback and please let me know if there is anything else we could do to further improve the document.
