+++ paper/paper.md
@@ -34,7 +34,7 @@ bibliography: paper.bib
 
 System Dynamics (SD) is a mathematical approach used to describe and simulate the dynamics of complex systems over time. The foundations of the methodology were laid in the 1950s by Professor Jay W. Forrester of the Massachusetts Institute of Technology [@Forrester:1971]. The building blocks of SD models are stocks, flows, variables, parameters, and lookup tables. Stocks represent cumulative quantities which take a certain value at each moment in time (integral), and flows are the rates at which those quantities change per unit of time (derivative). Variables express intermediate calculations, parameters set external conditions for the simulation, and lookup tables are single-valued functions that accept another model component as an argument. These components can combine to create feedback loops in which the state of the stock variables feeds back to influence flows in the model. The relationships between these model components can be represented using causal loop diagrams (Figure \ref{fig:CLD}).
 
-![An example Causal Loop Diagram showing how the various components of system dynamics models are visualized by domain-specific modeling environments. \label{fig:CLD}](CLD_example.png)
+![An example Causal Loop Diagram showing how the various components of system dynamics models are visualized by domain-specific modeling environments. \label{fig:CLD}](CLD_example.pdf)
 
 Since its inception, the SD methodology has been applied in different areas, including manufacturing, energy, climate, population, ecology and economics [@croads:2012;@Moallemi:2021;@FETENEADANE2019212;@harvey:2021]. In the 1990s, the popularity of the methodology gave rise to the development of several visual programming systems for building SD models. These domain-specific modeling environments were widely adopted by the SD community due to their convenient graphical user interfaces. Stella&#174; and Vensim&#174;, though not open-source, are two of the most popular commercial system dynamics modeling tools, but many others exist (see @wiki:Comparison_of_system_dynamics_software;@sdopentools).
 
@@ -46,7 +46,7 @@ In addition to the aforementioned core functionality, PySD also allows users to 
 
 Despite its maturity, PySD is currently in a very active development phase, and a proof to that is that most of the extra features listed in the paragraph above were implemented in the 14 months that separate releases *v0.11.0* and *v2.2.1*. The roadmap for the next major release (3.0) will focus on cutting down simulation times and on including additional built-in Stella&#174; and Vensim&#174; functions. The most relevant performance-oriented development will be the migration from xarray to numpy to perform array operations, which in addition to the expected overhead reduction will also allow for further potential model optimizations (i.e. JIT compilation). This migration will be facilitated by an even larger development scheduled for the 3.0 release, which corresponds to isolating the parsing process from the building process. This will be achieved by creating an intermediate or abstract model representation (AMR)(Figure \ref{fig:ABSTRACTMODEL}) that will embed all the information contained in the original models, plus additional information such as the order of operations, in pure Python objects. This intermediate AMR will open the door to the development of additional model builders (in Python) to write the output models in any programming language. Furthermore, the intermediate abstract model may encourage the development of new tools to generate and analyze the graphs associated with the models.
 
-![Comparison of the current and upcoming model parsing-building logic in PySD. PEG acronym stands for Parsing Expression Grammar. \label{fig:ABSTRACTMODEL}](abstract_model.png)
+![Comparison of the current and upcoming model parsing-building logic in PySD. PEG acronym stands for Parsing Expression Grammar. \label{fig:ABSTRACTMODEL}](abstract_model.pdf)
 
 ## Statement of need
 
