### 2022-04-26 JamesPHoughton/pysd#316 @rogersamso
Hi again @blsqr,
My bad. The two figures are now at higher res and in pdf.

> In that case I'm wondering: Have you considered postponing the publication of the JOSS paper until that point? It seems like a major structural change that is not too far in the future. Publishing at a later point would allow to focus on the new structure of PySD rather than describing the old one, making the publication more concise.

That is a very good point, indeed.

As we mention in the paper, PySD has been mature and functional for some time now (since 1.0). The basic problem PySD solves has not changed either, and will remain the same in further releases. That is what we aim to present in the paper.

However, we keep looking for ways to make the code more maintainable, to make simulations run faster, and to improve the user experience. Though these are not essential developments in most cases (unless performance is a must), the users (including ourselves) always welcome them.

That is what release 3.0 is all about: the average user may see a notable increase in simulation performance, but not much else. However, from a developers perspective, 3.0 will bring non backwards compatible changes which will cause inconveniences to the libraries that rely on PySD. Therefore, the objective of presenting it before it is released, is to catch the attention of developers (who may or may not be willing to write builders for other programming languages), and to give them the possibility to review and future-proof what we propose for the 3.0 release. As I mentioned in my previous comment, the PR may be ready before summer, but if we manage to onboard new contributors, we would be more than willing to delay the 3.0 release a few extra months.

Of course this is not a final decision, and we will adapt to what reviewers think is best.
