### 2022-04-26 JamesPHoughton/pysd#316 @blsqr
Thank you for the detailed response, @rogersamso.

> Therefore, the objective of presenting it before it is released, is to catch the attention of developers (who may or may not be willing to write builders for other programming languages), and to give them the possibility to review and future-proof what we propose for the 3.0 release.

This is a very good idea. I see more benefits in that approach compared to a postponed publication that would focus solely on v3.0. 👍
