### 2022-04-26 a32a120
Remove dependencies dict from model files
Remove dependencies dict from model files and also remove subscript dict when it is not necessary.
