+++ pysd/building/python/python_expressions_builder.py
@@ -292,11 +292,12 @@ class CallBuilder(StructureBuilder):
                           "%(args)s, '%(macro_name)s', "
                           "time_initialization=lambda: __data['time'], "
                           "py_name='%(name)s')" % arguments,
-            "calls": {
-                "initial": calls,
-                "step": calls
-            }
         }
+        self.element.other_dependencies[arguments["name"]] = {
+            "initial": calls,
+            "step": calls
+        }
+
         return BuildAST(
             expression="%s()" % arguments["name"],
             calls={arguments["name"]: 1},
@@ -367,15 +368,11 @@ class CallBuilder(StructureBuilder):
                 def_subs=self.def_subs,
                 force="equal"
             )
-            self.element.objects[name] = {
-                "name": name,
-                "expression": None,
-                "calls": {
-                    "initial": arguments["1"].calls,
-                    "step": arguments["0"].calls
-                }
-
+            self.element.other_dependencies[name] = {
+                "initial": arguments["1"].calls,
+                "step": arguments["0"].calls
             }
+
             calls = {name: 1}
         else:
             final_subscripts = self.reorder(
@@ -658,12 +655,12 @@ class InitialBuilder(StructureBuilder):
             "name": arguments["name"],
             "expression": "%(name)s = Initial(lambda: %(initial)s, "
                           "'%(name)s')" % arguments,
-            "calls": {
-                "initial": arguments["initial"].calls,
-                "step": {}
-            }
-
         }
+        self.element.other_dependencies[arguments["name"]] = {
+            "initial": arguments["initial"].calls,
+            "step": {}
+        }
+
         return BuildAST(
             expression=arguments["name"] + "()",
             calls={arguments["name"]: 1},
@@ -695,13 +692,13 @@ class IntegBuilder(StructureBuilder):
         self.element.objects[arguments["name"]] = {
             "name": arguments["name"],
             "expression": "%(name)s = Integ(lambda: %(flow)s, "
-                          "lambda: %(initial)s, '%(name)s')" % arguments,
-            "calls": {
-                "initial": arguments["initial"].calls,
-                "step": arguments["flow"].calls
-            }
-
+                          "lambda: %(initial)s, '%(name)s')" % arguments
         }
+        self.element.other_dependencies[arguments["name"]] = {
+            "initial": arguments["initial"].calls,
+            "step": arguments["flow"].calls
+        }
+
         return BuildAST(
             expression=arguments["name"] + "()",
             calls={arguments["name"]: 1},
@@ -742,17 +739,18 @@ class DelayBuilder(StructureBuilder):
                           "lambda: %(delay_time)s, lambda: %(initial)s, "
                           "lambda: %(order)s, "
                           "time_step, '%(name)s')" % arguments,
-            "calls": {
-                "initial": merge_dependencies(
-                    arguments["initial"].calls,
-                    arguments["delay_time"].calls,
-                    arguments["order"].calls),
-                "step": merge_dependencies(
-                    arguments["input"].calls,
-                    arguments["delay_time"].calls)
-
-            }
         }
+        self.element.other_dependencies[arguments["name"]] = {
+            "initial": merge_dependencies(
+                arguments["initial"].calls,
+                arguments["delay_time"].calls,
+                arguments["order"].calls),
+            "step": merge_dependencies(
+                arguments["input"].calls,
+                arguments["delay_time"].calls)
+
+        }
+
         return BuildAST(
             expression=arguments["name"] + "()",
             calls={arguments["name"]: 1},
@@ -787,13 +785,14 @@ class DelayFixedBuilder(StructureBuilder):
             "expression": "%(name)s = DelayFixed(lambda: %(input)s, "
                           "lambda: %(delay_time)s, lambda: %(initial)s, "
[…]
