### 2022-04-27 #4329 @enekomartinmartinez
Hi Andre,

The error is because of a missing feature. As the parser stops with a list
of subscript ranges I suposse that this will be part of an :EXCEPT:
declaration. If this is the case, this bug is already fixed in a branch and
will be solved in a future release. Please, could you share the full
equation where the string  '[mining],[petroleum' appears?

Thank you,
Eneko

Andre Cieplinski ***@***.***> igorleak hau idatzi zuen (2022
api. 27, az. 23:28):

> Hi @rogersamso <https://github.com/rogersamso>, I finally managed to use
> pysd and run some simple models. However, I am having some trouble
> converting a more complex model. I am not sure if I am trying to upload a
> model with features not supported in pysd or doing something wrong.
> Also I a not sure if this is where I should ask this, sorry, still trying
> to figure out Github.
>
> I get the following message:
> "model = pysd.read_vensim('model.mdl')
> Traceback (most recent call last):
>
> File
> "C:\Users\UTENTE\anaconda3\lib\site-packages\pysd\translation\vensim\vensim2py.py",
> line 470, in get_equation_components
> tree = parser.parse(equation_str)
>
> File
> "C:\Users\UTENTE\anaconda3\lib\site-packages\parsimonious\grammar.py", line
> 115, in parse
> return self.default_rule.parse(text, pos=pos)
>
> File
> "C:\Users\UTENTE\anaconda3\lib\site-packages\parsimonious\expressions.py",
> line 122, in parse
> raise IncompleteParseError(text, node.end, self)
>
> IncompleteParseError: Rule 'entry' matched in its entirety, but it didn't
> consume all the text. The non-matching portion of the text begins with '
> [mining],[petroleum' (line 1, column 36)."
>
> thanks a lot,
>
> andré
>
