### 2022-04-28 #4329 @rogersamso
Dear @andrecieplinski thanks for contacting. As @enekomartinmartinez mentions, this new feature will be added in the upcoming release 3.0 of pysd. Please let's not continue this discussion here. Either open a new issue in the pysd repo, or drop as an email. Thanks again.
