### 2022-04-28 #4329 @rogersamso
Dear @pdebuyl and @andrecieplinski , I just realized @andrecieplinski was assigned as reviewer for the paper. I see a potential conflict of interest, as @andrecieplinski is collaborating in the same research project than us (h2020 LOCOMOTION). please let us know if this may cause problems.
Regards
