### 2022-04-29 8ba76e5
created function get_namespace
