### 2022-05-04 8aece44
Put supported suffixes in one place
