### 2022-05-04 #4329 @pdebuyl
Hi all, we may proceed the review with @andrecieplinski , the association only by being member of a funding consortium -- provided there was no direct collaboration -- is not a sufficient conflict of interest for JOSS.
