+++ tests/more-tests/not_implemented_and_incomplete/test_not_implemented_and_incomplete.py
@@ -1,118 +1,0 @@
-"""
-Python model 'test_not_implemented_and_incomplete.py'
-Translated using PySD
-"""
-
-from pathlib import Path
-import numpy as np
-import xarray as xr
-
-from pysd.py_backend.functions import not_implemented_function, incomplete
-from pysd import Component
-
-__pysd_version__ = "3.0.0"
-
-__data = {"scope": None, "time": lambda: 0}
-
-_root = Path(__file__).parent
-
-
-component = Component()
-
-#######################################################################
-#                          CONTROL VARIABLES                          #
-#######################################################################
-
-_control_vars = {
-    "initial_time": lambda: 0,
-    "final_time": lambda: 1,
-    "time_step": lambda: 1,
-    "saveper": lambda: time_step(),
-}
-
-
-def _init_outer_references(data):
-    for key in data:
-        __data[key] = data[key]
-
-
-@component.add(name="Time")
-def time():
-    """
-    Current time of the model.
-    """
-    return __data["time"]()
-
-
-@component.add(
-    name="FINAL TIME", units="Month", comp_type="Constant", comp_subtype="Normal"
-)
-def final_time():
-    """
-    The final time for the simulation.
-    """
-    return __data["time"].final_time()
-
-
-@component.add(
-    name="INITIAL TIME", units="Month", comp_type="Constant", comp_subtype="Normal"
-)
-def initial_time():
-    """
-    The initial time for the simulation.
-    """
-    return __data["time"].initial_time()
-
-
-@component.add(
-    name="SAVEPER",
-    units="Month",
-    limits=(0.0, np.nan),
-    comp_type="Auxiliary",
-    comp_subtype="Normal",
-    depends_on={"time_step": 1},
-)
-def saveper():
-    """
-    The frequency with which output is stored.
-    """
-    return __data["time"].saveper()
-
-
-@component.add(
-    name="TIME STEP",
-    units="Month",
-    limits=(0.0, np.nan),
-    comp_type="Constant",
-    comp_subtype="Normal",
-)
-def time_step():
-    """
-    The time step for the simulation.
-    """
-    return __data["time"].time_step()
-
-
-#######################################################################
-#                           MODEL VARIABLES                           #
-#######################################################################
-
-
-@component.add(
-    name="incomplete var",
-    comp_type="Auxiliary",
-    comp_subtype="Normal",
-    depends_on={"time": 1},
-)
-def incomplete_var():
-    return incomplete(time())
-
-
-@component.add(
-    name="not implemented function",
-    comp_type="Auxiliary",
-    comp_subtype="Normal",
-    depends_on={"time": 1},
-)
-def not_implemented_function_1():
-    return not_implemented_function("my_func", time())
