### 2022-05-06 3f222c8
Create a AbstractSyntax class for typing
