+++ pysd/translation/structures/abstract_expressions.py
@@ -12,8 +12,16 @@ from dataclasses import dataclass
 from typing import Union
 
 
+class AbstractSyntax:
+    """
+    Generic class. All Abstract Synax structured are childs of that class.
+    Used for typing.
+    """
+    pass
+
+
 @dataclass
-class ArithmeticStructure:
+class ArithmeticStructure(AbstractSyntax):
     """
     Dataclass for an arithmetic structure.
 
@@ -34,7 +42,7 @@ class ArithmeticStructure:
 
 
 @dataclass
-class LogicStructure:
+class LogicStructure(AbstractSyntax):
     """
     Dataclass for a logic structure.
 
@@ -55,7 +63,7 @@ class LogicStructure:
 
 
 @dataclass
-class SubscriptsReferenceStructure:
+class SubscriptsReferenceStructure(AbstractSyntax):
     """
     Dataclass for a subscript reference structure.
 
@@ -72,7 +80,7 @@ class SubscriptsReferenceStructure:
 
 
 @dataclass
-class ReferenceStructure:
+class ReferenceStructure(AbstractSyntax):
     """
     Dataclass for an element reference structure.
 
@@ -94,7 +102,7 @@ class ReferenceStructure:
 
 
 @dataclass
-class CallStructure:
+class CallStructure(AbstractSyntax):
     """
     Dataclass for a call structure.
 
@@ -106,7 +114,7 @@ class CallStructure:
         The list of arguments used for calling the function.
 
     """
-    function: Union[str, object]
+    function: Union[str, ReferenceStructure]
     arguments: tuple
 
     def __str__(self) -> str:  # pragma: no cover
@@ -119,7 +127,7 @@ class CallStructure:
 
 
 @dataclass
-class GameStructure:
+class GameStructure(AbstractSyntax):
     """
     Dataclass for a game structure.
 
@@ -129,14 +137,14 @@ class GameStructure:
         The expression inside the game call.
 
     """
-    expression: object
+    expression: Union[AbstractSyntax, float]
 
     def __str__(self) -> str:  # pragma: no cover
         return "GameStructure:\n\t%s" % self.expression
 
 
 @dataclass
-class InitialStructure:
+class InitialStructure(AbstractSyntax):
     """
     Dataclass for a initial structure.
 
@@ -146,7 +154,7 @@ class InitialStructure:
         The expression inside the initial call.
 
     """
-    initial: object
+    initial: Union[AbstractSyntax, float]
 
     def __str__(self) -> str:  # pragma: no cover
         return "InitialStructure:\n\t%s" % (
@@ -154,7 +162,7 @@ class InitialStructure:
 
 
 @dataclass
-class IntegStructure:
+class IntegStructure(AbstractSyntax):
     """
     Dataclass for an integ/stock structure.
 
@@ -166,8 +174,8 @@ class IntegStructure:
         The initial value of the stock.
 
     """
-    flow: object
-    initial: object
+    flow: Union[AbstractSyntax, float]
+    initial: Union[AbstractSyntax, float]
 
     def __str__(self) -> str:  # pragma: no cover
         return "IntegStructure:\n\t%s,\n\t%s" % (
@@ -176,7 +184,7 @@ class IntegStructure:
 
 
 @dataclass
-class DelayStructure:
+class DelayStructure(AbstractSyntax):
     """
     Dataclass for a delay structure.
 
@@ -192,9 +200,9 @@ class DelayStructure:
         The order of the delay.
 
     """
-    input: object
-    delay_time: object
-    initial: object
+    input: Union[AbstractSyntax, float]
+    delay_time: Union[AbstractSyntax, float]
[…]
