### 2022-05-06 e6eb1e8
Add support for zeroled decimals
