### 2022-05-06 9632501
Remove unnecessary lines
