### 2022-05-09 50b6f80
Correct bug with return timestamps
