+++ pysd/py_backend/components.py
@@ -2,6 +2,7 @@
 Model components and time managing classes.
 """
 
+from warnings import warn
 import os
 import random
 import inspect
@@ -176,12 +177,30 @@ class Time(object):
 
     def in_return(self):
         """ Check if current time should be returned """
+        prec = self.time_step() * self.rprec
+
         if self.return_timestamps is not None:
-            return self._time in self.return_timestamps
+            # this allows managing float precission error
+            if self.next_return is None:
+                return False
+            if np.isclose(self._time, self.next_return, prec):
+                self._update_next_return()
+                return True
+            else:
+                while self.next_return is not None\
+                      and self._time > self.next_return:
+                    warn(
+                        f"The returning time stamp '{self.next_return}' "
+                        "seems to not be a multiple of the time step. "
+                        "This value will not be saved in the output. "
+                        "Please, modify the returning timestamps or the "
+                        "integration time step to avoid this."
+                        )
+                    self._update_next_return()
+                return False
 
         time_delay = self._time - self._initial_time
         save_per = self.saveper()
-        prec = self.time_step() * self.rprec
         return time_delay % save_per < prec or -time_delay % save_per < prec
 
     def round(self):
@@ -192,14 +211,28 @@ class Time(object):
 
     def add_return_timestamps(self, return_timestamps):
         """ Add return timestamps """
-        if return_timestamps is None or hasattr(return_timestamps, '__len__'):
-            self.return_timestamps = return_timestamps
+        if hasattr(return_timestamps, '__len__')\
+           and len(return_timestamps) > 0:
+            self.return_timestamps = list(return_timestamps)
+            self.return_timestamps.sort(reverse=True)
+            self.next_return = self.return_timestamps.pop()
+        elif isinstance(return_timestamps, (float, int)):
+            self.next_return = return_timestamps
+            self.return_timestamps = []
         else:
-            self.return_timestamps = [return_timestamps]
+            self.next_return = None
+            self.return_timestamps = None
 
     def update(self, value):
         """ Update current time value """
         self._time = value
+
+    def _update_next_return(self):
+        """ Update the next_return value """
+        if self.return_timestamps:
+            self.next_return = self.return_timestamps.pop()
+        else:
+            self.next_return = None
 
     def reset(self):
         """ Reset time value to the initial """
