### 2022-05-09 5bb06fe
Catch warnings in the integration tests
