+++ tests/pytest_integration/pytest_integration_vensim_pathway.py
@@ -1,5 +1,7 @@
-import pytest
+import warnings
 import shutil
+import pytest
+
 from pysd.tools.benchmarking import runner, assert_frames_close
 
 # TODO add warnings catcher per test
@@ -575,5 +577,7 @@ class TestIntegrateVensim:
         return kwargs
 
     def test_read_vensim_file(self, model_path, data_path, kwargs):
-        output, canon = runner(model_path, data_files=data_path)
+        with warnings.catch_warnings():
+            warnings.simplefilter("ignore")
+            output, canon = runner(model_path, data_files=data_path)
         assert_frames_close(output, canon, **kwargs)
