+++ pysd/building/python/python_model_builder.py
@@ -1,6 +1,15 @@
+"""
+The ModelBuilder class allows converting the AbstractModel into a
+PySD model writing the Python code in files that can be loaded later
+with PySD Model class. Each Abstract level has its own Builder. However,
+the user is only required to create a ModelBuilder object using the
+AbstractModel and call the `build_model` method.
+"""
 import textwrap
 import black
 import json
+from pathlib import Path
+from typing import Union
 
 from pysd.translation.structures.abstract_model import\
     AbstractComponent, AbstractElement, AbstractModel, AbstractSection
@@ -13,109 +22,155 @@ from pysd._version import __version__
 
 
 class ModelBuilder:
+    """
+    ModelBuilder allows building a PySD Python model from the
+    Abstract Model.
+
+    Parameters
+    ----------
+    abstract_model: AbstractModel
+        The abstract model to build.
+
+    """
 
     def __init__(self, abstract_model: AbstractModel):
         self.__dict__ = abstract_model.__dict__.copy()
+        # load sections
         self.sections = [
             SectionBuilder(section)
             for section in abstract_model.sections
         ]
+        # create the macrospace (namespace of macros)
         self.macrospace = {
             section.name: section for section in self.sections[1:]}
 
-    def build_model(self):
-        # TODO: add special building for main
+    def build_model(self) -> Path:
+        """
+        Build the python model in a file callled as the orginal model
+        but with '.py' suffix.
+
+        Returns
+        -------
+        path: pathlib.Path
+            The path to the new PySD model.
+
+        """
         for section in self.sections:
+            # add macrospace information to each section and build it
             section.macrospace = self.macrospace
             section.build_section()
 
+        # return the path to the main file
         return self.sections[0].path
 
 
 class SectionBuilder:
+    """
+    SectionBuilder allows building a section of the PySD model. Each
+    section will be a file unless the model has been setted to be
+    split in modules.
 
+    Parameters
+    ----------
+    abstract_section: AbstractSection
+        The abstract section to build.
+
+    """
     def __init__(self, abstract_section: AbstractSection):
         self.__dict__ = abstract_section.__dict__.copy()
-        self.root = self.path.parent
-        self.model_name = self.path.with_suffix("").name
+        self.root = self.path.parent  # the folder where the model is
+        self.model_name = self.path.with_suffix("").name  # name of the model
+        # Create subscript manager object with subscripts_dict
         self.subscripts = SubscriptManager(
             abstract_section.subscripts, self.root)
+        # Load the elements in the section
         self.elements = [
             ElementBuilder(element, self)
             for element in abstract_section.elements
         ]
+        # Create the namespace of the section
         self.namespace = NamespaceManager(self.params)
+        # Create an imports manager
         self.imports = ImportsManager()
+        # Create macrospace (namespace of macros)
         self.macrospace = {}
-
-        # create parameters dict necessary in macros
+        # Create parameters dict necessary in macros
         self.params = {
             key: self.namespace.namespace[key]
             for key in self.params
         }
 
-    def build_section(self):
-        # Create namespace
+    def build_section(self) -> None:
+        """
+        Build the python section in a file callled as the orginal model
+        if the section is main or in a file called as the macro name
+        if the section is a macro.
+        """
+        # Firts iteration over elements to recover their information
         for element in self.elements:
+            # Add element to namespace
             self.namespace.add_to_namespace(element.name)
             identifier = self.namespace.namespace[element.name]
             element.identifier = identifier
+            # Add element subscripts information to the subscript manager
             self.subscripts.elements[identifier] = element.subscripts
 
+        # Build elements
         for element in self.elements:
             element.build_element()
 
         if self.split:
+            # Build modular section
             self._build_modular(self.views_dict)
         else:
+            # Build one-file section
[…]
