+++ pysd/building/python/python_expressions_builder.py
@@ -1,3 +1,13 @@
+"""
+The translation from Abstract Syntax Tree to Python happens in both ways.
+The outer expression is visited with its builder, which will split its
+arguments and visit them with their respective builders. Once the lowest
+level is reached, it will be translated into Python returning a BuildAST
+object, this object will include the python expression, its subscripts,
+its calls to other and its arithmetic order (see Build AST for more info).
+BuildAST will be returned for each visited argument from the lower
+lever to the top level, giving the final expression.
+"""
 import warnings
 from dataclasses import dataclass
 from typing import Union
@@ -5,30 +15,79 @@ from typing import Union
 import numpy as np
 from pysd.py_backend.utils import compute_shape
 
-from pysd.translation.structures.abstract_expressions import AbstractSyntax
-from pysd.translation.structures import abstract_expressions as ae
+from pysd.translation.structures.abstract_expressions import\
+    AbstractSyntax, ArithmeticStructure, CallStructure, DataStructure,\
+    DelayFixedStructure, DelayStructure, DelayNStructure, ForecastStructure,\
+    GameStructure, GetConstantsStructure, GetDataStructure,\
+    GetLookupsStructure, InitialStructure, InlineLookupsStructure,\
+    IntegStructure, LogicStructure, LookupsStructure, ReferenceStructure,\
+    SampleIfTrueStructure, SmoothNStructure, SmoothStructure,\
+    SubscriptsReferenceStructure, TrendStructure
+
 from .python_functions import functionspace
+from .subscripts import SubscriptManager
 
 
 @dataclass
 class BuildAST:
+    """
+    Python expression holder.
+
+    Parameters
+    ----------
+    expression: str
+        The Python expression.
+    calls: dict
+        The calls to other variables for the dependencies dictionary.
+    subscripts: dict
+        The subscripts dict of the expression.
+    order: int
+        Arithmetic order of the expression. The arithmetic order depends
+        on the last arithmetic operation. If the expression is a number,
+        a call to a function, or is between parenthesis; its order will
+        be 0. If the expression its an exponential of two terms its order
+        will be 1. If the expression is a product or division its order
+        will be 2. If the expression is a sum or substraction its order
+        will be 3. If the expression is a logical comparison its order
+        will be 4.
+
+    """
     expression: str
     calls: dict
     subscripts: dict
     order: int
 
-    def __str__(self):
+    def __str__(self) -> str:
         # makes easier building
         return self.expression
 
-    def reshape(self, subscripts, final_subscripts, final_element=False):
+    def reshape(self, subscripts: SubscriptManager,
+                final_subscripts: dict,
+                final_element: bool = False) -> None:
+        """
+        Reshape the object to the desired subscripts. It will modify the
+        expression and lower the order if it is not 0.
+
+        Parameters
+        ----------
+        subscripts: SubscriptManager
+            The subscripts of the section.
+        final_subscripts: dict
+            The desired final subscripts.
+        final_element: bool (optional)
+            If True the array will be reshaped with the final subscripts
+            to have the shame shape. Otherwise, a length 1 dimension
+            will be included in the position to allow arithmetic
+            operations with other arrays. Default is False.
+
+        """
         if not final_subscripts or (
           self.subscripts == final_subscripts
           and list(self.subscripts) == list(final_subscripts)):
-            # same dictionary in the same orde, do nothing
+            # Same dictionary in the same order, do nothing
             pass
         elif not self.subscripts:
-            # original expression is not an array
+            # Original expression is not an array
             # NUMPY: object.expression = np.full(%s, %(shape)s)
             subscripts_out = subscripts.simplify_subscript_input(
                 final_subscripts)[1]
[…]
