+++ docs/structure/model_loading.rst
@@ -1,2 +1,11 @@
-Model model loading
-===================
+Model loading
+=============
+For loading a translated model with Python the function :py:func:`pysd.load` can be used:
+
+.. autofunction:: pysd.load
+
+To translate a load a model the :py:func:`pysd.read_vensim` and :py:func:`pysd.read_xmile` functions can be used:
+
+.. autofunction:: pysd.read_vensim
+
+.. autofunction:: pysd.read_xmile
