+++ docs/structure/python_builder.rst
@@ -9,35 +9,28 @@ In addition to translating individual commands between Vensim/XMILE and Python, 
 
 Main builders
 -------------
-
 .. automodule:: pysd.building.python.python_model_builder
    :members:
-   :undoc-members:
 
 Expression builders
 -------------------
 .. automodule:: pysd.building.python.python_expressions_builder
    :members:
-   :undoc-members:
 
 
 Namespace manager
 -----------------
-
 .. automodule:: pysd.building.python.namespace
    :members: NamespaceManager
-   :undoc-members:
 
 
 Subscript manager
 -----------------
 .. automodule:: pysd.building.python.subscripts
    :members:
-   :undoc-members:
 
 
 Imports manager
 ---------------
 .. automodule:: pysd.building.python.imports
    :members:
-   :undoc-members:
