+++ docs/functions.rst
@@ -1,12 +1,0 @@
-User Functions Reference
-========================
-
-These are the primary functions that control model import and execution.
-
-
-.. autofunction:: pysd.read_vensim
-
-.. autofunction:: pysd.read_xmile
-
-.. autofunction:: pysd.load
-
