+++ docs/structure/model_class.rst
@@ -1,0 +1,13 @@
+Python model class
+==================
+
+Model class
+-----------
+.. autoclass:: pysd.py_backend.statefuls.Model
+    :members:
+
+
+Macro class
+-----------
+.. autoclass:: pysd.py_backend.statefuls.Macro
+    :members:
