### 2022-05-11 35a71cf
Merge and correct PULSE functions
