+++ tests/pytest_pysd/pytest_functions.py
@@ -1,0 +1,352 @@
+import pytest
+import numpy as np
+import xarray as xr
+
+from pysd.py_backend.functions import\
+    ramp, step, pulse, xidz, zidz, if_then_else, sum, prod, vmin, vmax,\
+    invert_matrix
+
+
+class TestInputFunctions():
+    def test_ramp(self):
+        assert ramp(lambda: 14, .5, 10, 18) == 2
+
+        assert ramp(lambda: 4, .5, 10, 18) == 0
+
+        assert ramp(lambda: 24, .5, 10, 18) == 4
+
+        assert ramp(lambda: 24, .5, 10) == 7
+
+        assert ramp(lambda: 50, .5, 10) == 20
+
+    def test_step(self):
+        assert step(lambda: 5, 1, 10) == 0
+
+        assert step(lambda: 15, 1, 10) == 1
+
+        assert step(lambda: 10, 1, 10) == 1
+
+    def test_pulse(self):
+        assert pulse(lambda: 0, 1, width=3) == 0
+
+        assert pulse(lambda: 1, 1, width=3) == 1
+
+        assert pulse(lambda: 2, 1, width=3) == 1
+
+        assert pulse(lambda: 4, 1, width=3) == 0
+
+        assert pulse(lambda: 5, 1, width=3) == 0
+
+    def test_pulse_chain(self):
+        # before train starts
+        assert pulse(lambda: 0, 1, repeat_time=5, width=3, end=12) == 0
+        # on train start
+        assert pulse(lambda: 1, 1, repeat_time=5, width=3, end=12) == 1
+        # within first pulse
+        assert pulse(lambda: 2, 1, repeat_time=5, width=3, end=12) == 1
+        # end of first pulse
+        assert pulse(lambda: 4, 1, repeat_time=5, width=3, end=12) == 0
+        # after first pulse before second
+        assert pulse(lambda: 5, 1, repeat_time=5, width=3, end=12) == 0
+        # on start of second pulse
+        assert pulse(lambda: 6, 1, repeat_time=5, width=3, end=12) == 1
+        # within second pulse
+        assert pulse(lambda: 7, 1, repeat_time=5, width=3, end=12) == 1
+        # after second pulse
+        assert pulse(lambda: 10, 1, repeat_time=5, width=3, end=12) == 0
+        # on third pulse
+        assert pulse(lambda: 11, 1, repeat_time=5, width=3, end=12) == 1
+        # on train end
+        assert pulse(lambda: 12, 1, repeat_time=5, width=3, end=12) == 0
+        # after train
+        assert pulse(lambda: 15, 1, repeat_time=5, width=3, end=13) == 0
+
+    def test_pulse_magnitude(self):
+        from pysd.py_backend.statefuls import Time
+
+        # Pulse function with repeat time
+        # before first impulse
+        t = Time()
+        t.set_control_vars(initial_time=0, time_step=1)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 0
+        # first impulse
+        t.update(2)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 10
+        # after first impulse and before second
+        t.update(4)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 0
+        # second impulse
+        t.update(7)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 10
+        # after second and before third impulse
+        t.update(9)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 0
+        # third impulse
+        t.update(12)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 10
+        # after third impulse
+        t.update(14)
+        assert pulse(t, 2, repeat_time=5, magnitude=10) == 0
+
+        t = Time()
[…]
