+++ tests/unit_test_functions.py
@@ -1,322 +1,0 @@
-import unittest
-
-import numpy as np
-import xarray as xr
-
-
-class TestInputFunctions(unittest.TestCase):
-    def test_ramp(self):
-        from pysd.py_backend.functions import ramp
-
-        self.assertEqual(ramp(lambda: 14, .5, 10, 18), 2)
-
-        self.assertEqual(ramp(lambda: 4, .5, 10, 18), 0)
-
-        self.assertEqual(ramp(lambda: 24, .5, 10, 18), 4)
-
-        self.assertEqual(ramp(lambda: 24, .5, 10, -1), 7)
-
-    def test_step(self):
-        from pysd.py_backend.functions import step
-
-        self.assertEqual(step(lambda: 5, 1, 10), 0)
-
-        self.assertEqual(step(lambda: 15, 1, 10), 1)
-
-        self.assertEqual(step(lambda: 10, 1, 10), 1)
-
-    def test_pulse(self):
-        from pysd.py_backend.functions import pulse
-
-        self.assertEqual(pulse(lambda: 0, 1, 3), 0)
-
-        self.assertEqual(pulse(lambda: 1, 1, 3), 1)
-
-        self.assertEqual(pulse(lambda: 2, 1, 3), 1)
-
-        self.assertEqual(pulse(lambda: 4, 1, 3), 0)
-
-        self.assertEqual(pulse(lambda: 5, 1, 3), 0)
-
-    def test_pulse_chain(self):
-        from pysd.py_backend.functions import pulse_train
-        # before train starts
-        self.assertEqual(pulse_train(lambda: 0, 1, 3, 5, 12), 0)
-        # on train start
-        self.assertEqual(pulse_train(lambda: 1, 1, 3, 5, 12), 1)
-        # within first pulse
-        self.assertEqual(pulse_train(lambda: 2, 1, 3, 5, 12), 1)
-        # end of first pulse
-        self.assertEqual(pulse_train(lambda: 4, 1, 3, 5, 12), 0)
-        # after first pulse before second
-        self.assertEqual(pulse_train(lambda: 5, 1, 3, 5, 12), 0)
-        # on start of second pulse
-        self.assertEqual(pulse_train(lambda: 6, 1, 3, 5, 12), 1)
-        # within second pulse
-        self.assertEqual(pulse_train(lambda: 7, 1, 3, 5, 12), 1)
-        # after second pulse
-        self.assertEqual(pulse_train(lambda: 10, 1, 3, 5, 12), 0)
-        # on third pulse
-        self.assertEqual(pulse_train(lambda: 11, 1, 3, 5, 12), 1)
-        # on train end
-        self.assertEqual(pulse_train(lambda: 12, 1, 3, 5, 12), 0)
-        # after train
-        self.assertEqual(pulse_train(lambda: 15, 1, 3, 5, 13), 0)
-
-    def test_pulse_magnitude(self):
-        from pysd.py_backend.functions import pulse_magnitude
-        from pysd.py_backend.statefuls import Time
-
-        # Pulse function with repeat time
-        # before first impulse
-        t = Time()
-        t.set_control_vars(initial_time=0, time_step=1)
-        self.assertEqual(pulse_magnitude(t, 10, 2, 5), 0)
-        # first impulse
-        t.update(2)
-        self.assertEqual(pulse_magnitude(t, 10, 2, 5), 10)
-        # after first impulse and before second
-        t.update(4)
-        self.assertEqual(pulse_magnitude(t, 10, 2, 5), 0)
-        # second impulse
-        t.update(7)
-        self.assertEqual(pulse_magnitude(t, 10, 2, 5), 10)
-        # after second and before third impulse
[…]
