+++ pysd/_version.py
@@ -1,1 +1,1 @@
-__version__ = "3.0.0-dev"
+__version__ = "3.0.0"
