+++ pysd/py_backend/functions.py
@@ -4,8 +4,6 @@ stack, or require information about the internal state of the system beyond
 what is present in the function call. We provide them in a structure that
 makes it easy for the model elements to call.
 """
-
-from timeit import repeat
 import warnings
 
 import numpy as np
