+++ docs/structure/xmile_translation.rst
@@ -5,7 +5,7 @@ PySD allows parsing a Xmile file and translates the result to an :py:class:`Abst
 
 
 .. warning::
-    Currently no Xmile users are working in the development of PySD, this is causing a gap between the Xmile and Vensim developments. Stella users are encouraged to take part in the development of PySD by invcludying new test models and adding support for new functions and features.
+    Currently no Xmile users are working in the development of PySD, this is causing a gap between the Xmile and Vensim developments. Stella users are encouraged to take part in the development of PySD by invcludying new `test models <https://github.com/SDXorg/test-models>`_ and adding support for new functions and features.
 
 
 The translation workflow
@@ -86,9 +86,9 @@ Several subscript related features all supported. This include:
 - Basic subscript operations with different ranges.
 - Subscript ranges and subranges definitions.
 
-Grafic functions
-^^^^^^^^^^^^^^^^
-Xmile grafic functions (gf), also known as lookups, are supported. They can be given hardcoded or inline.
+Graphical functions
+^^^^^^^^^^^^^^^^^^^
+Xmile graphical functions (gf), also known as lookups, are supported. They can be given hardcoded or inline.
 
 .. warning::
    Interpolation methods 'extrapolate' and 'discrete' are implemented but not tested. Full integration models with this methods are required.
