+++ docs/development/complement.rst
@@ -3,8 +3,12 @@ Complementary Projects
 
 The most valuable component for better integrating models with *basically anything else* is a standard language for communicating the structure of those models. That language is `XMILE <http://www.iseesystems.com/community/support/XMILE.aspx>`_. The draft specifications for this have been finalized and the standard should be approved in the next few months.
 
-A python library for analyzing system dynamics models called the `Exploratory Modeling and Analysis (EMA) Workbench <http://simulation.tbm.tudelft.nl/ema-workbench/contents.html>`_ is being developed by `Erik Pruyt <http://www.tbm.tudelft.nl/en/about-faculty/departments/multi-actor-systems/policy-analysis/people/erik-pruyt/>`_ and `Jan Kwakkel <https://github.com/quaquel>`_ at TU Delft. This package implements a variety of analysis methods that are unique to dynamic models, and could work very tightly with PySD.
+A Python library for analyzing system dynamics models called the `Exploratory Modeling and Analysis (EMA) Workbench <http://simulation.tbm.tudelft.nl/ema-workbench/contents.html>`_ is being developed by `Erik Pruyt <http://www.tbm.tudelft.nl/en/about-faculty/departments/multi-actor-systems/policy-analysis/people/erik-pruyt/>`_ and `Jan Kwakkel <https://github.com/quaquel>`_ at TU Delft. This package implements a variety of analysis methods that are unique to dynamic models, and could work very tightly with PySD.
 
-An excellent javascript library called `sd.js <https://github.com/bpowers/sd.js/tree/master>`_ created by Bobby Powers at `SDlabs <http://sdlabs.io/>`_ exists as a standalone SD engine, and provides a beautiful front end. This front end could be rendered as an iPython widget to facilitate display of SD models.
+An excellent JavaScript library called `sd.js <https://github.com/bpowers/sd.js/tree/master>`_ created by Bobby Powers at `SDlabs <http://sdlabs.io/>`_ exists as a standalone SD engine, and provides a beautiful front end. This front end could be rendered as an iPython widget to facilitate display of SD models.
 
 The `Behavior Analysis and Testing Software(BATS) <http://www.ie.boun.edu.tr/labs/sesdyn/projects/bats/index.html>`_ delveloped by `Gönenç Yücel <http://www.ie.boun.edu.tr/people/pages/yucel.html>`_ includes a really neat method for categorizing behavior modes and exploring parameter space to determine the boundaries between them.
+
+The `SDQC library <https://gitlab.com/eneko.martin.martinez/sdqc>` developed by Eneko Martin Martinez may be used to check the quality of the data imported by Vensim models from speadsheet files.
+
+The `excels2vensim library <https://gitlab.com/eneko.martin.martinez/excels2vensim>`, also developed by Eneko Martin Martinez, aims to simplify the incorporation of equations from external data into Vensim. 
