### 2022-05-12 39c99dd
Remove variables in documentation when selecting a submodel
