### 2022-05-12 e9cbc77
Move model and macro to a separate file
