+++ pysd/py_backend/statefuls.py
@@ -1,26 +1,15 @@
 """
 The stateful objects are used and updated each time step with an update
-method. This include from basic Integ class objects until the Model
-class objects.
+method. This include Integs, Delays, Forecasts, Smooths, and Trends,
+between others. The Macro class and Model class are also Stateful class
+child. But defined in the file model.py.
 """
-import inspect
-import pickle
 import warnings
-from typing import Union
 
 import numpy as np
-import pandas as pd
 import xarray as xr
 
-from . import utils
 from .functions import zidz, if_then_else
-from .external import External, Excels
-from .cache import Cache, constant_cache
-from .data import TabData
-from .lookups import HardcodedLookups
-from .components import Components, Time
-
-from pysd._version import __version__
 
 
 small_vensim = 1e-6  # What is considered zero according to Vensim Help
@@ -575,1685 +564,3 @@ class Initial(Stateful):
 
     def export(self):
         return {'state': self.state}
-
-
-class Macro(DynamicStateful):
-    """
-    The Macro class implements a stateful representation of the system,
-    and contains the majority of methods for accessing and modifying
-    components.
-
-    When the instance in question also serves as the root model object
-    (as opposed to a macro or submodel within another model) it will have
-    added methods to facilitate execution.
-
-    The Macro object will be created with components drawn from a
-    translated python model file.
-
-    Parameters
-    ----------
-    py_model_file: str or pathlib.Path
-        Filename of a model or macro which has already been converted
-        into a python format.
-    params: dict or None (optional)
-        Dictionary of the macro parameters. Default is None.
-    return_func: str or None (optional)
-        The name of the function to return from the macro. Default is None.
-    time: components.Time or None (optional)
-        Time object for integration. If None a new time object will
-        be generated (for models), if passed the time object will be
-        used (for macros). Default is None.
-    time_initialization: callable or None
-        Time to set at the begginning of the Macro. Default is None.
-    data_files: dict or list or str or None
-        The dictionary with keys the name of file and variables to
-        load the data from there. Or the list of names or name of the
-        file to search the data in. Only works for TabData type object
-        and it is neccessary to provide it. Default is None.
-    py_name: str or None
-        The name of the Macro object. Default is None.
-
-    """
-    def __init__(self, py_model_file, params=None, return_func=None,
-                 time=None, time_initialization=None, data_files=None,
-                 py_name=None):
-        super().__init__()
-        self.time = time
-        self.time_initialization = time_initialization
-        self.cache = Cache()
-        self.py_name = py_name
-        self.external_loaded = False
-        self.lookups_loaded = False
-        self.components = Components(str(py_model_file), self.set_components)
-
-        if __version__.split(".")[0]\
-           != self.get_pysd_compiler_version().split(".")[0]:
-            raise ImportError(
-                "\n\nNot able to import the model. "
-                + "The model was translated with a "
-                + "not compatible version of PySD:"
-                + "\n\tPySD " + self.get_pysd_compiler_version()
-                + "\n\nThe current version of PySd is:"
-                + "\n\tPySD " + __version__ + "\n\n"
-                + "Please translate again the model with the function"
-                + " read_vensim or read_xmile.")
-
-        self._namespace = self.components._components.component.namespace
-        self._dependencies =\
-            self.components._components.component.dependencies.copy()
-        self._subscript_dict = getattr(
-            self.components._components, "_subscript_dict", {})
-        self._modules = getattr(
-            self.components._components, "_modules", {})
-
-        self._doc = self._build_doc()
-
-        if params is not None:
-            # add params to namespace
-            self._namespace.update(self.components._components._params)
-            # create new components with the params
-            self.set_components(params, new=True)
-            # update dependencies
-            for param in params:
-                self._dependencies[
[…]
