### 2022-05-12 ad54b03
Update path to model in documentation
