+++ docs/development/complement.rst
@@ -9,6 +9,6 @@ An excellent JavaScript library called `sd.js <https://github.com/bpowers/sd.js/
 
 The `Behavior Analysis and Testing Software(BATS) <http://www.ie.boun.edu.tr/labs/sesdyn/projects/bats/index.html>`_ delveloped by `Gönenç Yücel <http://www.ie.boun.edu.tr/people/pages/yucel.html>`_ includes a really neat method for categorizing behavior modes and exploring parameter space to determine the boundaries between them.
 
-The `SDQC library <https://gitlab.com/eneko.martin.martinez/sdqc>` developed by Eneko Martin Martinez may be used to check the quality of the data imported by Vensim models from speadsheet files.
+The `SDQC library <https://gitlab.com/eneko.martin.martinez/sdqc>`_ developed by Eneko Martin Martinez may be used to check the quality of the data imported by Vensim models from speadsheet files.
 
-The `excels2vensim library <https://gitlab.com/eneko.martin.martinez/excels2vensim>`, also developed by Eneko Martin Martinez, aims to simplify the incorporation of equations from external data into Vensim. 
+The `excels2vensim library <https://gitlab.com/eneko.martin.martinez/excels2vensim>`_, also developed by Eneko Martin Martinez, aims to simplify the incorporation of equations from external data into Vensim. 
