+++ docs/conf.py
@@ -23,7 +23,6 @@ sys.path.insert(0, os.path.abspath('../'))
 from docs.generate_tables import generate_tables
 
 
-
 # Generate tables used for documentation
 generate_tables()
 
