### 2022-05-12 f164338
Update paths in MANIFEST and setup
