+++ docs/basic_usage.rst
@@ -18,7 +18,7 @@ This code creates an instance of the :doc:`PySD Model class <structure/model_cla
 .. note::
    The teacup model can be found in the `samples of the test-models repository <https://github.com/SDXorg/test-models/tree/master/samples>`_.
 
-To view a synopsis of the model equations and documentation, use the :py:func:`.doc` property of the Model class. This will generate a listing of all model elements, their documentation, units, and initial values, where appropriate, and return them as a :py:class:`pandas.DataFrame`. Here is a sample from the teacup model::
+To view a synopsis of the model equations and documentation, use the :py:attr:`.doc` property of the Model class. This will generate a listing of all model elements, their documentation, units, and initial values, where appropriate, and return them as a :py:class:`pandas.DataFrame`. Here is a sample from the teacup model::
 
    >>> model.doc
 
@@ -49,7 +49,7 @@ To view a synopsis of the model equations and documentation, use the :py:func:`.
 
 Running the Model
 -----------------
-The simplest way to simulate the model is to use the :py:func:`.run()` command with no options. This runs the model with the default parameters supplied in the model file, and returns a :py:class:`pandas.DataFrame` of the values of the model components at every timestamp::
+The simplest way to simulate the model is to use the :py:meth:`.run` command with no options. This runs the model with the default parameters supplied in the model file, and returns a :py:class:`pandas.DataFrame` of the values of the model components at every timestamp::
 
    >>> stocks = model.run()
    >>> stocks
@@ -82,13 +82,13 @@ Pandas proovides a simple plotting capability, that we can use to see how the te
    :width: 400 px
    :align: center
 
-To show a progressbar during the model integration, the `progress` argument can be passed to the :py:func:`.run()` command::
+To show a progressbar during the model integration, the `progress` argument can be passed to the :py:meth:`.run` command::
 
    >>> stocks = model.run(progress=True)
 
 Running models with DATA type components
 ^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^^
-Venim allows to import DATA type data from binary `.vdf` files. Variables defined without an equation in the model, will attempt to read their values from the `.vdf`. PySD allows running models with this kind of data definition using the data_files argument when calling :py:func:`.run()` command, e.g.::
+Venim allows to import DATA type data from binary `.vdf` files. Variables defined without an equation in the model, will attempt to read their values from the `.vdf`. PySD allows running models with this kind of data definition using the data_files argument when calling :py:meth:`.run` command, e.g.::
 
    >>> stocks = model.run(data_files="input_data.tab")
 
@@ -109,7 +109,7 @@ If a variables are defined in different files, to choose the specific file a dic
 
 Outputting various run information
 ----------------------------------
-The :py:func:`.run()` command has a few options that make it more useful. In many situations we want to access components of the model other than merely the stocks - we can specify which components of the model should be included in the returned dataframe by including them in a list that we pass to the :py:func:`.run()` command, using the return_columns keyword argument::
+The :py:meth:`.run` command has a few options that make it more useful. In many situations we want to access components of the model other than merely the stocks - we can specify which components of the model should be included in the returned dataframe by including them in a list that we pass to the :py:meth:`.run` command, using the return_columns keyword argument::
 
    >>> model.run(return_columns=['Teacup Temperature', 'Room Temperature'])
 
@@ -129,7 +129,7 @@ The :py:func:`.run()` command has a few options that make it more useful. In man
    [241 rows x 2 columns]
 
 
-If the measured data that we are comparing with our model comes in at irregular timestamps, we may want to sample the model at timestamps to match. The :py:func:`.run()` function provides this functionality with the return_timestamps keyword argument::
[…]
