### 2022-05-13 ea4ebe9
review docs of vensim_element.py
