### 2022-05-13 906bee5
final revision of vensim_element.py docs
