### 2022-05-13 55259b1
review docs of vensim_translation.rst
