### 2022-05-13 95a3891
review docs for xmile_translation.rst
