+++ docs/structure/xmile_translation.rst
@@ -5,18 +5,18 @@ PySD allows parsing a Xmile file and translates the result to an :py:class:`Abst
 
 
 .. warning::
-    Currently no Xmile users are working in the development of PySD, this is causing a gap between the Xmile and Vensim developments. Stella users are encouraged to take part in the development of PySD by invcludying new `test models <https://github.com/SDXorg/test-models>`_ and adding support for new functions and features.
+    Currently no Xmile users are working on the development of PySD. This is causing a gap between the Xmile and Vensim developments. Stella users are encouraged to take part in the development of PySD by includying new `test models <https://github.com/SDXorg/test-models>`_ and adding support for new functions and features.
 
 
 The translation workflow
 -------------------------
-The following translation workflow allows splitting the Xmile file while parsing each part of it in order to make it possible to build an :py:class:`AbstractModel` type object. The workflow could be summarized as follows:
+The following translation workflow allows splitting the Xmile file while parsing each part of it to build an :py:class:`AbstractModel` type object. The workflow may be summarized as follows:
 
-1. Xmile file: Parses the file with etree library and creates a section for the model.
-2. Xmile section: Full set of varibles and definitions that can be integrated. Allows splitting the model elements.
-3. Xmile element: A variable definition. It includes units and commnets. Allows parsing its the expressions that contains and saving them inside AbstractComponents that are part of an AbstractElement.
+1. **Xmile file**: Parses the file with etree library and creates a section for the model.
+2. **Xmile section**: Full set of varibles and definitions that can be integrated. Allows splitting the model elements.
+3. **Xmile element**: A variable definition. It includes units and commnets. Allows parsing the expressions it contains and saving them inside AbstractComponents, that are part of an AbstractElement.
 
-Once the model is parsed and broken following the previous steps. The :py:class:`AbstractModel` can be returned.
+Once the model is parsed and split following the previous steps. The :py:class:`AbstractModel` can be returned.
 
 
 Xmile file
@@ -44,14 +44,14 @@ Xmile element
 Supported Functions and Features
 --------------------------------
 
-Ongoing development of the translator will support the full subset of Xmile functionality. The current release supports the following operators, functions and features.
+Ongoing development of the translator will support the full set of Xmile functionality. The current release supports the following operators, functions and features:
 
 .. warning::
    Not all the supported functions and features are properly tested. Any new test model to cover the missing functions test will be wellcome.
 
 Operators
 ^^^^^^^^^
-All the basic operators are supported, this includes the ones shown in the tables below.
+All the basic operators are supported, this includes the ones shown in the tables below.:
 
 .. csv-table:: Supported unary operators
    :file: ../tables/unary_xmile.csv
@@ -64,7 +64,7 @@ All the basic operators are supported, this includes the ones shown in the table
 
 Functions
 ^^^^^^^^^
-Not all the Xmile functions all included yet, the list of supported functions are included below.
+Not all the Xmile functions all included yet, the list of supported functions are included below:
 
 .. csv-table:: Supported basic functions
    :file: ../tables/functions_xmile.csv
@@ -77,27 +77,27 @@ Not all the Xmile functions all included yet, the list of supported functions ar
 
 Stocks
 ^^^^^^
-Stocks are supported with any number of inflows and outflows. The stocks are translated to the AST as `IntegStructure(flows, initial_value)`.
+Stocks are supported with any number of inflows and outflows. Stocks are translated to the AST as `IntegStructure(flows, initial_value)`.
 
 Subscripts
 ^^^^^^^^^^
-Several subscript related features all supported. This include:
+Several subscript related features are supported. Thiese include:
 
[…]
