### 2022-05-13 fec83c7
review vensim_section.py docs
