### 2022-05-18 JamesPHoughton/pysd#319 @blsqr
[JOSS Review] Documentation enhancements
Hi again @rogersamso, @enekomartinmartinez, and co-authors 👋

This issue is part of the [PySD review at JOSS](https://github.com/openjournals/joss-reviews/issues/4329). Following the [review checklist](https://github.com/openjournals/joss-reviews/issues/4329#issuecomment-1104329465), I'd like to make a few remarks regarding the PySD documentation.

First things first: The documentation is well written and very helpful for getting started with PySD and running some example models. With resources like the cookbook linked, I imagine this to be a good starting point for using PySD.
I also like the developer-focussed documentation that details program structure and philosophy!

For publication in JOSS, the following points need to be addressed, though:

* [x] **Statement of need:** This is well-explained in the software paper, but is (to my taste) a bit on the short side in the docs. I'd suggest to keep the first two sentences on the docs start page (because they are very crisp), but add another paragraph or so that offers a bit more insight into the motivation behind PySD and the possibilities it offers.
* [x] **Functionality documentation:** While you do include bits and pieces of API documentation in appropriate places, the User Functions Reference page does not appear to document PySD's full public API, correct? While I like that it highlights some important functions, I was not able to find a reference for other noteworthy structures like the `Model` class, it's `run` method etc.. There should be a page containing the *full* public API reference.

In addition, I'd have the following thoughts and suggestions regarding the documentation (all optional):

* [x] The start page currently shows a two-level listing of all pages; this is essentially redundant with the sidebar navigation. I would propose to remove the TOC from the start page and instead focus on more descriptive page titles, such that readers can use the sidebar navigation more efficiently. This would also make the content at the bottom of the start page more discoverable.
* [x] Consider renaming "Basic Usage" page to something like "Getting started" to more clearly direct new users to this page.
* [x] Move the "Supported Functions" section (currently on bottom of "Basic Usage" page) to a more appropriate place. To me, that table seems misplaced there; perhaps a separate page that focusses on "Loading vensim and xmile models" would be better?
  * [x] That page could also highlight potential limitations or difficulties when importing these models.
  * [x] Perhaps there is enough overlap with the "Advanced Usage" page to address it all there?
  * [x] Should there also be a table of supported Stella functions? 🤔
* [x] Make the "About the project" page more visible; I think that's very interesting but it's a bit hidden in the developer docs. This could easily fit on the root level of the docs, I think.
* [x] The "Contributing to PySD" page seems to be mostly about development guidelines. Given that there already is a remark about contributing to PySD on the start page, it may be clearer to rename this page to something like "Development guidelines"?
* [ ] Consider deploying more than just `master`/`latest` to readthedocs.

If there's something unclear in what I wrote above or if I can be of help with the necessary changes, please let me know.
