### 2022-05-18 #4329 @blsqr
Regarding the list of authors: Could you briefly explain the role of Jordi Solé in the creation of PySD and this paper?
