### 2022-05-18 JamesPHoughton/pysd#319 @enekomartinmartinez
@blsqr Thanks for those suggestions :)
I will implement them and mark the checkboxes when done.

@JamesPHoughton could you take charge of the deployment of more versions than master/latest in RTD?
