+++ docs/structure/structure_index.rst
@@ -14,16 +14,14 @@ For models translated to Python, all the necessary functions and classes to run 
 Translation
 -----------
 
-The internals of the translation process may be found in the following links of the documentation:
-
 .. toctree::
-   :maxdepth: 2
+   :hidden:
 
    vensim_translation
    xmile_translation
    abstract_model
 
-
+PySD currentlty supports translation :doc:`from Vensim <vensim_translation>` amb :doc:`from Xmile <xmile_translation>`.
 
 PySD can import models in Vensim's \*.mdl file format and in XMILE format (\*.xml, \*.xmile, or \*.stmx file). `Parsimonious <https://github.com/erikrose/parsimonious>`_ is the Parsing Expression Grammar `(PEG) <https://en.wikipedia.org/wiki/Parsing_expression_grammar>`_ parser library used in PySD to parse the original models and construct an abstract syntax tree. The translators then crawl the tree, using a set of classes to define the :doc:`Abstract Model <abstract_model>`.
 
@@ -53,31 +51,31 @@ It is important to maintain this order because although these operations by defi
 Building the model
 ------------------
 
-The builders allow to build the final model in any programming language (so long as there is a builder for that particular language). To do so, they use a series of classes that obtain the information from the Abstract Model and convert it into the desired code. Currently PySD only includes a builder to build the models in Python. Any contribution to add new builders (and solvers) for other programming languages is welcome.
+The builders allow to build the final model in any programming language (so long as there is a builder for that particular language). To do so, they use a series of classes that obtain the information from the Abstract Model and convert it into the desired code. Currently PySD only includes a :doc:`builder to build the models in Python <python_builder>` . Any contribution to add new builders (and solvers) for other programming languages is welcome.
 
 .. toctree::
-   :maxdepth: 2
+   :hidden:
 
    python_builder
 
 The Python model
 ----------------
-For loading a translated model with Python see :doc:`basic usage <../basic_usage>` or:
+For loading a translated model with Python see :doc:`Getting started <../../getting_started>` or :doc:`Model loading <model_loading>`:
 
 .. toctree::
-   :maxdepth: 2
+   :hidden:
 
    model_loading
 
 The Python builder constructs a Python class that represents the system dynamics model. The class maintains a dictionary representing the current values of each of the system stocks, and the current simulation time, making it a `statefull` model in much the same way that the system itself has a specific state at any point in time.
 
-The Model class also contains a function for each of the model components, representing the essential model equations. Each function contains its units, subcscripts type infromation and documentation as translated from the original model file. A query to any of the model functions will calculate and return its value according to the stored state of the system.
+The :doc:`Model class <model_class>` also contains a function for each of the model components, representing the essential model equations. Each function contains its units, subcscripts type infromation and documentation as translated from the original model file. A query to any of the model functions will calculate and return its value according to the stored state of the system.
 
-The Model class maintains only a single state of the system in memory, meaning that all functions must obey the Markov property  - that the future state of the system can be calculated entirely based upon its current state. In addition to simplifying integration, this requirement enables analyses that interact with the model at a step-by-step level.
+The :doc:`Model class <model_class>` maintains only a single state of the system in memory, meaning that all functions must obey the Markov property  - that the future state of the system can be calculated entirely based upon its current state. In addition to simplifying integration, this requirement enables analyses that interact with the model at a step-by-step level.
 
[…]
