### 2022-05-18 JamesPHoughton/pysd#319 @enekomartinmartinez
@blsqr I have been working on the PR #312. We have this PR almost ready for merge which includes part of the future roadmap of the paper (splitting parsing from building and Abstract Model representation, it doesn't include the numpy array based backend). We will soon modify the paper slightly to keep it updated.

We are working now on finishing the documentation of that branch and we will include your suggestion. As soon everything is ready I will let you know.

Thanks :)
