### 2022-05-18 JamesPHoughton/pysd#319 @JamesPHoughton
> Consider deploying more than just master/latest to readthedocs
Happy to look at it, can you help me understand more about what this means? Would we want to have a separate version for 2.0 vs 3.0?
