### 2022-05-19 JamesPHoughton/pysd#319 @blsqr
> … can you help me understand more about what this means?

@JamesPHoughton ReadTheDocs offers to build [Versioned Documentation](https://docs.readthedocs.io/en/stable/versions.html) for tags, such that you can select between the docs of version 2.0, 2.2.4, … via a selector like this, including the new `stable` tag which points to the latest *release* (not commit like `latest`):

Whether you *want* to do that is another question, I'd say. The main advantage of versioned documentation may be of an archival nature: it allows users to refer to a certain version and have the corresponding docs available on RTD. This may also come in handy when publishing PySD models that were created with a specific version.
If, however, you'd want to nudge users to always working with the latest PySD version (or even stay on HEAD), I think it's understandable to not differentiate versions.

Regarding implementation: In the RTD admin interface you might want to check the Advanced Settings and make sure the "Single Version" setting is not checked. RTD typically sets up a webhook automatically upon project import, but you may have to do this [manually](https://docs.readthedocs.io/en/stable/integrations.html#integration-creation), if it does not work. After the webhook is set up, each new tag that matches a semantic version number will trigger a build.

Also note that you can manually edit which versions will be shown publicly. I believe you can also manually trigger builds for older tags, if needed.
