### 2022-05-19 7c719cc
Document and set flatten_output=True by default
