+++ docs/whats_new.rst
@@ -8,24 +8,24 @@ v3.0.0 (unreleased)
 New Features
 ~~~~~~~~~~~~
 
-- Properties added to the :py:class:`Model` to make more accessible some information: :py:attr:`.namespace`, :py:attr:`.subscripts`, :py:attr:`.dependencies`, :py:attr:`.modules`, :py:attr:`.doc`.
+- Properties added to the :py:class:`pysd.py_backend.model.Macro` to make more accessible some information: :py:attr:`.namespace`, :py:attr:`.subscripts`, :py:attr:`.dependencies`, :py:attr:`.modules`, :py:attr:`.doc`.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
 
 - The argument :py:data:`flatten_output` from :py:meth:`.run` is now set to :py:data:`True` by default.
 - The docstring of the model is now a property and thus is it not callable,:py:attr:`.doc`.
-- Allow the function :py:func:`py_backend.functions.pulse` to also perform the operations performed by :py:data:`py_backend.functions.pulse_train()` and :py:data:`py_backend.functions.pulse_magnitude()`.
-- The first argument of :py:func:`py_backend.functions.active_initial` now is the stage and not the time.
-- The function :py:data:`py_backend.utils.rearrange` now its mutch simpler oriented to perform simple rearrange cases for user interaction.
+- Allow the function :py:func:`pysd.py_backend.functions.pulse` to also perform the operations performed by :py:data:`pysd.py_backend.functions.pulse_train()` and :py:data:`pysd.py_backend.functions.pulse_magnitude()`.
+- The first argument of :py:func:`pysd.py_backend.functions.active_initial` now is the stage and not the time.
+- The function :py:data:`pysd.py_backend.utils.rearrange()` now its mutch simpler oriented to perform simple rearrange cases for user interaction.
 - The translation and the building of models has been totally modified to use the :doc:`Abstract Model Representation <structure/structure_index>`.
-- Move :py:data:`py_backend.statefuls.Model` and  :py:data:`py_backend.statefuls.Macro` to  :py:class:`py_backend.model.Model` and :py:class:`py_backend.model.Macro`, respectively.
-- All kinds of lookups are now managed with the :py:class:`py_backend.lookups.Lookups` class.
+- Move :py:data:`pysd.py_backend.statefuls.Model` and  :py:data:`pysd.py_backend.statefuls.Macro` to  :py:class:`pysd.py_backend.model.Model` and :py:class:`pysd.py_backend.model.Macro`, respectively.
+- All kinds of lookups are now managed with the :py:class:`pysd.py_backend.lookups.Lookups` class.
 
 Deprecations
 ~~~~~~~~~~~~
 
-- Remove :py:data:`py_backend.utils.xrmerge()`, :py:data:`py_backend.functions.pulse_train()`, :py:data:`py_backend.functions.pulse_magnitude()`, :py:data:`py_backend.functions.lookup()`, :py:data:`py_backend.functions.lookup_discrete()`, :py:data:`py_backend.functions.lookup_extrapolation()`, :py:data:`py_backend.functions.logical_and()`, :py:data:`py_backend.functions.logical_or()`, :py:data:`py_backend.functions.bounded_normal()`, :py:data:`py_backend.functions.log()`.
[…]
