+++ docs/whats_new.rst
@@ -9,31 +9,30 @@ New Features
 ~~~~~~~~~~~~
 
 - The new :doc:`Abstract Model Representation <structure/structure_index>` translation and building workflow will allow to add new output languages in the future.
-- Properties added to the :py:class:`pysd.py_backend.model.Macro` to make more accessible some information: :py:attr:`.namespace`, :py:attr:`.subscripts`, :py:attr:`.dependencies`, :py:attr:`.modules`, :py:attr:`.doc`.
-- The Python models now look cleaner:
-    - :py:data:`_namespace` and :py:data:`_dependencies` dictionaries are removed from the file.
-    - Variables original names, dependencies metadata are given through :py:meth:`pysd.py_backend.components.Component.add` decorator, instead of having them in the docstring.
-    - The merging of variable equations is done using the coordinates to a pre-allocated array, instead of using the `magic` function :py:data:`pysd.py_backend.utils.xrmerge()`.
-    - The arranging and subseting arrays are now done inplace instead of using the magic function :py:data:`pysd.py_backend.utils.rearrange()`.
+- Added new properties to the :py:class:`pysd.py_backend.model.Macro` to make more accessible some information: :py:attr:`.namespace`, :py:attr:`.subscripts`, :py:attr:`.dependencies`, :py:attr:`.modules`, :py:attr:`.doc`.
+- Cleaner Python models:
+    - :py:data:`_namespace` and :py:data:`_dependencies` dictionaries have been removed from the file.
+    - Variables original names, dependencies metadata now are given through :py:meth:`pysd.py_backend.components.Component.add` decorator, instead of having them in the docstring.
+    - Merging of variable equations is now done using the coordinates to a pre-allocated array, instead of using the `magic` function :py:data:`pysd.py_backend.utils.xrmerge()`.
+    - Arranging and subseting arrays are now done inplace instead of using the magic function :py:data:`pysd.py_backend.utils.rearrange()`.
 
 Breaking changes
 ~~~~~~~~~~~~~~~~
 
-- The argument :py:data:`flatten_output` from :py:meth:`.run` is now set to :py:data:`True` by default.
-- The docstring of the model is now a property and thus is it not callable,:py:attr:`.doc`.
+- Set the argument :py:data:`flatten_output` from :py:meth:`.run` to :py:data:`True` by default. Previously it was set to :py:data:`False` by default.
+- Move the docstring of the model to a property, :py:attr:`.doc`. Thus, it is not callable anymore.
 - Allow the function :py:func:`pysd.py_backend.functions.pulse` to also perform the operations performed by :py:data:`pysd.py_backend.functions.pulse_train()` and :py:data:`pysd.py_backend.functions.pulse_magnitude()`.
-- The first argument of :py:func:`pysd.py_backend.functions.active_initial` now is the stage and not the time.
-- The function :py:data:`pysd.py_backend.utils.rearrange()` now its mutch simpler oriented to perform simple rearrange cases for user interaction.
-- The translation and the building of models has been totally modified to use the :doc:`Abstract Model Representation <structure/structure_index>`.
+- Change first argument of :py:func:`pysd.py_backend.functions.active_initial`, now it is the `stage of the model` and not the `time`.
+- Simplify the function :py:data:`pysd.py_backend.utils.rearrange()` orienting it to perform simple rearrange cases for user interaction.
 - Move :py:data:`pysd.py_backend.statefuls.Model` and  :py:data:`pysd.py_backend.statefuls.Macro` to  :py:class:`pysd.py_backend.model.Model` and :py:class:`pysd.py_backend.model.Macro`, respectively.
[…]
