### 2022-05-19 #4329 @rogersamso
Apart from being the one who bet on the use of PySD in the H2020 projects MEDEAS and LOCMOTION, which led to the major changes that took place from 0.9 to the newest 3.0, Jordi Solé has been part of all design decisions that have been taken during the whole process. He has also reviewed the paper several times, and is now following the revision process.
