### 2022-05-19 #4329 @blsqr
Thanks, @rogersamso! I did not find [direct code contributions](https://github.com/JamesPHoughton/pysd/graphs/contributors), that's what triggered the question – but that's obviously only one of many reasons for co-authorship.
