+++ paper/paper.md
@@ -42,13 +42,15 @@ PySD [@pysd:2014] is a Python library that transpiles models built in Stella&#17
 
 The main functional elements of PySD are 1) a set of parsing expression grammars (PEGs) (and their corresponding node visitor classes) to parse models built using Stella&#174; and Vensim&#174; (in *.xmile* and *.mdl* formats, respectively); 2) isomorphic implementations of the most frequently used Stella&#174; and Vensim&#174; built-in functions and other basic operations; 3) a builder, to write the parsed model code in Python and 4) a fordward Euler solver to run the models.
 
-Since release *v3.0.0*, the parsing process has been isolated from the building process. This has been achieved by creating an intermediate abstract model representation (AMR)(Figure \ref{fig:ABSTRACTMODEL}) that embeds all the information contained in the original models, plus additional information such as the order of operations, in pure Python objects. This intermediate AMR opens the door to the development of additional model builders (in Python) to write the output models in any programming language. Furthermore, the intermediate AMR may encourage the development of new tools to generate and analyze the graphs associated with the models.
+In addition to the aforementioned core functionality, PySD also allows users to import model inputs from external sources (from spreadsheet files), modify model variables at runtime, split models into any number of modules and submodules (corresponding to Vensim&#174; views), isolate parts of a model to be run individually, store intermediate simulation results and resume the simulation from that particular state, and run models using PySD alone (without Stella&#174; or Vensim&#174;). All these features are made available to the user through a command-line interface.
+
+Despite its maturity, PySD is currently in a very active development phase, and a proof to that is that most of the extra features listed in the paragraph above were implemented in the 18 months that separate releases *v0.11.0* and *v3.0.0*.
+
+The main novelty introduced by the recent *v3.0.0* release is that the parsing process is now isolated from the building process, which opens the door to the development of additional model builders (in Python) to write the output models in any programming language. This was achieved by creating an intermediate abstract model representation (AMR)(Figure \ref{fig:ABSTRACTMODEL}) that embeds all the information contained in the original models, plus additional information such as the order of operations, in pure Python objects.
 
 ![New model parsing-building logic in PySD v3.0.0. PEG acronym stands for Parsing Expression Grammar. \label{fig:ABSTRACTMODEL}](abstract_model.pdf)
 
-In addition to the aforementioned core functionality, PySD also allows users to import model inputs from external sources (from spreadsheet files), modify model variables at runtime, split models into any number of modules and submodules (corresponding to Vensim&#174; views), isolate parts of a model to be run individually, store intermediate simulation results and resume the simulation from that particular state, and run models using PySD alone (without Stella&#174; or Vensim&#174;). Finally, all these features are made available to the user through a command-line interface.
-
-Despite its maturity, PySD is currently in a very active development phase, and a proof to that is that most of the extra features listed in the paragraph above were implemented in the 18 months that separate releases *v0.11.0* and *v3.0.0*. The roadmap for new releases will focus on cutting down simulation times and on including additional built-in Stella&#174; and Vensim&#174; functions. The most relevant performance-oriented development will be the migration from xarray to numpy to perform array operations, which in addition to the expected overhead reduction will also allow for further potential model optimizations (i.e. JIT compilation).
+Future releases will focus on cutting down simulation times and on including additional built-in Stella&#174; and Vensim&#174; functions. The most relevant performance-oriented development will be the migration from xarray to numpy to perform array operations, which in addition to the expected overhead reduction will also allow for further potential model optimizations (i.e. JIT compilation).
 
 ## Statement of need
 
