+++ pysd/py_backend/statefuls.py
@@ -1,8 +1,8 @@
 """
-The stateful objects are used and updated each time step with an update
+The Stateful objects are used and updated each time step with an update
 method. This include Integs, Delays, Forecasts, Smooths, and Trends,
-between others. The Macro class and Model class are also Stateful class
-child. But defined in the file model.py.
+between others. The Macro class and Model class are also Stateful type.
+However, they are defined appart as they are more complex.
 """
 import warnings
 
@@ -61,20 +61,24 @@ class DynamicStateful(Stateful):
 
 class Integ(DynamicStateful):
     """
-    Implements INTEG function
+    Implements INTEG function.
+
+    Parameters
+    ----------
+    ddt: callable
+        Derivate to integrate.
+    initial_value: callable
+        Initial value.
+    py_name: str
+        Python name to identify the object.
+
+    Attributes
+    ----------
+    state: float or xarray.DataArray
+        Current state of the object. Value of the stock.
+
     """
     def __init__(self, ddt, initial_value, py_name):
-        """
-
-        Parameters
-        ----------
-        ddt: function
-          This will become an attribute of the object
-        initial_value: function
-          Initial value
-        py_name: str
-          Python name to identify the object
-        """
         super().__init__()
         self.init_func = initial_value
         self.ddt = ddt
@@ -96,7 +100,29 @@ class Integ(DynamicStateful):
 
 class Delay(DynamicStateful):
     """
-    Implements DELAY function
+    Implements DELAY function.
+
+    Parameters
+    ----------
+    delay_input: callable
+        Input of the delay.
+    delay_time: callable
+        Delay time.
+    initial_value: callable
+        Initial value.
+    order: callable
+        Delay order.
+    tsetp: callable
+        The time step of the model.
+    py_name: str
+        Python name to identify the object.
+
+    Attributes
+    ----------
+    state: numpy.array or xarray.DataArray
+        Current state of the object. Array of the delays values multiplied
+        by their corresponding average time.
+
     """
     # note that we could have put the `delay_input` argument as a parameter to
     # the `__call__` function, and more closely mirrored the vensim syntax.
@@ -107,17 +133,6 @@ class Delay(DynamicStateful):
 
     def __init__(self, delay_input, delay_time, initial_value, order, tstep,
                  py_name):
-        """
-
-        Parameters
-        ----------
-        delay_input: function
-        delay_time: function
-        initial_value: function
-        order: function
-        py_name: str
-          Python name to identify the object
-        """
         super().__init__()
         self.init_func = initial_value
         self.delay_time_func = delay_time
@@ -180,7 +195,34 @@ class Delay(DynamicStateful):
 
 class DelayN(DynamicStateful):
     """
-    Implements DELAY N function
+    Implements DELAY N function.
+
+    Parameters
+    ----------
+    delay_input: callable
+        Input of the delay.
+    delay_time: callable
+        Delay time.
+    initial_value: callable
+        Initial value.
+    order: callable
+        Delay order.
+    tsetp: callable
+        The time step of the model.
+    py_name: str
+        Python name to identify the object.
+
+    Attributes
+    ----------
+    state: numpy.array or xarray.DataArray
+        Current state of the object. Array of the delays values multiplied
+        by their corresponding average time.
+
+    times: numpy.array or xarray.DataArray
+        Array of delay times used for computing the delay output.
+        If delay_time is constant, this array will be constant and
+        DelayN will behave ad Delay.
+
     """
     # note that we could have put the `delay_input` argument as a parameter to
     # the `__call__` function, and more closely mirrored the vensim syntax.
@@ -191,17 +233,6 @@ class DelayN(DynamicStateful):
 
     def __init__(self, delay_input, delay_time, initial_value, order, tstep,
[…]
