+++ pysd/py_backend/functions.py
@@ -1,8 +1,10 @@
 """
-These functions have no direct analog in the standard Python data analytics
-stack, or require information about the internal state of the system beyond
-what is present in the function call. We provide them in a structure that
-makes it easy for the model elements to call.
+The provided functions have no direct analog in the standard Python data
+analytics stack, or require information about the internal state of the
+system beyond what is present in the function call. They are provided
+in a structure that makes it easy for the model elements to call. The
+functions may be similar to the original functions given by Vensim or
+Stella, but sometimes the number or order of arguments may change.
 """
 import warnings
 
@@ -18,7 +20,7 @@ def ramp(time, slope, start, finish=None):
 
     Parameters
     ----------
-    time: function
+    time: callable
         Function that returns the current time.
     slope: float
         The slope of the ramp starting at zero at time start.
@@ -30,7 +32,7 @@ def ramp(time, slope, start, finish=None):
 
     Returns
     -------
-    response: float
+    float or xarray.DataArray:
         If prior to ramp start, returns zero.
         If after ramp ends, returns top of ramp.
 
@@ -53,7 +55,7 @@ def step(time, value, tstep):
 
     Parameters
     ----------
-    time: function
+    time: callable
         Function that returns the current time.
     value: float
         The height of the step.
@@ -62,7 +64,7 @@ def step(time, value, tstep):
 
     Returns
     -------
-    float:
+    float or xarray.DataArray:
         - In range [-inf, tstep):
             returns 0
         - In range [tstep, +inf]:
@@ -78,7 +80,7 @@ def pulse(time, start, repeat_time=0, width=None, magnitude=None, end=None):
 
     Parameters
     ----------
-    time: function
+    time: callable
         Function that returns the current time.
     start: float
         Starting time of the pulse.
@@ -97,13 +99,12 @@ def pulse(time, start, repeat_time=0, width=None, magnitude=None, end=None):
 
     Returns
     -------
-    float:
+    float or xarray.DataArray:
         - In range [-inf, start):
             returns 0
         - In range [start + n*repeat_time, start + n*repeat_time + width):
             returns magnitude/time_step or 1
-        - In range [start + n*repeat_time + width,
-                    start + (n+1)*repeat_time):
+        - In range [start + n*repeat_time + width, start + (n+1)*repeat_time):
             returns 0
 
     """
@@ -126,14 +127,15 @@ def if_then_else(condition, val_if_true, val_if_false):
     Parameters
     ----------
     condition: bool or xarray.DataArray of bools
-    val_if_true: function
+    val_if_true: callable
         Value to evaluate and return when condition is true.
-    val_if_false: function
+    val_if_false: callable
         Value to evaluate and return when condition is false.
 
     Returns
     -------
-    The value depending on the condition.
+    float or xarray.DataArray:
+        The value depending on the condition.
 
     """
     # NUMPY: replace xr by np
@@ -178,8 +180,9 @@ def xidz(numerator, denominator, x):
 
     Returns
     -------
-    numerator/denominator if denominator > small_vensim
-    otherwise, returns value_if_denom_is_zero
+    float or xarray.DataArray:
+        - numerator/denominator if denominator > small_vensim
+        - value_if_denom_is_zero otherwise
 
     """
     # NUMPY: replace DataArray by np.ndarray, xr.where -> np.where
@@ -215,8 +218,9 @@ def zidz(numerator, denominator):
 
     Returns
     -------
-    result of division numerator/denominator if denominator is not zero,
-    otherwise zero.
+    float or xarray.DataArray:
+        - numerator/denominator if denominator > small_vensim
+        - 0 or 0s array otherwise
 
     """
[…]
