### 2022-05-20 #4329 @enekomartinmartinez
Dear @andrecieplinski, @blsqr and @pdebuyl
We are close to the release 3.0.0 in PySD, see [PR 312](https://github.com/JamesPHoughton/pysd/pull/312).
Therefore we have updated the last part of the summary section of the paper, when talking about the future roadmap.

We have also included in the documentation the suggestions made in [issue 319](https://github.com/JamesPHoughton/pysd/issues/319). Moreover, new sections are available in the documentation for release notes and explaining the structure of PySD. As the PR is still not built the documentation should be built locally using _paper_pysd_ or _split-parsing-building_ branches, you can also check the documentation from [RDT auto-build in the PR](https://pysd--312.org.readthedocs.build/en/312/). I wait for any comment about the documentation before merging 312 so we can include it in version 3.0.0.

Thanks a lot :)
