### 2022-05-20 JamesPHoughton/pysd#319 @enekomartinmartinez
Hi @blsqr ,
I think I have finished implementing your suggestions in the documentation. They have been very helpful to make it cleaner and more useful. If you think some other improvements needs to be done, please, let me know so I can include them before merging #312
