### 2022-05-20 JamesPHoughton/pysd#319 @blsqr
@rogersamso Having looked at the [PR build of the docs](https://pysd--312.org.readthedocs.build/en/312/index.html) for #312, I think it looks great and these are valuable changes. I'm happy you found my suggestions helpful.

There are no further suggestions from my side regarding #312. Whether you keep this issue open (for tracking the changes to RTD deployment) or close it is up to you.
