### 2022-05-20 JamesPHoughton/pysd#316 @blsqr
With #312 nearing a merge, I find the [revised software paper](https://github.com/openjournals/joss-papers/blob/f6672eb0488e6229682be2ace73c716fbcb1adbe/joss.04329/10.21105.joss.04329.pdf) more concise now. Very nice! 👍

Two minor details:
- In the bib entry of Forrester1971, I believe the `[by] jay w. forrester` can be removed? If not, the name should be made title case in the bib file.
- It is a bit unfortunate that the "co-first author" footnote symbols are different and the footnote are repeated. I don't know if JOSS's markdown parser (which I believe is pandoc) supports this, but it may be worth trying if [this syntax](https://www.markdownguide.org/extended-syntax/#footnotes) is working? It would probably mean adding something like `[^*]` to both author names in the table (`*` becoming the symbol), and adding `[^*]: co-first author` at the beginning of the text? If that does not work, I would propose to inquire regarding the best practice of denoting the authorship status. What do you think?

Also, are there representative publications that can be cited for the LOCOMOTION project, despite it being still ongoing? Given that it's not unlikely that a project website breaks in the future, I find it more robust to (additionally) cite a journal publication, if available.
