+++ paper/paper.bib
@@ -1,6 +1,6 @@
 @book{Forrester:1971,
 author = { {Forrester}, Jay W. },
-title = { World dynamics [by] Jay W. Forrester },
+title = { World dynamics },
 publisher = { Wright-Allen Press Cambridge, Mass },
 pages = { xiii, 142 p. },
 year = { 1971 },
