### 2022-05-20 JamesPHoughton/pysd#316 @enekomartinmartinez
> In the bib entry of Forrester1971, I believe the [by] jay w. forrester can be removed? If not, the name should be made title case in the bib file.

Corrected

> It is a bit unfortunate that the "co-first author" footnote symbols are different and the footnote are repeated. I don't know if JOSS's markdown parser (which I believe is pandoc) supports this, but it may be worth trying if [this syntax](https://www.markdownguide.org/extended-syntax/#footnotes) is working? It would probably mean adding something like [^*] to both author names in the table (* becoming the symbol), and adding [^*]: co-first author at the beginning of the text? If that does not work, I would propose to inquire regarding the best practice of denoting the authorship status. What do you think?

The [example given in the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) do as current version of the paper. I didn't manage to find a workaround to that, I tried your suggestion but it didn't work. I know that is possible to have a workaround in latex with `cleveref` package and something like:
```
Hello\footnote{\label{fnote:my_note}My foot note.}

World\footnoteref{fnote:my_note}
```
But I don't know if there is a way to have the same workaround with pandoc.
