### 2022-05-20 JamesPHoughton/pysd#316 @blsqr
> The [example given in the JOSS documentation](https://joss.readthedocs.io/en/latest/submitting.html#example-paper-and-bibliography) do as current version of the paper

I missed that example, sorry. In that case, let's just leave it as it is. Thanks for looking into it!
