+++ docs/whats_new.rst
@@ -2,8 +2,8 @@
 What's New
 ==========
 
-v3.0.0 (unreleased)
------------------------
+v3.0.0 (23/05/2022)
+-------------------
 
 New Features
 ~~~~~~~~~~~~
