### 2022-05-23 JamesPHoughton/pysd#321 @blsqr
[JOSS Review] Add citation.cff file
This is a suggestion that can probably only be addressed once the JOSS review has finished and the paper is published – or once the DOI is fixed. Perhaps it can be prepared on the `paper_pysd` branch, though?

The [`citation.cff` file](https://citation-file-format.github.io) would provide citation information in a format that is becoming more and more wide-spread. Most importantly, GitHub has recently added support and will display that information in the project's sidebar:
