### 2022-05-23 #4329 @blsqr
Just as a brief update from my side: I am almost finished with the review and everything is looking really good. Given that the paper will be about v3.0, I'd like to wait for its official release and give it another go before concluding the review.
