### 2022-05-23 872a62e
Simplify subscripts dictionaries for TabData objects
