### 2022-05-24 JamesPHoughton/pysd#322 @blsqr
[JOSS Review] Python 3.10 support
Python v3.10 is not mentioned as being supported in `setup.py`, despite tests passing (at least for me locally) – have you considered updating the CI accordingly and then adding that classifier in `setup.py`?
