### 2022-05-24 JamesPHoughton/pysd#323 @blsqr
[JOSS Review] Warnings from tests
The PySD test suite seems very thorough to me, which is great for further development. However, the tests emit a number of warnings, some of which may soon require your attention; see below.

Some of these warnings are from PySD itself; in such a case, a [pytest warnings capture](https://docs.pytest.org/en/6.2.x/warnings.html#warns) may be helpful, i.e.: asserting the warning.
Another one is from the use of [`load_module`](https://github.com/JamesPHoughton/pysd/search?q=load_module), which will apparently be removed in Python 3.12 (to be released Oct. 2023 – still some time off).

For me, this is not an issue that blocks us going forward with the JOSS review. I mainly wanted to bring this to your attention and enquire about procedures or automations you may have in place to address upcoming warnings like these? One such option would be to configure the CI jobs to handle warnings as errors, such that you get notified about new warnings and prompted to act on them. This would be ideal for a workflow that relies on GitHub PRs, where these warnings would then be shown.

---
```
<frozen importlib._bootstrap>:283: 1 warning
unit_test_benchmarking.py: 2 warnings
unit_test_cli.py: 4 warnings
unit_test_components.py: 2 warnings
unit_test_pysd.py: 77 warnings
unit_test_statefuls.py: 2 warnings
pytest_pysd/pytest_errors.py: 3 warnings
pytest_pysd/pytest_model_attributes.py: 4 warnings
pytest_pysd/pytest_random.py: 1 warning
pytest_pysd/pytest_select_submodel.py: 14 warnings
pytest_types/data/pytest_data_with_model.py: 11 warnings
  <frozen importlib._bootstrap>:283: DeprecationWarning: the load_module() method is deprecated and slated for removal in Python 3.12; use exec_module() instead

pytest_translators/pytest_vensim.py:38
  /Users/yunus/Documents/Science/Coding/pysd/tests/pytest_translators/pytest_vensim.py:38: PytestUnknownMarkWarning: Unknown pytest.mark.dependency - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.dependency(name="read_vensim_file")

pytest_translators/pytest_vensim.py:51
  /Users/yunus/Documents/Science/Coding/pysd/tests/pytest_translators/pytest_vensim.py:51: PytestUnknownMarkWarning: Unknown pytest.mark.dependency - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.dependency(depends=["read_vensim_file"])

unit_test_cli.py::TestPySD::test_run_model_with_data
  /Users/yunus/Documents/Science/Coding/pysd/pysd/tools/benchmarking.py:299: SettingWithCopyWarning:
  A value is trying to be set on a copy of a slice from a DataFrame

  See the caveats in the documentation: https://pandas.pydata.org/pandas-docs/stable/user_guide/indexing.html#returning-a-view-versus-a-copy
    df.loc[:, nan_cols] = df.loc[:, nan_cols].iloc[0].values

unit_test_pysd.py: 182 warnings
unit_test_utils.py: 1 warning
  /Users/yunus/.virtualenvs/pysd/lib/python3.10/site-packages/pandas/core/internals/blocks.py:938: VisibleDeprecationWarning: Creating an ndarray from ragged nested sequences (which is a list-or-tuple of lists-or-tuples-or ndarrays with different lengths or shapes) is deprecated. If you meant to do this, you must specify 'dtype=object' when creating the ndarray.
    arr_value = np.asarray(value)

unit_test_pysd.py::TestPySD::test_run_progress
unit_test_utils.py::TestProgressbar::test_progressbar
  /Users/yunus/.virtualenvs/pysd/lib/python3.10/site-packages/progressbar/bar.py:294: DeprecationWarning: The usage of `maxval` is deprecated, please use `max_value` instead
    warnings.warn('The usage of `maxval` is deprecated, please use '

unit_test_utils.py::TestLoadOutputs::test_load_columns
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_get_data_and_run[transposed_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_get_data_and_run[transposed_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_get_data_and_run[transposed_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_get_data_and_run[dict_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_modify_data[transposed_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_modify_data[transposed_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_modify_data[transposed_file]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_modify_data[dict_file]
  /Users/yunus/Documents/Science/Coding/pysd/pysd/py_backend/utils.py:431: FutureWarning: Passing a set as an indexer is deprecated and will raise in a future version. Use a list instead.
    out = out[columns]

pytest_types/data/pytest_columns.py: 8 warnings
pytest_types/data/pytest_data_with_model.py: 3 warnings
  /Users/yunus/Documents/Science/Coding/pysd/pysd/py_backend/data.py:67: DeprecationWarning: Sampling from a set deprecated
  since Python 3.9 and will be removed in a subsequent version.
    [float(col) for col in random.sample(out, min(3, len(out)))]

pytest_types/data/pytest_columns.py::TestColumns::test_transposed_frame
pytest_types/data/pytest_columns.py::TestColumns::test_get_columns
pytest_types/data/pytest_columns.py::TestColumnsErrors::test_columns_errors[invalid_file_format]
pytest_types/data/pytest_data_with_model.py::TestPySDData::test_get_data_and_run[transposed_file]
  /Users/yunus/Documents/Science/Coding/pysd/pysd/py_backend/data.py:74: DeprecationWarning: Sampling from a set deprecated
  since Python 3.9 and will be removed in a subsequent version.
    [float(col) for col in random.sample(out, min(3, len(out)))]
```
