### 2022-05-24 JamesPHoughton/pysd#324 @blsqr
[JOSS Review] CI Automation
A topic adjacent to #323 would be that of further CI automations.

This could be useful for code formatting, for instance. I saw that `black` is one of your requirements although I read nothing about black specifically in the docs. You do mention pep8 formatting in the Development Guidelines, but I see no place where code formatting is *enforced*.
Have you considered introducing something like [pre-commit](https://pre-commit.com) to automate code formatting on the user side? This could also become part of a GitHub action, leading to a failing job in case a developer did not properly apply code formatting.

To be clear: I absolutely do not want to push you towards using these tools and automations, just want to bring up the topic. I am happy to discuss benefits and disadvantages, if you are interested.
