### 2022-05-24 #4329 @blsqr
Thanks for the update, @enekomartinmartinez!

I wanted to give an overview of the issues I opened in the [PySD project](https://github.com/JamesPHoughton/pysd/issues):
- [x] [Questions and suggestions regarding the software paper](https://github.com/JamesPHoughton/pysd/issues/316)
- [x] [Documentation enhancements](https://github.com/JamesPHoughton/pysd/issues/319)
- [ ] [Python 3.10 support](https://github.com/JamesPHoughton/pysd/issues/322)
- [ ] [Address warnings](https://github.com/JamesPHoughton/pysd/issues/323)
- [ ] [Consider using CI automations](https://github.com/JamesPHoughton/pysd/issues/324) *(suggestion)*
- [ ] [Consider adding a citation.cff file](https://github.com/JamesPHoughton/pysd/issues/321) *(suggestion)*

@pdebuyl Contingent on the above issues being addressed, I would conclude my review and **strongly support publication of PySD in JOSS**: The PySD package provides highly valuable functionality to the System Dynamics community by creating a bridge from domain-specific languages to the Python ecosystem. In addition to translating models, the recent changes using the Abstract Model Representation open an interesting development pathway.
Furthermore, the package appears matured and is well documented, including many examples, further reading material and cook books.
