### 2022-05-24 JamesPHoughton/pysd#322 @enekomartinmartinez
HI @blsqr,
I think when I tried I could make GitHub Actions run with Python 3.10 because it was still not available there. Therefore, I did not include it in the supported list because I didn't want to include something that is not tested. The main idea with the Python versions was to test only the lower and the higher supported versions.
I will try configuring GH Actions again to run with Python 3.10 or 3.11 if available.
