### 2022-05-24 JamesPHoughton/pysd#324 @enekomartinmartinez
Hi @blsqr,
Thanks for pointing that out. We use _black_ for formatting the Python output from translated models, I will mention that in the Python building section.
We currently use [PEP 8 Speaks](https://github.com/OrkoHunter/pep8speaks) in PRs, which automatically checks the style in the modified lines of code and publishes a message in the PR. But, having some kind of test that may fail could be also interesting. I will consider it.
