### 2022-05-24 JamesPHoughton/pysd#319 @blsqr
> If we understand by being feature complete supporting all the Vensim and Xmile functions …

Fair enough, I didn't think about that! But maybe that raises an interesting point: the alpha/beta/… status may not be that informative overall, because it's unclear what it relates to for PySD. In my perception, PySD's public interface looks rather stable and it is feature complete regarding *many* intended use cases. For me that would already warrant a "release" or "production" status, even if there are still some further features and translation functions to be implemented.

At the same time, I see your point: It should be clear that not all Vensim and Xmile functions are available yet. However, would that not be better communicated by explicitly having something like a comparison table that shows which functions are available and which still need to be implemented? An alpha or beta status does not really convey that, in my opinion.

Either way: I think you have to feel happy with the release status and I can understand if you'd want to keep it as alpha or beta 👍
