### 2022-05-24 JamesPHoughton/pysd#322 @blsqr
Sounds good! And I agree that it should be part of the test pipeline before the classifier is added.

> test only the lower and the higher supported versions.

That's typically a robust approach, I'd say 👍
