### 2022-05-24 JamesPHoughton/pysd#324 @blsqr
Thanks, @enekomartinmartinez!

> We currently use [PEP 8 Speaks](https://github.com/OrkoHunter/pep8speaks) in PRs, which automatically checks the style in the modified lines of code and publishes a message in the PR.

That's a good solution as well! I'm typically working on GitLab where these kinds of tasks are often handled via CI jobs, but the PEP 8 Speaks bot seems to do task just as well, so there's probably no need for a test.
