### 2022-05-24 JamesPHoughton/pysd#322 @blsqr
Sure! :)
