### 2022-05-24 JamesPHoughton/pysd#324 @enekomartinmartinez
Yes, PEP 8 Speaks doesn't work on GitLab, I will consider adding the test for some other repos I have there.
