### 2022-05-24 JamesPHoughton/pysd#319 @enekomartinmartinez
@JamesPHoughton have you been able to change the configuration in RTD? And what do you think about moving the release status?
