### 2022-05-24 a98cd7e
Remove some warnings and some other improvements
