+++ setup.py
@@ -11,7 +11,7 @@ setup(
     author_email='james.p.houghton@gmail.com',
     packages=find_packages(exclude=['docs', 'tests', 'dist', 'build']),
     url='https://github.com/JamesPHoughton/pysd',
-    license='LICENSE.txt',
+    license='LICENSE',
     description='System Dynamics Modeling in Python',
     long_description=open('README.md').read(),
     long_description_content_type='text/markdown',
@@ -26,6 +26,7 @@ setup(
         'Programming Language :: Python :: 3.7',
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
+        'Programming Language :: Python :: 3.10',
     ],
     install_requires=open('requirements.txt').read().strip().split('\n'),
     package_data={
