+++ docs/whats_new.rst
@@ -20,12 +20,17 @@ Bug fixes
 
 Documentation
 ~~~~~~~~~~~~~
+- Improve tests/README.md.
+- Minor improvements in the documentation.
 
 Performance
 ~~~~~~~~~~~
 
 Internal Changes
 ~~~~~~~~~~~~~~~~
+- Add Python 3.10 to CI pipeline and include it in the supported versions list.
+- Correct LICENSE file extension in the `setup.py`.
+- Move from `importlib`'s :py:func:`load_module` to :py:func:`exec_module`.
 
 
 v3.0.0 (2022/05/23)
