+++ tests/pytest.ini
@@ -1,2 +1,5 @@
 [pytest]
 python_files = unit_test_*.py pytest_*/**/*.py pytest_*/*.py
+filterwarnings =
+    error::DeprecationWarning
+    error::FutureWarning
