### 2022-05-25 JamesPHoughton/pysd#319 @JamesPHoughton
Hey @enekomartinmartinez - I was able to manually set up some active versions in the RTD, but didn't add every release, as there are so many. Here's what's currently being shown:
![image](https://user-images.githubusercontent.com/4304478/170155069-3d543355-33b5-4d77-b492-bcf011b4cc18.png)
Do you think that's a reasonable set?
