### 2022-05-25 JamesPHoughton/pysd#319 @JamesPHoughton
>Fair enough, I didn't think about that! But maybe that raises an interesting point: the alpha/beta/… status may not be that informative overall, because it's unclear what it relates to for PySD. In my perception, PySD's public interface looks rather stable and it is feature complete regarding many intended use cases. For me that would already warrant a "release" or "production" status, even if there are still some further features and translation functions to be implemented.
>
>At the same time, I see your point: It should be clear that not all Vensim and Xmile functions are available yet. However, would that not be better communicated by explicitly having something like a comparison table that shows which functions are available and which still need to be implemented? An alpha or beta status does not really convey that, in my opinion.

I don't personally think we need to tie our alpha/beta status to feature completeness - the feature set of Vensim is notoriously bloated, and I don't think we can or even should aim to support all of Vensim's functionality. From my perspective, the alpha/beta/production status signals to the user more about how ready the software is for the average user from a reliability standpoint, and how involved we expect them to be in the bugfinding process! I'm more than happy to move to 'beta' at this point to signal the maturity that @enekomartinmartinez has achieved in the last 3(!!!) major releases. =)
