### 2022-05-25 e9907c5
Remove warnings from tests and move to pytest backend for all
