+++ tests/pytest_tools/pytest_benchmarking.py
@@ -1,0 +1,135 @@
+import pytest
+
+import pandas as pd
+
+from pysd.tools.benchmarking import runner, load_outputs, assert_frames_close
+
+# most of the features of this script are already tested indirectly when
+# running vensim and xmile integration tests
+
+
+class TestErrors():
+
+    def test_canonical_file_not_found(self, _root):
+        error_message = 'Canonical output file not found.'
+        with pytest.raises(FileNotFoundError, match=error_message):
+            runner(_root.joinpath("more-tests/not_existent.mdl"))
+
+    def test_non_valid_model(self, _root):
+        error_message = r"The model file name must be a Vensim \(\.mdl\),"\
+            r" a Xmile \(\.xmile, \.xml, \.stmx\) or a PySD \(\.py\) "\
+            r"model file\.\.\.$"
+        with pytest.raises(ValueError, match=error_message):
+            runner(_root.joinpath("more-tests/not_vensim/test_not_vensim.txt"))
+
+    def test_different_frames_error(self, _root):
+
+        error_message = r"\nFollowing columns are not close:\n\t"\
+            r"Teacup Temperature\n\nFirst false values \(30\.0\):\n\t"\
+            r"Teacup Temperature$"
+
+        with pytest.raises(AssertionError, match=error_message):
+            assert_frames_close(
+                load_outputs(_root.joinpath("data/out_teacup.csv")),
+                load_outputs(_root.joinpath("data/out_teacup_modified.csv")))
+
+        error_message = r"\nFollowing columns are not close:\n\tTeacup "\
+            r"Temperature\n\nFirst false values \(30\.0\):\n\tTeacup "\
+            r"Temperature\n\nColumn 'Teacup Temperature' is not close\."\
+            r"\n\nExpected values:\n\t\[[\-0-9\s\n,\.]*\]"\
+            r"\n\nActual values:\n\t\[[\-0-9\s\n,\.]*\]"\
+            r"\n\nDifference:\n\t\[[\-0-9\s\n,\.]*\]$"
+
+        with pytest.raises(AssertionError, match=error_message):
+            assert_frames_close(
+                load_outputs(_root.joinpath("data/out_teacup.csv")),
+                load_outputs(_root.joinpath("data/out_teacup_modified.csv")),
+                verbose=True)
+
+    def test_different_frames_warning(self, _root):
+
+        warn_message = r"\nFollowing columns are not close:\n\t"\
+            r"Teacup Temperature\n\nFirst false values \(30\.0\):\n\t"\
+            r"Teacup Temperature$"
+        with pytest.warns(UserWarning, match=warn_message):
+            assert_frames_close(
+                load_outputs(_root.joinpath("data/out_teacup.csv")),
+                load_outputs(_root.joinpath("data/out_teacup_modified.csv")),
+                assertion="warn")
+
+        warn_message = r"\nFollowing columns are not close:\n\tTeacup "\
+            r"Temperature\n\nFirst false values \(30\.0\):\n\tTeacup "\
+            r"Temperature\n\nColumn 'Teacup Temperature' is not close\."\
+            r"\n\nExpected values:\n\t\[[\-0-9\s\n,\.]*\]"\
+            r"\n\nActual values:\n\t\[[\-0-9\s\n,\.]*\]"\
+            r"\n\nDifference:\n\t\[[\-0-9\s\n,\.]*\]$"
+
+        with pytest.warns(UserWarning, match=warn_message):
+            assert_frames_close(
+                load_outputs(_root.joinpath("data/out_teacup.csv")),
+                load_outputs(_root.joinpath("data/out_teacup_modified.csv")),
+                assertion="warn", verbose=True)
+
[…]
