+++ tests/pytest_pysd/pytest_cli.py
@@ -1,8 +1,10 @@
 import sys
 import os
 import shutil
-import unittest
 import subprocess
+from pathlib import Path
+
+import pytest
 
 import pandas as pd
 import numpy as np
@@ -10,7 +12,11 @@ import numpy as np
 from pysd.tools.benchmarking import load_outputs, assert_frames_close
 from pysd import __version__
 
-_root = os.path.dirname(__file__)
+# TODO replace test paths by fixtures and translate and run the models
+# in temporal directories
+# TODO make CLI tests run on Windows
+
+_root = Path(__file__).parent.parent
 
 test_model = os.path.join(_root, "test-models/samples/teacup/teacup.mdl")
 test_model_xmile = os.path.join(
@@ -54,8 +60,11 @@ def split_bash(string):
     return spl
 
 
-@unittest.skipIf(sys.platform.startswith("win"), "not working on Windows")
-class TestPySD(unittest.TestCase):
+@pytest.mark.skipif(
+    sys.platform.startswith("win"),
+    reason="not working on Windows"
+)
+class TestPySD():
     """ These tests are similar to unit_test_pysd but adapted for cli """
     def test_read_not_model(self):
 
@@ -64,11 +73,10 @@ class TestPySD(unittest.TestCase):
         command = f"{call} {model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {model}", stderr)
-        self.assertIn(
-            "The model file name must be a Vensim (.mdl), a Xmile (.xmile, "
-            ".xml, .stmx) or a PySD (.py) model file...", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {model}" in stderr
+        assert "The model file name must be a Vensim (.mdl), a Xmile "\
+            "(.xmile, .xml, .stmx) or a PySD (.py) model file..." in stderr
 
     def test_read_model_not_exists(self):
 
@@ -77,10 +85,9 @@ class TestPySD(unittest.TestCase):
         command = f"{call} {model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {model}", stderr)
-        self.assertIn(
-            "The model file does not exist...", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {model}" in stderr
+        assert "The model file does not exist..." in stderr
 
     def test_read_not_valid_output(self):
 
@@ -88,10 +95,9 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -o {out_xls_file} {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
-        self.assertNotEqual(out.returncode, 0)
-        self.assertIn(f"PySD: error: when parsing {out_xls_file}", stderr)
-        self.assertIn(
-            "The output file name must be .tab or .csv...", stderr)
+        assert out.returncode != 0
+        assert f"PySD: error: when parsing {out_xls_file}" in stderr
+        assert "The output file name must be .tab or .csv..." in stderr
 
     def test_read_not_valid_time_stamps(self):
 
@@ -99,19 +105,19 @@ class TestPySD(unittest.TestCase):
         command = f"{call} -R '{time_stamps}' {test_model}"
         out = subprocess.run(split_bash(command), capture_output=True)
[…]
