+++ docs/whats_new.rst
@@ -32,6 +32,8 @@ Internal Changes
 - Correct LICENSE file extension in the `setup.py`.
 - Move from `importlib`'s :py:func:`load_module` to :py:func:`exec_module`.
 - Remove warnings related to :py:data:`set` usage.
+- Move all the missing test to :py:mod:`pytest`.
+- Remove warning messages from test and make test fail if there is any warning.
 
 
 v3.0.0 (2022/05/23)
