+++ tests/pytest.ini
@@ -1,5 +1,5 @@
 [pytest]
-python_files = unit_test_*.py pytest_*/**/*.py pytest_*/*.py
+python_files = pytest_*/**/*.py pytest_*/*.py
 filterwarnings =
-    error::DeprecationWarning
-    error::FutureWarning
+    error
+    ignore:Creating an ndarray from ragged nested sequences

[…]
