### 2022-05-26 JamesPHoughton/pysd#323 @enekomartinmartinez
Hi @blsqr,
Thank you for the recommendation. I removed all the warnings a long time ago but with the newer versions, new warnings appear. Therefore, I think is a very good idea to make the tests fail if there is any warning.
I have almost completed the objective of cleaning them in branch `dev`, which also contains the other suggestions.
I will merge this branch with I manage to make all the tests pass again (currently 2 tests are failing on windows).
