### 2022-05-26 JamesPHoughton/pysd#323 @blsqr
Hi @enekomartinmartinez, yep, that's a constant struggle :D I think automation really can help a lot in staying on top of those things.

> I will merge this branch with I manage to make all the tests pass again (currently 2 tests are failing on windows).

👍
