+++ tests/pytest_pysd/pytest_cli.py
@@ -1,41 +1,50 @@
 import sys
+import re
 import os
+from pathlib import Path
 import shutil
 import subprocess
-from pathlib import Path
 
 import pytest
 
 import pandas as pd
 import numpy as np
 
+import pysd
 from pysd.tools.benchmarking import load_outputs, assert_frames_close
-from pysd import __version__
 
-# TODO replace test paths by fixtures and translate and run the models
-# in temporal directories
-# TODO make CLI tests run on Windows
-
-_root = Path(__file__).parent.parent
-
-test_model = os.path.join(_root, "test-models/samples/teacup/teacup.mdl")
-test_model_xmile = os.path.join(
-    _root, "test-models/samples/teacup/teacup.xmile")
-test_model_subs = os.path.join(
-    _root,
-    "test-models/tests/subscript_2d_arrays/test_subscript_2d_arrays.mdl")
-test_model_look = os.path.join(
-    _root,
-    "test-models/tests/get_lookups_subscripted_args/"
-    + "test_get_lookups_subscripted_args.mdl")
-
-out_tab_file = os.path.join(_root, "cli_output.tab")
-out_csv_file = os.path.join(_root, "cli_output.csv")
+test_model_ven = "test-models/samples/teacup/teacup.mdl"
 
 encoding_stdout = sys.stdout.encoding or "utf-8"
 encoding_stderr = sys.stderr.encoding or "utf-8"
 
 call = "python -m pysd"
+
+
+@pytest.fixture
+def out_tab(tmp_path):
+    return tmp_path.joinpath("cli_output.tab")
+
+
+@pytest.fixture
+def out_csv(tmp_path):
+    return tmp_path.joinpath("cli_output.csv")
+
+
+@pytest.fixture
+def test_model(model, _root):
+    return _root.joinpath(model)
+
+
+@pytest.fixture
+def test_copy(tmp_path, test_model):
+    """
+    Copy test folder to a temporary folder therefore we avoid creating
+    PySD model files in the original folder
+    """
+    test_folder = tmp_path.joinpath(test_model.parent.name)
+    shutil.copytree(test_model.parent, test_folder)
+    return test_folder.joinpath(test_model.name)
 
 
 def split_bash(string):
@@ -60,16 +69,11 @@ def split_bash(string):
     return spl
 
 
-@pytest.mark.skipif(
-    sys.platform.startswith("win"),
-    reason="not working on Windows"
-)
 class TestPySD():
-    """ These tests are similar to unit_test_pysd but adapted for cli """
-    def test_read_not_model(self):
+    """ These tests are similar to pytest_pysd but adapted for cli """
+    def test_read_not_model(self, _root):
 
-        model = os.path.join(
-            _root, "more-tests/not_vensim/test_not_vensim.txt")
+        model = _root.joinpath("more-tests/not_vensim/test_not_vensim.txt")
         command = f"{call} {model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
@@ -78,10 +82,9 @@ class TestPySD():
         assert "The model file name must be a Vensim (.mdl), a Xmile "\
             "(.xmile, .xml, .stmx) or a PySD (.py) model file..." in stderr
 
-    def test_read_model_not_exists(self):
+    def test_read_model_not_exists(self, _root):
 
-        model = os.path.join(
-            _root, "more-tests/not_vensim/test_not_vensim.mdl")
+        model = _root.joinpath("more-tests/not_vensim/test_not_vensim.mdl")
         command = f"{call} {model}"
         out = subprocess.run(split_bash(command), capture_output=True)
         stderr = out.stderr.decode(encoding_stderr)
[…]
