### 2022-05-26 9f33528
Minor improvements in docs
