### 2022-05-26 #4329 @enekomartinmartinez
Thanks a lot @blsqr!
All the suggestions you proposed have been already included in the master branch and release 3.0.1. I have also merged master to paper_pysd. I will include the _citation.cff_ file in the paper_pysd branch soon.
