### 2022-05-26 JamesPHoughton/pysd#321 @enekomartinmartinez
I have updated the file to the paper_pysd branch, I am wondering how to show the journal in the final citation widget from GitHub. Currently, it doesn't appear.

@JamesPHoughton @rogersamso does that look good to you?

@blsqr do you know if there is a way to add multiple cite entries? I think it would be nice to also keep Zenodo's DOI somehow, that way all the contributors are also mentioned.
