### 2022-05-26 JamesPHoughton/pysd#321 @enekomartinmartinez
> I am wondering how to show the journal in the final citation widget from GitHub. Currently, it doesn't appear.
I managed to show the journal :)
