### 2022-05-26 JamesPHoughton/pysd#321 @blsqr
> multiple cite entries

The citation file focusses on a single citation, afaik, but the [`references` section of the file](https://github.com/citation-file-format/citation-file-format/blob/main/schema-guide.md#referencing-other-work) can be used to refer to other work.
This does put the references on a different level than the main work though ...

In your case I would propose to expand the "How to cite" section in the docs/readme to contain both citations and corresponding BibTeX data. What do you think?
