### 2022-05-27 JamesPHoughton/pysd#321 @enekomartinmartinez
Thanks a lot @blsqr
I have make the following changes in the paper_pysd branch:
- Remove how to cite from the README.md and added a link to the citation section in RTD (this way we only have the information in one place and is easier to update)
- Update the way to cite the introductory paper of PySD in the docs. I have added information about the "book" of the conference, otherwise, it seemed that the paper was not published, as no doi or isbn were provided. Let me know @JamesPHoughton if this works for you
- Ceate the citation for the new paper, the doi and url must be updated when published.
- Add BibTeX citations in the docs, to make it easier to export the citation.

I read that Zenodo updates the DOI from the citation.cff file once a release with this file is published. I don't know how exactly this works but will be fine if the Zenodo's badges can be updated with the paper's doi once published.
