### 2022-05-27 #4329 @rogersamso
Thanks a lot for your excellent review @blsqr , both the paper and the library have been greatly improved thanks to your comments and suggestions.

Looking forward to @andrecieplinski review to see if there's still room for further improvements!
