+++ docs/development/development_index.rst
@@ -6,9 +6,11 @@ Developer Documentation
 
    guidelines
    pathway
+   adding_functions
    pysd_architecture_views/4+1view_model
 
 In order to contribut to PySD check the :doc:`guidelines` and the :doc:`pathway`.
 You also will find helpful the :doc:`Structure of the PySD library <../../structure/structure_index>` to understand better how it works.
+If you want to add a missing function to the PySD Python builder you may find useful the example in :doc:`adding_functions`.
 
 
