### 2022-05-31 JamesPHoughton/pysd#316 @pdebuyl
Hi all. I read the discussion above and I am not sure that I understand fully the authors' choice regarding the version that serves as a reference.

I would find it unusual to refer to a version <3 if 3 is out. Although it is an important release, it seems to build on the same overall package organization, and it would make sense to consider version 3.0.1 the reviewed version. @blsqr and authors, what do you think?

PS: JOSS changed some of its infrastructure recently, hence the mismatch for the co-first author syntax. You can remove the footnotes and add a `equal-contrib: true` line below the co-first authors instead. Similarly, corresponding authors are indicated by a line `corresponding: true` below the authors. See the related issue that I opened at JOSS.
