### 2022-06-01 #4329 @pdebuyl
@blsqr thank you very much for the review! I commented in the issue here https://github.com/JamesPHoughton/pysd/issues/316 about the version of reference.
