### 2022-06-01 JamesPHoughton/pysd#316 @enekomartinmartinez
Hi @pdebuyl
Thanks for your suggestions, I will include them in the paper.

The current version of the paper is about 3.0.0, we updated it and noticed the reviewers take that into account. Version 3.0.1 includes the suggestion made by @blsqr during the revision process and a minor bug fix.
