+++ paper/paper.bib
@@ -111,11 +111,20 @@ doi = {10.5281/zenodo.5654824},
 url = {https://doi.org/10.5281/zenodo.5654824}
 }
 
-@inbook{houghton_siegel_2015,
-chapter = {Advanced data analytics for system dynamics models using PySD},
-title = {Proceedings of the  33rd International  Conference of the System Dynamics Society},
-author = {Houghton, James P and Siegel, Michael},
-year = {2015}
+@inproceedings{houghton_siegel_2015,
+    author = {Houghton, James P and Siegel, Michael},
+    booktitle = {{Proceedings of the 33rd International Conference of the System Dynamics Society}},
+    publisher = {{System Dynamics Society}},
+    title = {{Advanced data analytics for system dynamics models using PySD}},
+    url = {https://www.proceedings.com/28517.html},
+    isbn = {9781510815056},
+    volume = {2},
+    pages = {1436-1462},
+    eventdate = {2015-07-19/2015-07-23},
+    location = {Cambridge, Massachusetts, USA},
+    year = {2015},
+    month = {7},
+    keywords = {System Dynamics, Vensim, Python}
 }
 
 @article{FETENEADANE2019212,
