### 2022-06-01 #4329 @enekomartinmartinez
@pdebuyl @andrecieplinski with the new version you will be able to translate that model but still not able to run it as some functions that it uses are still missing in PySD.

As we discussed with @blsqr in [issue 319](https://github.com/JamesPHoughton/pysd/issues/319), some functions of Vensim and XMILE may be missing, that's why we updated the status of the library from _alpha_ to _beta_ and not to _production_. However, we will continue including these functions in the future or encourage PySD users to work on including them. In particular, the functions that are needed by @andrecieplinski's model will be included by me or my colleague @rogersamso as we need them for the project we are working on.

In any case, in the paper we focus in the structure and capabilities of PySD, which are completed. As we mention in line 52 we will continue improving the performance and support for Vensim and XMILE functions
> Future releases will focus on cutting down simulation times and on including additional built-in Stella® and Vensim® functions.
