### 2022-06-02 JamesPHoughton/pysd#316 @blsqr
> it would make sense to consider version 3.0.1 the reviewed version. @blsqr and authors, what do you think?

Would fully agree: some version 3.0.x should be considered the reviewed version.

The version probably has to be adjusted in the review issue to no longer say 2.x?

> new infrastructure

That's a very nice improvement! 👍
