### 2022-06-02 f01954b
Update release notes
