### 2022-06-02 08156b5
Add support for vector ordering functions and add examples of development to docs (#327)
* Add support for vector ordering functions

* Add examples of development to docs
