+++ .github/pull_request_template.md
@@ -1,0 +1,25 @@
+## Description
+
+(add the description of your changes here)
+
+## Related issues
+
+(add any related issues here)
+
+## Type of change
+
+(please delete options that are not relevant)
+
+- Bug fix (non-breaking change which fixes an issue)
+- New feature (non-breaking change which adds functionality)
+- Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- Other (specify)
+
+## PR verification (to be filled by reviewers)
+
+- [ ] The code follows the [PEP 8 style](https://peps.python.org/pep-0008/)
+- [ ] The new code has been tested properly for all the possible cases
+- [ ] The overall coverage has not dropped and other features have not been broken.
+- [ ] If new features have been added, they have been properly documented
+- [ ] *docs/whats_new.rst* has been updated
+- [ ] PySD version has been updated
