### 2022-06-03 e5f4427
Add support for GET TIME VALUE
