+++ docs/structure/vensim_translation.rst
@@ -117,6 +117,5 @@ Vensim macros are supported. The macro content between the keywords \:MACRO: and
 Planed New Functions and Features
 ---------------------------------
 - ALLOCATE BY PRIORITY
-- GET TIME VALUE
 - SHIFT IF TRUE
 - VECTOR SELECT
