### 2022-06-08 #4329 @pdebuyl
Thank you for the clarification. The limitation should be stated more explicitly in the article and the documentation. The current phrasing is "forward looking", which I can appreciate, but does not give a clear view on the code capabilities. This is not a criticism of the software. A "limitation" paragraph in the docs, for instance, could be one way to do this.
