### 2022-06-09 3d7d834
Add support for VECTOR SELECT
