### 2022-06-09 d380fd1
Add support for VECTOR SELECT
