### 2022-06-09 9d6b6fc
Add a paragraph about limitations to the documentation
