### 2022-06-09 #4329 @enekomartinmartinez
Thank you @pdebuyl for the suggestion, this could also help PySD users to understand why PySD doesn't work sometimes.

I have...

- Merged master (3.1.0) in paper_pysd. This includes some extra Vensim functions and a section with examples for adding new functions to the builder. See the [latest release notes](https://pysd.readthedocs.io/en/latest/whats_new.html).
- Included the paragraph you suggest in the [index of the documentation of paper_pysd](https://github.com/JamesPHoughton/pysd/blob/paper_pysd/docs/index.rst#limitations).

Let me know if you find this enough or if you prefer us to also include this information in the paper explicitly.
