### 2022-06-14 a09ccbf
Solve bug of defining subscripts with subranges
