### 2022-06-16 #4329 @rogersamso
> @andrecieplinski could you solve the issue you encountered? You can discuss the usage of the package here, this is fine.

Dear @pdebuyl for the lack of response to all direct mentions made in this thread, I am concerned @andrecieplinski is not receiving Github's notifications.
I was wondering if reviewers of JOSS have a time limit to submit their reviews?
Regards
