### 2022-06-16 34670ca
Add allocate by priority function
