+++ docs/python_api/functions.rst
@@ -5,6 +5,11 @@ Functions
 .. automodule:: pysd.py_backend.functions
    :members:
 
+Allocation functions
+--------------------
+.. automodule:: pysd.py_backend.allocation
+   :members:
+
 Statefuls
 ---------
 .. automodule:: pysd.py_backend.statefuls
