### 2022-06-20 55c6464
Add release date in docs
