### 2022-06-21 f80d662
Update status of SHIFT IF TRUE
