### 2022-06-22 e7bb0df
Improve error messages for External objects
