### 2022-06-22 a73da49
Add support for allocate available
