### 2022-06-23 0e11225
Add support for allocate available
