+++ pysd/py_backend/allocation.py
@@ -2,58 +2,221 @@
 The provided allocation functions have no direct analog in the standard
 Python data analytics stack. They are provided in a structure that makes
 it easy for the model elements to call. The functions may be similar to
-the original functions given by Vensim or Stella, but sometimes the
-number or order of arguments may change. The allocation functions may
-call a protected function or class method thatintegrates the algorithm
-to compute the allocation. The algorithms are briefly explained in these
+the original functions given by Vensim, but sometimes the number or
+order of arguments may change. The allocation functions may call a
+protected function or class method thatintegrates the algorithm to
+compute the allocation. The algorithms are briefly explained in these
 functions docstring.
+
+Note
+----
+The Allocation functions basis is explained in the Vensim documentation.
+https://www.vensim.com/documentation/allocation_overview.html
+
+Warning
+-------
+Some allocation function's results may differ from the result given by
+Vensim as optimization functions are used to solve the allocation
+problems. Those algorithms may not work in the same way or may
+have differences in the numerical error propagation.
+
 """
 import itertools
 from math import erfc
 
 import numpy as np
 import xarray as xr
-from scipy.optimize import fsolve
+from scipy.optimize import least_squares
+import portion as p
 
 
 class Priorities:
     @classmethod
-    def get_functions(cls, array):
-        return [cls.get_function(array[i]) for i in range(array.shape[0])]
+    def get_functions(cls, q0, pp, kind):
+        """
+        Get priority functions based on the demand/supply and priority profile.
+
+        Parameters
+        ----------
+        q0: numpy.array
+            values of maximum demand or supply of each component.
+            Its shape should be (n,)
+        pp: numpy.array
+            pp values array. Its shape should be (n, m).
+        kind: str ("demand" or "supply")
+            The kind of priority "demand" or "supply".
+
+        Returns
+        -------
+        functions: list of functions
+            List of allocation supply or demand function for each element.
+
+        full_allocation: function
+            Full allocation function. It is the result function of
+            addying all the functions.
+
+        def_intervals: list of tuples
+            List of (supply interval, priority interval, mean priority)
+            where the full_allocation function is extrictly monotonous
+            (injective). Givin a supply value, this is used to compute
+            the limits and starting point of the optimization problem.
+
+        """
+        if np.any(pp[:, 2] <= 0):
+            # pwidth values smaller than 0
+            raise ValueError("pwidth values must be positive.")
+
+        if kind == "demand":
+            # Get the list of priority functions and the intervals where
+            # they are strictly monotonous (injective function)
+            func_int = [
+                cls.get_function_demand(q0[i], pp[i])
+                for i in range(pp.shape[0])
+            ]
+            # In order to get the range of the full_allocation function,
+            # we need to flip the lower and the upper value, as it is a
+            # decreasing function for demand
+            int_attr = {"lower": "upper", "upper": "lower"}
+        elif kind == "supply":  # pragma: no cover
+            # Get the list of priority functions and the intervals where
+            # they are strictly monotonous (injective function)
+            func_int = [
+                cls.get_function_supply(q0[i], pp[i])
+                for i in range(pp.shape[0])
+            ]
+            # In order to get the range of the full_allocation function,
+            # we need to keep the lower and the upper value, as it is a
+            # increasing function for supply
+            int_attr = {"lower": "lower", "upper": "upper"}
+        else:
+            raise ValueError(
[…]
