+++ setup.py
@@ -27,6 +27,7 @@ setup(
         'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
+        'Programming Language :: Python :: 3.11',
     ],
     install_requires=open('requirements.txt').read().strip().split('\n'),
     package_data={
