### 2022-06-28 3eb1970
Recover python 3.10 as newest supported
