+++ .github/workflows/ci.yml
@@ -10,7 +10,7 @@ jobs:
     strategy:
       matrix:
         os: [ubuntu-latest, windows-latest]
-        python-version: ['3.7', '3.11.0-alpha.3']
+        python-version: ['3.7', '3.10']
 
     steps:
       - uses: actions/checkout@v3
