### 2022-07-04 a22979f
Add support for ALLOCATE AVAILABLE
Add support for Vensim's ALLOCATE AVAILABLE
