### 2022-07-05 78752c1
Rename SectionElementsParser to SectionElementsVisitor
