### 2022-07-08 #4329 @enekomartinmartinez
Hi @pdebuyl
I was wondering if you have any updates about the revision status.
I am concerned about the summer break causing additional delays to the review.
Thank you!
