### 2022-07-13 0ee4580
Initial implementation of stan translator
