### 2022-07-15 4d1f3bc
Solve bug about coordinates order in GET calls
