### 2022-07-15 98ca9ab
Solve bug about coordinates order in GET calls
