### 2022-07-15 3eb4989
Add support for subscripted arguments in step and ramp
