### 2022-07-18 3ec67c6
Fix bug in request 0 values in allocate by priority
