### 2022-07-18 cf5f25f
Fix bug in request 0 values in allocate by priority
