### 2022-07-19 27d23fe
Fix a numerical error in starting time of step and ramp
