+++ tests/pytest_pysd/pytest_functions.py
@@ -12,15 +12,22 @@ from pysd.py_backend.functions import\
 
 class TestInputFunctions():
     def test_ramp(self):
-        assert ramp(lambda: 14, .5, 10, 18) == 2
+        t = Time()
+        t.set_control_vars(time_step=1)
 
-        assert ramp(lambda: 4, .5, 10, 18) == 0
+        t.update(4)
+        assert ramp(t, .5, 10, 18) == 0
 
-        assert ramp(lambda: 24, .5, 10, 18) == 4
+        t.update(14)
+        assert ramp(t, .5, 10, 18) == 2
 
-        assert ramp(lambda: 24, .5, 10) == 7
+        t.update(24)
+        assert ramp(t, .5, 10, 18) == 4
 
-        assert ramp(lambda: 50, .5, 10) == 20
+        assert ramp(t, .5, 10) == 7
+
+        t.update(50)
+        assert ramp(t, .5, 10) == 20
 
         # arrays start
         coords = {"dim1": ["A", "B"], "dim2": ["X", "Y", "Z"]}
@@ -29,16 +36,19 @@ class TestInputFunctions():
             coords
         )
 
-        assert ramp(lambda: 14, .5, start_array, 18).equals(
-            xr.DataArray([[3., 3.5, 2.5], [5., 4.5, 4.]], coords))
-
-        assert ramp(lambda: 4, .5, start_array, 18).equals(
+        t.update(4)
+        assert ramp(t, .5, start_array, 18).equals(
             xr.DataArray([[0., 0., 0.], [0., 0., 0.]], coords))
 
-        assert ramp(lambda: 24, .5, start_array, 18).equals(
+        t.update(14)
+        assert ramp(t, .5, start_array, 18).equals(
+            xr.DataArray([[3., 3.5, 2.5], [5., 4.5, 4.]], coords))
+
+        t.update(24)
+        assert ramp(t, .5, start_array, 18).equals(
             xr.DataArray([[5., 5.5, 4.5], [7., 6.5, 6.]], coords))
 
-        assert ramp(lambda: 24, .5, start_array).equals(
+        assert ramp(t, .5, start_array).equals(
             xr.DataArray([[8., 8.5, 7.5], [10., 9.5, 9.]], coords))
 
         # arrays finish
@@ -46,10 +56,13 @@ class TestInputFunctions():
             [[10., 15., 18.], [20., 16., 12.]],
             coords
         )
-        assert ramp(lambda: 14, .5, 10, finish_array).equals(
+
+        t.update(14)
+        assert ramp(t, .5, 10, finish_array).equals(
             xr.DataArray([[0., 2., 2.], [2., 2., 1.]], coords))
 
-        assert ramp(lambda: 25, .5, 10, finish_array).equals(
+        t.update(25)
+        assert ramp(t, .5, 10, finish_array).equals(
             xr.DataArray([[0., 2.5, 4.], [5., 3., 1.]], coords))
 
         # arrays slope
@@ -57,37 +70,52 @@ class TestInputFunctions():
             [[0.5, .1, 1.], [2., -1., -0.5]],
             coords
         )
-        assert ramp(lambda: 14, slope_array, 10, 18).equals(
-            xr.DataArray([[2., .4, 4.], [8., -4, -2.]], coords))
 
-        assert ramp(lambda: 4, slope_array, 10, 18).equals(
+        t.update(4)
+        assert ramp(t, slope_array, 10, 18).equals(
             xr.DataArray([[0., 0., 0.], [0., 0., 0.]], coords))
 
-        assert ramp(lambda: 24, slope_array, 10, 18).equals(
+        t.update(14)
[…]
