### 2022-07-19 dbf3ea3
Fix a numerical error in starting time of step and ramp
