+++ test_scripts/testing.py
@@ -2,16 +2,15 @@ from pysd.translators.vensim.vensim_file import VensimFile
 from pysd.translators.xmile.xmile_file import XmileFile
 from pysd.builders.stan.stan_model_builder import *
 
-
-vf = VensimFile("vensim_models/Inventory.mdl")
-#vf = XmileFile("vensim_models/repair.xmile")
+vf = VensimFile("test_scripts/vensim_models/Inventory.mdl")
+#vf = VensimFile("test_scripts/vensim_models/repair.mdl")
 vf.parse()
 
 am = vf.get_abstract_model()
 
 stan_builder = StanModelBuilder(am)
-# print(stan_builder.build_function_block(["failure_count", "repair_time"], ["battle_field", "repair_shop"])) # repair
-print(stan_builder.create_stan_program(["demand"], ["inventory", "backlog"]))
+#print(stan_builder.create_stan_program([(int, "failure_count"), "repair_time"], ["battle_field", "repair_shop"])) # repair
+#stan_builder.create_stan_program(["demand"], ["inventory", "backlog"])
 
 # for section in am.sections:
 #     for element in section.elements:
