### 2022-07-21 6246c21
change model input and output signature
will document on sub-types of input and output
