### 2022-07-24 b651530
Add LOOKUP implementation. Refactor AST walkers to classes
