+++ pysd/builders/stan/stan_model_builder.py
@@ -1,40 +1,34 @@
 import os
 from pathlib import Path
-from typing import Union, List, Dict, Set, Iterable
+from typing import Union, List, Dict, Set, Iterable, Type
 
 from .ast_walker import *
-
+from .utilities import *
 from pysd.translators.structures.abstract_model import\
     AbstractComponent, AbstractElement, AbstractModel, AbstractSection
-
-
-class IndentedString:
-    def __init__(self, indent_level=0):
-        self.indent_level = indent_level
-        self.string = " " * 4 * self.indent_level
-
-    def __iadd__(self, other: str):
-        prefix = " " * 4 * self.indent_level
-        if other != "\n":
-            self.string += prefix
-        self.string += other
-        return self
-
-    def __str__(self):
-        return self.string
-
-
-def name_to_identifier(name: str):
-    return name.lower().replace(" ", "_")
 
 
 class StanModelBuilder:
     def __init__(self, abstract_model: AbstractModel):
         self.abstract_model = abstract_model
 
-    def create_stan_program(self, predictor_variable_names, outcome_variable_names, function_name="vensim_func"):
-        self.code = IndentedString()
 
+    def create_stan_program(self, predictor_variable_names: List[Union[str, Tuple[str, str]]], outcome_variable_names: List[str], function_name="vensim_func"):
+        # Santize vensim names to stan-compliant identifiers
+        sanitized_predictor_variable_names = []
+        for var in predictor_variable_names:
+            match var:
+                case str(x):
+                    sanitized_predictor_variable_names.append(name_to_identifier(x))
+                case (str(type), str(var_name)):
+                    sanitized_predictor_variable_names.append((type, name_to_identifier(var_name)))
+                case _:
+                    raise Exception("predictor_variable_names must be a list of strings and/or a tuple of the form(T, Name), where T is a string denoting the variable's stan type and Name a string denoting the variable name")
+
+        predictor_variable_names = sanitized_predictor_variable_names
+        outcome_variable_names = [name_to_identifier(name) for name in outcome_variable_names]
+
+        self.code = IndentedString()
 
         self.code += StanFunctionBuilder(self.abstract_model).build_function_block(predictor_variable_names,
                                                                                    outcome_variable_names, function_name)
@@ -49,6 +43,26 @@ class StanModelBuilder:
         self.code += "generated quantities{\n}"
 
         return self.code
+
+    def print_variable_info(self):
+        var_names = []
+        max_length = len("original name") + 1
+        for element in self.abstract_model.sections[0].elements:
+            is_stock = False
+            for component in element.components:
+                if isinstance(component.ast, IntegStructure):
+                    is_stock = True
+                    break
+
+            var_names.append((element.name, name_to_identifier(element.name), is_stock))
+            max_length = max(max_length, len(element.name) + 1)
+
+        header = 'original name'.ljust(max_length) + "stan variable name".ljust(max_length) + "is stock"
+        print(header)
+        print("-" * len(header))
+        for x in var_names:
+            print(x[0].ljust(max_length) + x[1].ljust(max_length) + ("V" if x[2] else ""))
+
 """ class StanDataBuilder:
     def __init__(self, abstract_model: AbstractModel):
         self.abstract_model = abstract_model
@@ -65,6 +79,7 @@ class StanModelBuilder:
         self.code.indent_level -= 1
         self.code += "}\n" """
 
+
 class StanTransformedParametersBuilder:
     def __init__(self, abstract_model: AbstractModel):
         self.abstract_model = abstract_model
@@ -76,10 +91,8 @@ class StanTransformedParametersBuilder:
 
         argument_variables = []
         for var in predictor_variable_names:
-            print(var)
             match var:
                 case str(x):
-                    print("dfsdfsdfs")
                     argument_variables.append(x)
[…]

[…]
