### 2022-07-24 64eb140
Automatically identify stock variables and initial values
