+++ pysd/builders/stan/stan_model_builder.py
@@ -13,31 +13,34 @@ class StanModelBuilder:
         self.abstract_model = abstract_model
 
 
-    def create_stan_program(self, predictor_variable_names: List[Union[str, Tuple[str, str]]], outcome_variable_names: List[str], function_name="vensim_func"):
+    def create_stan_program(self, predictor_variable_names: List[Union[str, Tuple[str, str]]], function_name="vensim_func"):
         # Santize vensim names to stan-compliant identifiers
         sanitized_predictor_variable_names = []
         for var in predictor_variable_names:
             match var:
                 case str(x):
-                    sanitized_predictor_variable_names.append(name_to_identifier(x))
+                    sanitized_predictor_variable_names.append(vensim_name_to_identifier(x))
                 case (str(type), str(var_name)):
-                    sanitized_predictor_variable_names.append((type, name_to_identifier(var_name)))
+                    sanitized_predictor_variable_names.append((type, vensim_name_to_identifier(var_name)))
                 case _:
                     raise Exception("predictor_variable_names must be a list of strings and/or a tuple of the form(T, Name), where T is a string denoting the variable's stan type and Name a string denoting the variable name")
 
         predictor_variable_names = sanitized_predictor_variable_names
-        outcome_variable_names = [name_to_identifier(name) for name in outcome_variable_names]
+        outcome_variable_names = self.get_stock_variable_stan_names()
+        if not outcome_variable_names:
+            raise Exception("There are no stock variables defined in the model.")
 
         self.code = IndentedString()
 
-        self.code += StanFunctionBuilder(self.abstract_model).build_function_block(predictor_variable_names,
-                                                                                   outcome_variable_names, function_name)
+        function_block_builder = StanFunctionBuilder(self.abstract_model)
+
+        self.code += function_block_builder.build_function_block(predictor_variable_names, outcome_variable_names, function_name)
 
         self.code += "data{\n}\n"
         # self.code += StanDataBuilder(self.abstract_model).build_block(predictor_variable_names, outcome_variable_names)
         self.code += "transformed data{\n}\n"
         self.code += "parameters{\n}\n"
-        self.code += StanTransformedParametersBuilder(self.abstract_model).build_block(predictor_variable_names, outcome_variable_names, function_name)
+        self.code += StanTransformedParametersBuilder(self.abstract_model).build_block(predictor_variable_names, outcome_variable_names, function_block_builder.lookup_builder_walker.generated_lookup_function_names, function_name)
         self.code += "model{\n}\n"
 
         self.code += "generated quantities{\n}"
@@ -54,7 +57,7 @@ class StanModelBuilder:
                     is_stock = True
                     break
 
-            var_names.append((element.name, name_to_identifier(element.name), is_stock))
+            var_names.append((element.name, vensim_name_to_identifier(element.name), is_stock))
             max_length = max(max_length, len(element.name) + 1)
 
         header = 'original name'.ljust(max_length) + "stan variable name".ljust(max_length) + "is stock"
@@ -62,6 +65,22 @@ class StanModelBuilder:
         print("-" * len(header))
         for x in var_names:
             print(x[0].ljust(max_length) + x[1].ljust(max_length) + ("V" if x[2] else ""))
+
+    def get_stock_variable_stan_names(self) -> List[str]:
+        """
+        Iterate through the AST and find stock variables
+        Returns
+        -------
+
+        """
+        stock_varible_names = []
+        for element in self.abstract_model.sections[0].elements:
+            for component in element.components:
+                if isinstance(component.ast, IntegStructure):
+                    stock_varible_names.append(vensim_name_to_identifier(element.name))
+                    break
[…]
