### 2022-07-24 4f1e37e
WIP static RNG function
