### 2022-07-25 1a18cc5
Add a warning message when a subscript range is duplicated in a variable reference
