### 2022-07-25 9539518
Include new PySD logo
