### 2022-07-25 02fcbd3
Include new PySD logo
