+++ docs/whats_new.rst
@@ -1,7 +1,7 @@
 What's New
 ==========
 
-v3.5.0 (2022/07/19)
+v3.5.0 (2022/07/25)
 -------------------
 
 New Features
