### 2022-07-25 8aa50ac
Replace relative link to image by raw link to show the image on PyPI
