### 2022-07-26 #4329 @enekomartinmartinez
I have merged PySD 3.5.0 on paper_pysd branch. There are no changes on the paper, only bug fixes and support for new functions.
