### 2022-07-31 e7fa259
Allow non-default outcome variable passage
