### 2022-07-31 ce5bc65
Add float to auxillary name walker
