### 2022-07-31 f37c720
Update explanation on outcome signature
