+++ test_scripts/readme.md
@@ -3,7 +3,7 @@
 
 The main contribution of this PR is inference; its benefits are estimating parameter values and calibration. Currently pysd is doing a good job in data generation conditional on parameter values but lacks modules that retrieves parameter from the generated data. This pr aims to fill the gap; Stan is a computational Bayesian statistics package, with a large eco-system, providing state of the art inference algorithms including MCMC (HMC-NUTS) and variational inference. Given the following input from users, our output is posterior samples of estimated parameters.
 
-1. dynamic model in .mdl or .xmile or .stmx format (SQ1. can we support .stmx?)
+1. dynamic model in .mdl or .xmile or .stmx format
 
 2. user's classification of 
 
@@ -28,23 +28,8 @@ Our work is casting dynamic model into statistical model structure (JQ1. may I c
 
 Two works are planned for the near future and it would be tremendously helpful some support could be given in the second.
 
-- Developing case studies on workflow which includes the following:
-```
-1. prior predictive
-2. calibration 
-- posterior credible interval and SBC
-- sensitivity check on prior distribution and prior parameter 
-- compare posterior with prior
-3. posterior predictive
-```
 - Translating vensim and/or stella function using python [here](https://github.com/Dashadower/pysd/blob/master/pysd/builders/stan/ast_walker.py)
-Currently `integration`, `random generation`, `lookup` are completed (but need testing) and `smooth` is in progress. (SQ4. am I missing something? JQ2. are we missing any main structure?) 
+Currently `integration`, `random generation`, `lookup` are completed (but need testing) and `smooth` is in progress. Some parts we are missing: `if then else`, `step` (using `if then else`), `pulse`(point mass + divide by dt), `delay` (hard..), `material memory` (not possible in stan as it does not allow).
 
 ## 3. Our question
 May we ask what test or other functions would be need for pysd for our PR to be most helpful? Thanks for the awesome package!
-
----
-Q. I am making the notebook but is this necessary for pr? Could the pr and making notebook be proceeded parallelly?
-SQ5. Could you give us a code review as we can receive Jair's feedback? I am curious on 
-- the interface btw `ast_walker` and `stan_model_builder` 
-- the role of `RNGCodegenWalker`
