### 2022-08-03 SDXorg/pysd#348 @enekomartinmartinez
Move PySD to SDXorg
We want to move the PySD official repo to [SDXorg](https://github.com/SDXorg) organization. Nevertheless, GitHub takes care of migrating everything a redirecting the links we may need to take care of several things.

### **Expected migration date: mid-September**

**TODOs before migration**
- [x] Inform all current and past developers about the incoming change. Includying those who have projects that depend on PySD (use dependencies tree from the repo).
- [x] Check with JOSS pending review if this may cause any conflicts on the process.
- [x] Migrate PySD CookBook and  checkif there are conflicts with RTD

**TODOs after migration**
- [x] Update all the links in PySD to the new repo.
- [ ] Update external links that point to PySD (that we know about)
  - [x] https://pysd-cookbook.readthedocs.io/en/latest/
  - [ ] http://www.jamesphoughton.com/projects/
  - [x] https://systemdynamics.org/tools/useful-open-source-tools/
  - [x] https://github.com/SDXorg/SD-Tools
  - [x] coveralls and badge
- [x] Update all the links in libraries that depen on PySD to the new repo.
  - [x] excels2vensim
  - [x] sdqc
  - [x] pysd2r
  - [x]  SDVerificationToolKit
- [x] Check if secrets where properly migrated, this includes pip tokens for automatic publications of new releases.
- [x] Check if new releases are properly uploaded to pip.
- [x] Update the conda-forge recipe with the repo link.
- [x] Update permissions/access controls
- [x] Check if RTD works properly after migration and/or make neccesary updates.
- [x] Check if pep8 linter works properly in PR or include it again.
- [x] Check if coveralls works properly on pushes and PR.
- [x] Create an about the project section explaining the origin of the project mentioning all the main contributions, evolution and linking to the contributors page.
- [x] Create a table with "Who to contact" with the main developers information per "section" of PySD.
- [x] Update the license to extent it to 2022 and make a more general autorship.
- [x] Inform all current and past developers about the incoming change. Includying those who have projects that depend on PySD (use dependencies tree from the repo).
