### 2022-08-03 JamesPHoughton/pysd#348 @enekomartinmartinez
@JamesPHoughton are we moving as well the PySD cookbook?
