### 2022-08-03 JamesPHoughton/pysd#348 @JamesPHoughton
> @JamesPHoughton are we moving as well the PySD cookbook?

I think that would make sense. It could use some attention anyways, maybe it would be a good way to get that going. =)
