### 2022-08-03 #4329 @enekomartinmartinez
Hi @pdebuyl !
We are about to migrate PySD to [SDXorg organization](https://github.com/SDXorg), find the [issue he have created to track this process](https://github.com/JamesPHoughton/pysd/issues/348).

We wanted to ask you if this could make any conflict with the pending review or future publication process. GitHub is supposed to automatically redirect the links to the new repository. However, I am not sure if this will work properly with the links that the bot may be using to generate the paper or link the repo.

If any conflict happens, a potential fix could be to create a new review process from the new repo.

Please, let me know what do you think could happen and possible ways to avoid future conflicts. We will not migrate the repo until the second week of September when both @JamesPHoughton and I are back from vacation.

Enjoy the summer 🙂
