### 2022-08-03 JamesPHoughton/pysd#348 @enekomartinmartinez
I agree, we need to update the cookbook with the latest version. I think it could also be a good test to move the cookbook first so we can see how RDT reacts to moving the repo. Maybe we can move it before the end of this week so I can make some tests during next week?
