### 2022-08-06 JamesPHoughton/pysd#348 @JamesPHoughton
I think this is going to need to be after I get back from travel. I'll have some time the morning of the 7th or 8th of Sept. Should we work through it together?
