### 2022-08-08 JamesPHoughton/pysd#348 @enekomartinmartinez
@JamesPHoughton great.
For now both days work for me. But I will confirm it in the first week of September.
