### 2022-08-09 2ff8339
Improve performance of reading cellranges
