### 2022-08-09 JamesPHoughton/pysd#348 @enekomartinmartinez
About the paper review, I have found the following information:
https://joss.readthedocs.io/en/latest/editorial_bot.html?highlight=repository#changing-the-repository
