### 2022-08-10 181eff9
Fix bug of EXCEPT subranges
