### 2022-08-10 dbf96dd
Fix allocate_by_priority bug
