### 2022-08-10 d54c36b
Fix allocate_by_priority bug
