### 2022-08-11 1903312
Substitute match with isinstance.
