+++ pysd/builders/stan/ast_walker.py
@@ -1,4 +1,4 @@
-from typing import Union, List, Iterable, Dict, Tuple
+from typing import Union, List, Iterable, Dict, Tuple, Any
 from itertools import chain
 from dataclasses import dataclass, field
 from .utilities import IndentedString
@@ -15,21 +15,22 @@ class BaseNodeWaler:
 
 class AuxNameWalker(BaseNodeWaler):
     def walk(self, ast_node) -> List[str]:
-        match ast_node:
-            case int():
-                return []
-            case float():
-                return []
-            case ArithmeticStructure(operators, arguments):
-                return list(chain.from_iterable([self.walk(argument) for argument in arguments]))
-            case ReferenceStructure(reference, subscripts):
-                return [ast_node.reference]
-            case CallStructure(function, arguments):
-                return list(chain.from_iterable([self.walk(argument) for argument in arguments]))
-            case IntegStructure(flow, initial):
-                return self.walk(flow) + self.walk(initial)
-            case InlineLookupsStructure(argument, lookups):
-                return self.walk(lookups)
+        if isinstance(ast_node, int):
+            return []
+        elif isinstance(ast_node, float):
+            return []
+        elif isinstance(ast_node, ArithmeticStructure):
+            return list(chain.from_iterable([self.walk(argument) for argument in ast_node.arguments]))
+        elif isinstance(ast_node, ReferenceStructure):
+            return [ast_node.reference]
+        elif isinstance(ast_node, CallStructure):
+            return list(chain.from_iterable([self.walk(argument) for argument in ast_node.arguments]))
+        elif isinstance(ast_node, IntegStructure):
+            return self.walk(ast_node.flow) + self.walk(ast_node.initial)
+        elif isinstance(ast_node, InlineLookupsStructure):
+            return self.walk(ast_node.lookups)
+        else:
+            raise Exception(f"AST node of type {ast_node.__class__.__name__} is not supported.")
 
 @dataclass
 class LookupCodegenWalker(BaseNodeWaler):
@@ -44,45 +45,43 @@ class LookupCodegenWalker(BaseNodeWaler):
         return lookup_node.x + lookup_node.y + lookup_node.x_limits + lookup_node.y_limits
 
     def walk(self, ast_node) -> None:
-        match ast_node:
-            case InlineLookupsStructure(argument, lookups):
-                self.walk(lookups)
-            case LookupsStructure(x, y, x_limits, y_limits, type):
-                assert type == "interpolate", "Type of Lookup must be 'interpolate'"
-                identifier_key = LookupCodegenWalker.get_lookup_keyname(ast_node)
-                function_name = f"lookupFunc_{self.n_lookups}"
-                self.generated_lookup_function_names[identifier_key] = function_name
-                self.n_lookups += 1
-                self.code += f"real {function_name}(real x){{\n"
+        if isinstance(ast_node, InlineLookupsStructure):
+            self.walk(ast_node.lookups)
+        elif isinstance(ast_node, LookupsStructure):
+            assert ast_node.type == "interpolate", "Type of Lookup must be 'interpolate'"
+            identifier_key = LookupCodegenWalker.get_lookup_keyname(ast_node)
+            function_name = f"lookupFunc_{self.n_lookups}"
+            self.generated_lookup_function_names[identifier_key] = function_name
+            self.n_lookups += 1
+            self.code += f"real {function_name}(real x){{\n"
+            self.code.indent_level += 1
+            # Enter function body
+            self.code += f"# x {ast_node.x_limits} = {ast_node.x}\n"
+            self.code += f"# y {ast_node.y_limits} = {ast_node.y}\n"
+            self.code += "real slope;\n"
+            self.code += "real intercept;\n\n"
[…]
