+++ pysd/builders/stan/ast_walker.py
@@ -2,8 +2,12 @@ from typing import Union, List, Iterable, Dict, Tuple, Any
 from itertools import chain
 from dataclasses import dataclass, field
 from .utilities import IndentedString
-from pysd.translators.structures.abstract_model import\
-    AbstractComponent, AbstractElement, AbstractModel, AbstractSection
+from pysd.translators.structures.abstract_model import (
+    AbstractComponent,
+    AbstractElement,
+    AbstractModel,
+    AbstractSection,
+)
 
 from pysd.translators.structures.abstract_expressions import *
 
@@ -20,21 +24,34 @@ class AuxNameWalker(BaseNodeWaler):
         elif isinstance(ast_node, float):
             return []
         elif isinstance(ast_node, ArithmeticStructure):
-            return list(chain.from_iterable([self.walk(argument) for argument in ast_node.arguments]))
+            return list(
+                chain.from_iterable(
+                    [self.walk(argument) for argument in ast_node.arguments]
+                )
+            )
         elif isinstance(ast_node, ReferenceStructure):
             return [ast_node.reference]
         elif isinstance(ast_node, CallStructure):
-            return list(chain.from_iterable([self.walk(argument) for argument in ast_node.arguments]))
+            return list(
+                chain.from_iterable(
+                    [self.walk(argument) for argument in ast_node.arguments]
+                )
+            )
         elif isinstance(ast_node, IntegStructure):
             return self.walk(ast_node.flow) + self.walk(ast_node.initial)
         elif isinstance(ast_node, InlineLookupsStructure):
             return self.walk(ast_node.lookups)
         else:
-            raise Exception(f"AST node of type {ast_node.__class__.__name__} is not supported.")
+            raise Exception(
+                f"AST node of type {ast_node.__class__.__name__} is not supported."
+            )
+
 
 @dataclass
 class LookupCodegenWalker(BaseNodeWaler):
-    generated_lookup_function_names: Dict[Tuple, str] = field(default_factory=dict)
+    generated_lookup_function_names: Dict[Tuple, str] = field(
+        default_factory=dict
+    )
     # This dict holds the generated function names of each individual lookup function.
     # Key is x + y + x_limits + y_limits, value is function name
     n_lookups = 0
@@ -42,16 +59,25 @@ class LookupCodegenWalker(BaseNodeWaler):
 
     @staticmethod
     def get_lookup_keyname(lookup_node: LookupsStructure):
-        return lookup_node.x + lookup_node.y + lookup_node.x_limits + lookup_node.y_limits
+        return (
+            lookup_node.x
+            + lookup_node.y
+            + lookup_node.x_limits
+            + lookup_node.y_limits
+        )
 
     def walk(self, ast_node) -> None:
         if isinstance(ast_node, InlineLookupsStructure):
             self.walk(ast_node.lookups)
         elif isinstance(ast_node, LookupsStructure):
-            assert ast_node.type == "interpolate", "Type of Lookup must be 'interpolate'"
+            assert (
+                ast_node.type == "interpolate"
+            ), "Type of Lookup must be 'interpolate'"
             identifier_key = LookupCodegenWalker.get_lookup_keyname(ast_node)
             function_name = f"lookupFunc_{self.n_lookups}"
-            self.generated_lookup_function_names[identifier_key] = function_name
+            self.generated_lookup_function_names[
+                identifier_key
+            ] = function_name
             self.n_lookups += 1
             self.code += f"real {function_name}(real x){{\n"
             self.code.indent_level += 1
@@ -122,17 +148,25 @@ class BlockCodegenWalker(BaseNodeWaler):
             elif function_name == "max":
                 function_name = "fmax"
             elif function_name == "xidz":
[…]
