+++ pysd/builders/stan/stan_model_builder.py
@@ -4,8 +4,12 @@ from typing import Union, List, Dict, Set, Sequence
 
 from .ast_walker import *
 from .utilities import *
-from pysd.translators.structures.abstract_model import\
-    AbstractComponent, AbstractElement, AbstractModel, AbstractSection
+from pysd.translators.structures.abstract_model import (
+    AbstractComponent,
+    AbstractElement,
+    AbstractModel,
+    AbstractSection,
+)
 
 
 class StanModelBuilder:
@@ -13,14 +17,22 @@ class StanModelBuilder:
         self.abstract_model = abstract_model
 
         self.variable_ast_dict: Dict[str, AbstractSyntax] = {}
-        assert len(self.abstract_model.sections) == 1, "Number of sections in AbstractModel must be 1."
+        assert (
+            len(self.abstract_model.sections) == 1
+        ), "Number of sections in AbstractModel must be 1."
         for element in self.abstract_model.sections[0].elements:
             stan_varname = vensim_name_to_identifier(element.name)
-            assert len(element.components) == 1, f"Number of components in AbstractElement must be 1, but {element.name} has {len(element.components)}"
+            assert (
+                len(element.components) == 1
+            ), f"Number of components in AbstractElement must be 1, but {element.name} has {len(element.components)}"
             self.variable_ast_dict[stan_varname] = element.components[0].ast
 
-
-    def create_stan_program(self, predictor_variable_names: List[Union[str, Tuple[str, str]]], outcome_variable_names: Sequence[str] = (), function_name="vensim_func"):
+    def create_stan_program(
+        self,
+        predictor_variable_names: List[Union[str, Tuple[str, str]]],
+        outcome_variable_names: Sequence[str] = (),
+        function_name="vensim_func",
+    ):
         """
 
         Parameters
@@ -40,29 +52,53 @@ class StanModelBuilder:
         sanitized_predictor_variable_names = []
         for var in predictor_variable_names:
             if isinstance(var, str):
-                sanitized_predictor_variable_names.append(vensim_name_to_identifier(var))
+                sanitized_predictor_variable_names.append(
+                    vensim_name_to_identifier(var)
+                )
             elif isinstance(var, tuple):
                 var_name = var[1]
-                sanitized_predictor_variable_names.append((type, vensim_name_to_identifier(var_name)))
+                sanitized_predictor_variable_names.append(
+                    (type, vensim_name_to_identifier(var_name))
+                )
             else:
-                raise Exception("predictor_variable_names must be a list consisting of: strings and/or a tuple of the form(T, Name), where T is a string denoting the variable's stan type and Name a string denoting the variable name")
+                raise Exception(
+                    "predictor_variable_names must be a list consisting of: strings and/or a tuple of the form(T, Name), where T is a string denoting the variable's stan type and Name a string denoting the variable name"
+                )
 
         predictor_variable_names = sanitized_predictor_variable_names
-        outcome_variable_names = self.get_stock_variable_stan_names() if not outcome_variable_names else [vensim_name_to_identifier(name) for name in outcome_variable_names]
+        outcome_variable_names = (
+            self.get_stock_variable_stan_names()
+            if not outcome_variable_names
+            else [
+                vensim_name_to_identifier(name)
+                for name in outcome_variable_names
+            ]
+        )
         if not outcome_variable_names:
-            raise Exception("There are no stock variables defined in the model, hence nothing to integrate.")
+            raise Exception(
+                "There are no stock variables defined in the model, hence nothing to integrate."
+            )
 
         self.code = IndentedString()
 
         function_block_builder = StanFunctionBuilder(self.abstract_model)
 
[…]
