### 2022-08-11 c744e17
Minor update to logic
