### 2022-08-11 c16dbf3
Fix bug of constant cache assignment
