+++ docs/whats_new.rst
@@ -16,6 +16,7 @@ Bug fixes
 ~~~~~~~~~
 - Fix bug generated when :EXCEPT: keyword is used with subscript subranges (:issue:`352`).
 - Fix bug of precision error for :py:func:`pysd.py_backend.allocation.allocate_by_priority` (:issue:`353`).
+- Fix bug of constant cache assignment.
 
 Documentation
 ~~~~~~~~~~~~~
