### 2022-08-12 JamesPHoughton/pysd#348 @enekomartinmartinez
@JamesPHoughton @alexprey @rogersamso @julienmalard @SimonStrong @ConnectedSystems @pbreach @DVRodri8 @ainar @keith-citrenbaum @JoonaTuovinen @rugue1 @mhy05 @marrobl @sepiatone
Hi all,
I just wanted to let you know that we will move PySD to [SDXorg organization](https://github.com/SDXorg) in September. As there have been several contributions to PySD I would like to add an about section in the documentation where we reflect who made each contribution to the code. So if you want to appear in this section please write a short of message about your contributions to PySD and when they were done.
Mention any other developers if you miss them here.

Once the migration has been done GitHub will automatically redirect https://github.com/JamesPHoughton/pysd to https://github.com/SDXorg/pysd, however, it would be better if you update your links to PySD.
