### 2022-08-12 #4329 @sixpearls
Hi @rogersamso and co-authors, I have a few questions/comments for you before I can recommend acceptance:

- [ ] I found the PySD references confusing. The J. Houghton 2014 reference points to a zenodo archive of version 3.5.0 released in July 2022. The earliest version on that DOI is version 2.0.0 from November 2021. I don't think it makes sense for the reference to be dated 2014 since nothing about the archive is from 2014. It is better to state that the first release was in 2014 with no reference.

- [ ] I prefer references to the figures to be in complete sentences, like ",as shown in Figure #," rather than in a paranehtetical. I won't make this a blocker since (as far as I know) JOSS does not have a styleguide where this is dictated.

- [ ] Does the paper need to update references to the version number from 3.0.0 to 3.5.0?

- [ ] I found a few typos. It might be helpful to run the docs through a spell checker, do a thorough read through of the current version, and create a process for review of new docs. At least fix the ones I happened to find:
    - [ ] [Guidelines](https://pysd.readthedocs.io/en/master/development/guidelines.html): "encoraged" to "encouraged"
    - [ ] [Structure](https://pysd.readthedocs.io/en/master/structure/structure_index.html): "amb" to "and" and extra tick on 1+2*3-5
    - [ ] [Getting started](https://pysd.readthedocs.io/en/master/getting_started.html): "model component using curly brackets" should say square brackets

- [ ] The depth within the "Structure" section is confusing; I am not sure if I am seeing everything because sometimes the links on the menu take me to a section of an index page and sometimes to a separate sub-page. For reference, the structure of the Developer documentation seems clear.

- [ ] Can you add a page or modify the "getting started" guide to include a complete walk through of the features and research application of PySD on a simple model? The current "getting started" tutorial is disjointed with no clear story behind the sequence of feature demonstrations. As someone not in the "System Dynamics" field, it is not clear how someone would actually use PySD for the cited projects (MEDEAS, LOCOMOTION, etc).

### Technical

- [ ] Are you following semantic versioning?
- [ ] Are all of the file IO functions case insensitive? If so, can that be documented?
- [ ] Why does this library only use a forward euler solver? Is there a reason why scipy solvers were not used? Do you test your solver for mathematical properties, for example verifying accuracy of known solutions to within some tolerance?
- [ ] Is there a way to build or modify a model using the python API? Or do all model operations need to be performed in Stella or Vensim?
