### 2022-08-12 #4329 @enekomartinmartinez
Thanks for your comments @sixpearls
- I will review the references, in any case, I want to drop Zenodo's DOI as soon this paper is published. That may be an error I have done
- I will update the references to figures in the paper
- The main changes from 3.0.0 to the lastest have been new functions and some bug fixes. We can update the version is mentioned in the paper just to keep the last one. However, we want to mention 3.0.0 as it is the first version with the AMR
- I will correct the typos, thank you for pointing them
- I will see if I can improve the Structure section
- About the features demosntration on a simple model, I think it would make more sense to include them or update the [PySD CookBook](https://pysd-cookbook.readthedocs.io/en/latest/). Let me know if you still think that it should be in Getting started
- Yes, we are following semantic versioning, since version 3.0.0 all the changes have been tracked in the "what's new" section
- The case insensitivity is used only when parsing a model file to manage references to variables. Currently, when asking for a value of a function through PySD API there is case sensitivity, however, we could easily implement it following the same logic that Vensim/Xmile use.
- The generated objects are in a stock and flows representation, as far as I know scipy solvers only let solve a differential equations representation. Therefore the solver must to be given with PySD, so if we want to implement any other solver, we need to change how the Stateful objects work in the library, the main difficulty is to work with this object that require to save previous times values.
- You can modify some parameters values as explained in getting started or include own functions as explained in advanced usage. We have been trying to make the translated Python model file as simple as possible so users can also modify it to include new variables and processes.

As soon as I have applied all your suggestions I will let you know. Thank you!
