### 2022-08-12 #4329 @sixpearls
The PySD cookbook provides the type of documentation that is needed. Please provide a prominent link the readme and docs website. Another possibility is to insert the teacup page from the cookbook in place of the "getting started" page of the current documentation.
