### 2022-08-15 2102a9f
Convert sketch's font size optional
