+++ docs/whats_new.rst
@@ -18,6 +18,7 @@ Bug fixes
 
 Documentation
 ~~~~~~~~~~~~~
+- Correct typos.
 
 Performance
 ~~~~~~~~~~~
