### 2022-08-15 6a5b35d
Add full reference to figures and update last version
