+++ paper/paper.bib
@@ -101,16 +101,6 @@ eprint = {https://onlinelibrary.wiley.com/doi/pdf/10.1002/sdr.1695},
 year = {2021}
 }
 
-@software{pysd:2014,
-author = {James Houghton and Eneko Martin-Martinez and Alexey and Roger Samsó and Julien Malard-Adam and SimonStrong and Takuya Iwanaga and Patrick Breach and Diego Rodrigo and ... and ainar and Joona Tuovinen and Manuel Ruh and María Robles del Blanco and mhy05},
-title = {JamesPHoughton/pysd},
-month = dec,
-year = 2014,
-publisher = {Zenodo},
-doi = {10.5281/zenodo.5654824},
-url = {https://doi.org/10.5281/zenodo.5654824}
-}
-
 @inproceedings{houghton_siegel_2015,
     author = {Houghton, James P and Siegel, Michael},
     booktitle = {{Proceedings of the 33rd International Conference of the System Dynamics Society}},
