### 2022-08-15 #4329 @enekomartinmartinez
Changes done:
- PySD reference to Zenodo removed
- Complete sentence references
- Update the version number in the paper when talking about the changes done
- Correct documentation typos

I will take care of the other task when I am back from vacations (end-August/beginning-September)
