### 2022-08-18 ae761d3
Update return signature of ode function and change deprecated stan functions
