+++ test_scripts/testing.py
@@ -2,7 +2,7 @@ from pysd.translators.vensim.vensim_file import VensimFile
 from pysd.translators.xmile.xmile_file import XmileFile
 from pysd.builders.stan.stan_model_builder import *
 
-vf = VensimFile("vensim_models/demand-supply.mdl")
+vf = VensimFile("vensim_models/ds_white_sterman.mdl")
 #vf = VensimFile("vensim_models/arithmetic.mdl")
 #vf = XmileFile("vensim_models/repair.xmile")
 vf.parse()
@@ -11,8 +11,14 @@ am = vf.get_abstract_model()
 
 stan_builder = StanModelBuilder(am)
 stan_builder.print_variable_info()
-#print(stan_builder.create_stan_program([("int", "failure_count"), "repair_time"], ["battle_field", "repair_shop"])) # repair
-print(stan_builder.create_stan_program(["Demand"]))
+
+ass_param_lst = ["customer_order_rate", "inventory_coverage", "manufacturing_cycle_time", "time_to_average_order_rate", "wip_adjustment_time"]
+obs_stock_lst = ["work_in_process_inventory", "inventory"]
+
+#print(stan_builder.create_stan_program(ass_param_lst, obs_stock_lst))
+
+f_builder = StanFunctionBuilder(am)
+print(f_builder.build_function_block(ass_param_lst, obs_stock_lst))
 # for section in am.sections:
 #     for element in section.elements:
 #         print("*" * 10)
