### 2022-08-18 dc961b6
Change output signature of ode call
