### 2022-08-22 ad76f2a
WIP stan model interface
