### 2022-08-29 #4329 @enekomartinmartinez
Hi @sixpearls
I have been reviewing the depth in the Structure section:

- Structure of the PySD library (Page 1)
  - Translation (section on Page 1)
    * Vensim Translation (page 2)
      + Sections on page 2
    * XMILE Translation (page 3)
      + Sections on page 3
    * AMR Representation (page 4)
      + Sections on page 4
  - Building the Model (section on Page 1)
    * Python Builder (page 5)
      + Sections on page 5
  - The Python Model (section on Page 1)

Having something similar to the development documentation would be:

- Structure of the PySD library (Page 1)
  - Translation (Page 2)
    * Vensim Translation (Section on page 2)
      + Subsections on page 2
    * XMILE Translation (Section on page 2)
      + Subsections on page 2
    * AMR Representation (Section on page 2)
      + Subsections 0n page 2
  - Building the Model (Page 3)
    * Python Builder (section on page 3)
      + Subsections on page 3
  - The Python Model (Page 4)

I am not convinced by this way as it will make page 2 extremely long, and potentially page 3 in the future when more builders are added. Also the current separation of pages allows a developer in Vensim to have to read the Vensim translation page, otherwise they will have to see where the relevant section starts and ends which I see as more complicated.

I would keep the current classification. Let me know if you think there are any intermediate solutions that you think might work.
