### 2022-08-30 97b7777
Correct bugs when classifying variables per module
