+++ pysd/builders/python/python_model_builder.py
@@ -5,6 +5,7 @@ with PySD Model class. Each Abstract level has its own Builder. However,
 the user is only required to create a ModelBuilder object using the
 AbstractModel and call the `build_model` method.
 """
+from warnings import warn
 import textwrap
 import black
 import json
@@ -12,7 +13,8 @@ from pathlib import Path
 from typing import Union
 
 from pysd.translators.structures.abstract_model import\
-    AbstractComponent, AbstractElement, AbstractModel, AbstractSection
+    AbstractComponent, AbstractElement, AbstractControlElement,\
+    AbstractModel, AbstractSection
 
 from . import python_expressions_builder as vs
 from .namespace import NamespaceManager
@@ -142,34 +144,88 @@ class SectionBuilder:
             view_content = {
                 self.namespace.cleanspace[var] for var in view_content
             }
-            # Get subview elements
-            subview_elems = [
-                element for element in self.elements_remaining
-                if element.identifier in view_content
+
+            # Get subview elements (ordered)
+            subview_elems = sorted(
+                [
+                    element for element in self.elements_remaining
+                    if element.identifier in view_content
+                    and not element.control_var
+                ],
+                key=lambda x: x.identifier)
+
+            # Get the names of the elements and include their
+            # information in the elements_added dictionary
+            subview_elems_names = [
+                element.identifier for element in subview_elems
             ]
+            view_path = ".".join(view_name.parts[1:])
+            self.elements_added.update({
+                var: view_path for var in subview_elems_names
+            })
+
+            if len(view_content) != len(subview_elems_names):
+                # Some elements from the view where not added
+                for var in view_content.difference(subview_elems_names):
+                    original_name = self.namespace.get_original_name(var)
+                    if var in self.elements_added:
+                        # Element already added in another view
+                        warn(
+                            f"Variable '{original_name}' is declared as "
+                            f"a workbench variable in '{view_path}' but "
+                            "it has been already added in "
+                            f"'{self.elements_added[var]}'."
+                        )
+                    else:
+                        # Element is a control variable
+                        warn(
+                            f"Control variable '{original_name}' is "
+                            "declared as a workbench variable in "
+                            f"'{view_path}'. As it is a control "
+                            "variable, this declaration will be ignored "
+                            "and added to the main module only."
+                        )
+
             # Remove elements from remaining ones
             [
                 self.elements_remaining.remove(element)
                 for element in subview_elems
             ]
-            # Build the module
-            self._build_separate_module(subview_elems, view_name, wdir)
-            return sorted(view_content)
+
+            if subview_elems:
+                # Build the module (only when they are variables)
+                self._build_separate_module(subview_elems, view_name, wdir)
+
+            return list(subview_elems_names)
         else:
             # The current view has subviews
-            wdir = wdir.joinpath(view_name)
-            wdir.mkdir(exist_ok=True)
-            return {
-                subview_name:
-                self._process_views_tree(subview_name, subview_content, wdir)
+            (wdir / view_name).mkdir(exist_ok=True)
+            subviews = {
+                subview_name: self._process_views_tree(
+                    view_name / subview_name, subview_content, wdir)
                 for subview_name, subview_content in view_content.items()
+            }
+            # Avoid includying empty views to the dictionary
+            return {
+                subview_name: subview_content
+                for subview_name, subview_content in subviews.items()
+                if subview_content
             }
 
     def _build_modular(self, elements_per_view: dict) -> None:
         """ Build modular section """
         self.elements_remaining = self.elements.copy()
+        self.elements_added = {}
[…]
