+++ tests/pytest_translators/pytest_split_views.py
@@ -1,7 +1,8 @@
-
-import pytest
+import re
 from pathlib import Path
 import shutil
+
+import pytest
 
 import pysd
 from pysd.tools.benchmarking import assert_frames_close
@@ -16,66 +17,6 @@ def input_file(shared_tmpdir, _root):
     )
 
 
-@pytest.mark.parametrize(
-    "model_path,subview_sep,modules,macros,original_vars,py_vars,"
-    + "stateful_objs",
-    [
-        (  # split_views
-            Path("more-tests/split_model/test_split_model.mdl"),
-            [],
-            ["view_1", "view2", "view_3"],
-            [],
-            ["Stock"],
-            ["another_var", "rate1", "varn", "variablex", "stock"],
-            ["_integ_stock"]
-        ),
-        (  # split_subviews
-            Path("more-tests/split_model/test_split_model_subviews.mdl"),
-            ["."],
-            ["view_1/submodule_1", "view_1/submodule_2", "view_2"],
-            [],
-            ["Stock"],
-            ["another_var", "rate1", "varn", "variablex", "stock"],
-            ["_integ_stock"]
-        ),
-        (  # split_sub_subviews
-            Path("more-tests/split_model/test_split_model_sub_subviews.mdl"),
-            [".", "-"],
-            [
-                "view_1/submodule_1", "view_1/submodule_2", "view_2",
-                "view_3/subview_1/sview_1", "view_3/subview_1/sview_2",
-                "view_3/subview_2/sview_3", "view_3/subview_2/sview_4"
-            ],
-            [],
-            ["Stock"],
-            ["another_var", "rate1", "varn", "variablex", "stock",
-             "interesting_var_2", "great_var"],
-            ["_integ_stock"]
-        ),
-        (  # split_macro
-            Path("more-tests/split_model_with_macro/"
-                 + "test_split_model_with_macro.mdl"),
-            [".", "-"],
-            ["view_1", "view_2"],
-            ["expression_macro"],
-            ["new var"],
-            ["new_var"],
-            ["_macro_macro_output"]
-        ),
-        (  # split_vensim_8_2_1
-            Path("more-tests/split_model_vensim_8_2_1/"
-                 + "test_split_model_vensim_8_2_1.mdl"),
-            [],
-            ["teacup", "cream"],
-            [],
-            ["Teacup Temperature", "Cream Temperature"],
-            ["teacup_temperature", "cream_temperature"],
-            ["integ_teacup_temperature", "integ_cream_temperature"]
-        )
-    ],
-    ids=["split_views", "split_subviews", "split_sub_subviews", "split_macro",
-         "split_vensim_8_2_1"]
-)
 class TestSplitViews:
     """
     Test for splitting Vensim views in modules and submodules
@@ -101,6 +42,81 @@ class TestSplitViews:
         [files.add(shared_tmpdir.joinpath(macro + ".py")) for macro in macros]
         return files
 
+    @pytest.fixture
+    def nonexpected_files(self, shared_tmpdir, _root, model_path,
+                          model_file, emptymodules):
+        model_name = model_path.stem
+        shutil.copy(
+            _root.joinpath(model_path),
+            model_file
+        )
+        modules_dir = shared_tmpdir.joinpath("modules_" + model_name)
+        files = {
+            modules_dir.joinpath(module + ".py") for module in emptymodules
+        }
+        return files
+
+    @pytest.mark.parametrize(
+        "model_path,subview_sep,modules,macros,original_vars,py_vars,"
+        + "stateful_objs",
+        [
+            (  # split_views
+                Path("more-tests/split_model/test_split_model.mdl"),
+                [],
+                ["view_1", "view2", "view_3"],
+                [],
+                ["Stock"],
[…]
