### 2022-08-30 6a2b3e2
Correct bugs when classifying variables per module
