+++ docs/advanced_usage.rst
@@ -68,13 +68,16 @@ In a Vensim model with three separate views (e.g. `view_1`, `view_2` and `view_3
 | ├── _subscripts_many_views_model.json
 | ├── many_views_model.py
 
-
+The variables in each file will be sorted alphabetically, using their Python name.
 
 .. note ::
     Often, modelers wish to organise views further. To that end, a common practice is to include a particular character in the View name to indicate that what comes after it is the name of the subview. For instance, we could name one view as `ENERGY.Supply` and another one as `ENERGY.Demand`.
     In that particular case, setting the `subview_sep` kwarg equal to `["."]`, as in the code below, would name the translated views as `demand.py` and `supply.py` and place them inside the `ENERGY` folder::
 
       read_vensim("many_views_model.mdl", split_views=True, subview_sep=["."])
+
+.. note ::
+    If a variable appears as a `workbench variable` in more than one view, it will be added only to the module corresponding to the first view and a warning message will be printed. If a variable does not appear as a workbench variable in any view, it will be added to the main model file printing a warning message.
 
 If macros are present, they will be self-contained in files named after the macro itself. The macro inner variables will be placed inside the module that corresponds with the view in which they were defined.
 
