### 2022-08-31 JamesPHoughton/pysd#348 @enekomartinmartinez
@JamesPHoughton I was trying to migrate **excels2vensim** from GitLab to SDXorg. I am having the following error with RTD, I think it means that we need to give permisions to RTD to acces SDXorg projects.

```
[rtd-command-info] start-time: 2022-08-31T11:11:20.075404Z, end-time: 2022-08-31T11:11:20.224909Z, duration: 0, exit-code: 128
git clone --no-single-branch --depth 50 https://github.com/SDXorg/excels2vensim .
Cloning into '.'...
fatal: could not read Username for 'https://github.com': terminal prompts disabled
```

We need also to give access to coveralls
