### 2022-08-31 161e55e
Check if variables with priors exist in the ODE function declaration
