+++ pysd/builders/stan/stan_model.py
@@ -1,7 +1,7 @@
 from typing import List, Set, Type, Tuple
 from numbers import Number
 from dataclasses import dataclass, field
-import ast, os, pathlib, warnings, glob
+import ast, os, pathlib, warnings, glob, re
 from .stan_block_builder import *
 from .utilities import vensim_name_to_identifier
 from pysd.translators.structures.abstract_expressions import *
@@ -12,11 +12,13 @@ class SamplingStatement:
     distribution_type: str
     distribution_return_type: Type
     distribution_args: Tuple[str]
+    init_state: bool
 
-    def __init__(self, lhs_name, distribution_type, *distribution_args):
+    def __init__(self, lhs_name, distribution_type, *distribution_args, init_state=False):
         self.lhs_name = lhs_name
         self.distribution_type = distribution_type
         self.distribution_args = distribution_args
+        self.init_state = init_state
 
     def __post_init__(self):
         if self.distribution_type in ("bernoulli", "binomial", "beta_binomial", "neg_binomial", "poisson"):
@@ -78,20 +80,39 @@ class StanVensimModel:
         self.integration_times = integration_times
         self.stan_model_context = StanModelContext()
         self.vensim_model_context = VensimModelContext(self.abstract_model)
+        if initial_time in integration_times:
+            raise Exception("initial_time shouldn't be present in integration_times")
 
-    def set_prior(self, variable_name: str, distribution_type: str, *args):
-        for arg in args:
-            if isinstance(arg, str):
-                # If the distribution argument is an expression, parse the dependant variables
-                # We're using the python parser here, which might be problematic
-                used_variable_names = [node.id for node in ast.walk(ast.parse(arg)) if isinstance(node, ast.Name)]
-                for name in used_variable_names:
-                    if name in self.vensim_model_context.variable_names:
-                        self.stan_model_context.exposed_parameters.update(used_variable_names)
+        self.init_variable_regex = re.compile(".+?(?=_init$)")
+        # This regex is to match all preceding characters that come before '_init' at the end of the string.
+        # So something like stock_var_init_init would match into stock_var_init.
+        # This is used to parse out the corresponding stock names for init parameters.
 
-        if variable_name in self.vensim_model_context.variable_names:
-            self.stan_model_context.exposed_parameters.add(variable_name)
-        self.stan_model_context.sample_statements.append(SamplingStatement(variable_name, distribution_type, *args))
+    def print_info(self):
+        print("- Vensim model information:")
+        self.vensim_model_context.print_variable_info(self.abstract_model)
+        print("*" * 10)
+        print("- Stan model information:")
+
+    def set_prior(self, variable_name: str, distribution_type: str, *args, init_state=False):
+        if init_state:
+            # This means the initial value of the ODE state variable.
+            if variable_name not in self.vensim_model_context.stock_variable_names:
+                raise Exception("init_state may be set to True only for stock variables.")
+            self.stan_model_context.sample_statements.append(SamplingStatement(f"{variable_name}_init", distribution_type, *args, init_state=init_state))
+        else:
+            for arg in args:
+                if isinstance(arg, str):
+                    # If the distribution argument is an expression, parse the dependant variables
+                    # We're using the python parser here, which might be problematic
+                    used_variable_names = [node.id for node in ast.walk(ast.parse(arg)) if isinstance(node, ast.Name)]
+                    for name in used_variable_names:
+                        if name in self.vensim_model_context.variable_names:
+                            self.stan_model_context.exposed_parameters.update(used_variable_names)
+
+            if variable_name in self.vensim_model_context.variable_names:
+                self.stan_model_context.exposed_parameters.add(variable_name)
[…]
