+++ test_scripts/stan_vensim_integration.py
@@ -6,10 +6,13 @@ vf = VensimFile("vensim_models/ds_white_sterman.mdl")
 vf.parse()
 am = vf.get_abstract_model()
 
-model = StanVensimModel("ds_white_sterman", am, 0.0, list(range(0, 10)))
+model = StanVensimModel("ds_white_sterman", am, 0.0, list(range(1, 10)))
+model.print_info()
+
 model.set_prior("inventory_adjustment_time", "normal", 0, 1)
 model.set_prior("minimum_order_processing_time", "normal", 0, 1)
 model.set_prior("alpha", "normal", 0, 1)
+model.set_prior("inventory", "normal", 0, 1, init_state=True)
 
 print(model.vensim_model_context.variable_names)
 

[…]
