### 2022-08-31 5584248
Update lookup codegen logic and create cmdstan model
