+++ test_scripts/stan_vensim_integration.py
@@ -18,4 +18,6 @@ print(model.vensim_model_context.variable_names)
 
 model.build_stan_functions()
 
-model.data2draws("")
+cmdstan_model = model.data2draws("")
+result = cmdstan_model.sample()
+result.summary()
