### 2022-09-02 1ac6847
first implementation of netCDF storage
