+++ pysd/py_backend/model.py
@@ -1172,21 +1172,20 @@ class Model(Macro):
         # need to clean cache to remove the values from active_initial
         self.clean_caches()
 
-        if output_file:
-            self._integrate_binary_output(capture_elements['step'], out_file=output_file)
-            print(f"Resulsts stored in {output_file}")
-            return None
-
-        res = self._integrate(capture_elements['step'])
+        out_obj = self._integrate(capture_elements['step'], output_file)
 
         del self._dependencies["OUTPUTS"]
 
-        self._add_run_elements(res, capture_elements['run'])
+        out_obj.add_run_elements(self, capture_elements['run'])
+
         self._remove_constant_cache()
 
-        return_df = utils.make_flat_df(res, return_addresses, flatten_output)
+        if output_file:
+            out_obj.handler.ds.close()
+            return None
 
-        return return_df
+        return utils.make_flat_df(out_obj.handler.ds, return_addresses, flatten_output)
+
 
     def select_submodel(self, vars=[], modules=[], exogenous_components={}):
         """
@@ -1654,60 +1653,8 @@ class Model(Macro):
         """
         self.state = self.state + self.ddt() * dt
 
-    def _integrate(self, capture_elements):
-        """
-        Performs euler integration.
 
-        Parameters
-        ----------
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        outputs: pandas.DataFrame
-            Output capture_elements data.
-
-        """
-        # necessary to have always a non-xaray object for appending objects
-        # to the DataFrame time will always be a model element and not saved
-        # TODO: find a better way of saving outputs
-        capture_elements.add("time")
-        outputs = pd.DataFrame(columns=capture_elements)
-
-        if self.progress:
-            # initialize progress bar
-            progressbar = utils.ProgressBar(
-                int((self.time.final_time()-self.time())/self.time.time_step())
-            )
-        else:
-            # when None is used the update will do nothing
-            progressbar = utils.ProgressBar(None)
-
-        while self.time.in_bounds():
-            if self.time.in_return():
-                outputs.at[self.time.round()] = [
-                    getattr(self.components, key)()
-                    for key in capture_elements]
-            self._euler_step(self.time.time_step())
-            self.time.update(self.time()+self.time.time_step())
-            self.clean_caches()
-            progressbar.update()
-
-        # need to add one more time step, because we run only the state
-        # updates in the previous loop and thus may be one short.
-        if self.time.in_return():
-            outputs.at[self.time.round()] = [getattr(self.components, key)()
-                                             for key in capture_elements]
-
-        progressbar.finish()
-
-        # delete time column as it was created only for avoiding errors
-        # of appending data. See previous TODO.
-        del outputs["time"]
-        return outputs
-
-    def _integrate_binary_output(self, capture_elements, out_file):
+    def _integrate(self, capture_elements, out_file):
         """
         Performs euler integration and stores the results in netCDF4 file.
 
@@ -1722,6 +1669,9 @@ class Model(Macro):
         nothing
 
         """
+        # instantiating output object
+        output = ModelOutput(capture_elements, out_file)
+        output.initialize(self)
 
         if self.progress:
             # initialize progress bar
@@ -1732,104 +1682,239 @@ class Model(Macro):
             # when None is used the update will do nothing
             progressbar = utils.ProgressBar(None)
 
-
-        with nc.Dataset(out_file, "w") as ds:
-            # generating global attributes
-            ds.description = "Results for simulation run on" \
-            f"{t.ctime(t.time())} using PySD version {__version__}"
-            ds.model_file = self.py_model_file
-            ds.timestep = "{}".format(self.components.time_step())
-            ds.initial_time = "{}".format(self.components.initial_time())
[…]
