+++ pysd/py_backend/model.py
@@ -1160,6 +1160,7 @@ class Model(Macro):
         self._dependencies["OUTPUTS"] = {
             element: 1 for element in capture_elements["step"]
         }
+
         if cache_output:
             # udate the cache type taking into account the outputs
             self._assign_cache_type()
@@ -1172,19 +1173,19 @@ class Model(Macro):
         # need to clean cache to remove the values from active_initial
         self.clean_caches()
 
-        out_obj = self._integrate(capture_elements['step'], output_file)
+        # instantiating output object
+        output = ModelOutput(self, capture_elements['step'], output_file)
+
+        self._integrate(output)
 
         del self._dependencies["OUTPUTS"]
 
-        out_obj.add_run_elements(self, capture_elements['run'])
+        output.add_run_elements(self, capture_elements['run'])
 
         self._remove_constant_cache()
 
-        if output_file:
-            out_obj.handler.ds.close()
-            return None
-
-        return utils.make_flat_df(out_obj.handler.ds, return_addresses, flatten_output)
+        return output.postprocess(return_addresses=return_addresses,
+                                  flatten=flatten_output)
 
 
     def select_submodel(self, vars=[], modules=[], exogenous_components={}):
@@ -1654,24 +1655,19 @@ class Model(Macro):
         self.state = self.state + self.ddt() * dt
 
 
-    def _integrate(self, capture_elements, out_file):
+    def _integrate(self, out_obj):
         """
-        Performs euler integration and stores the results in netCDF4 file.
+        Performs euler integration and writes results to the out_obj.
 
         Parameters
         ----------
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-        out_file: str or pathlib.Path
+        out_obj: pysd.ModelOutput
 
         Returns
         -------
-        nothing
+        None
 
         """
-        # instantiating output object
-        output = ModelOutput(capture_elements, out_file)
-        output.initialize(self)
 
         if self.progress:
             # initialize progress bar
@@ -1682,30 +1678,44 @@ class Model(Macro):
             # when None is used the update will do nothing
             progressbar = utils.ProgressBar(None)
 
+        # performs the time stepping
         while self.time.in_bounds():
             if self.time.in_return():
-                output.update(self)
+                out_obj.update(self)
 
-                self._euler_step(self.time.time_step())
-                self.time.update(self.time()+self.time.time_step())
-                self.clean_caches()
-                progressbar.update()
+            self._euler_step(self.time.time_step())
+            self.time.update(self.time()+self.time.time_step())
+            self.clean_caches()
+            progressbar.update()
 
         # need to add one more time step, because we run only the state
         # updates in the previous loop and thus may be one short.
         if self.time.in_return():
-            output.update(self)
+            out_obj.update(self)
 
         progressbar.finish()
 
-        output.postprocess()
-        return output
-
-
 class OutputHandler(object):
-    pass
+    """
+    Interface for the different output handlers.
+    """
+
+    def initialize(self):
+        pass
+
+    def update(self):
+        pass
+
+    def postprocess(self):
+        pass
+
+    def add_run_elements(self):
+        pass
 
 class DatasetHandler(OutputHandler):
+    """
+    Manages simulation results stored as netCDF4 Dataset.
+    """
 
     def __init__(self, out_file):
         self.out_file = out_file
@@ -1714,23 +1724,23 @@ class DatasetHandler(OutputHandler):
 
     def initialize(self, model, capture_elements):
         """
-        Creates a netCDF4 dataset and adds model dimensions and variables
+        Creates a netCDF4 Dataset and adds model dimensions and variables
         present in the capture elements to it.
 
         Parameters
         ----------
+        model: pysd.Model
[…]
