### 2022-09-05 c85a8bc
move output classes to separate file
