+++ pysd/py_backend/output.py
@@ -1,0 +1,381 @@
+"""
+ModelOutput class is used to build different output objects based on
+user input. For now, available output types are pandas DataFrame or
+netCDF4 Dataset.
+The OutputHandlerInterface class is an interface for the creation of handlers
+for other output object types.
+"""
+import abc
+import warnings
+import time as t
+
+import numpy as np
+import xarray as xr
+import pandas as pd
+import netCDF4 as nc
+
+from pysd._version import __version__
+
+from . import utils
+
+
+class ModelOutput():
+    """
+    Handles different types of outputs by dispatchinging the tasks to adequate
+    object handlers.
+
+    Parameters
+    ----------
+    model: pysd.Model
+        PySD Model object
+    capture_elements: set
+        Which model elements to capture - uses pysafe names.
+    out_file: str or pathlib.Path
+        Path to the file where the results will be written.
+    """
+    valid_output_files = [".nc"]
+
+    def __init__(self, model, capture_elements, out_file=None):
+
+        self.handler = self.__handle(out_file)
+
+        capture_elements.add("time")
+        self.capture_elements = capture_elements
+
+        self.initialize(model)
+
+    def __handle(self, out_file):
+        # TODO improve the handler to avoid if then else statements
+        if out_file:
+            if out_file.suffix == ".nc":
+                return DatasetHandler(out_file)
+            raise ValueError(
+                    f"Unsupported output file format {out_file.suffix}")
+        return DataFrameHandler()
+
+    def initialize(self, model):
+        """ Delegating the creation of the results object and its elements to
+        the appropriate handler."""
+        self.handler.initialize(model, self.capture_elements)
+
+    def update(self, model):
+        """ Delegating the update of the results object and its elements to the
+        appropriate handler."""
+        self.handler.update(model, self.capture_elements)
+
+    def postprocess(self, **kwargs):
+        """ Delegating the postprocessing of the results object to the
+        appropriate handler."""
+        return self.handler.postprocess(**kwargs)
+
+    def add_run_elements(self, model, run_elements):
+        """ Delegating the addition of results with run cache in the output
+        object to the appropriate handler."""
+        self.handler.add_run_elements(model, run_elements)
+
+
+class OutputHandlerInterface(metaclass=abc.ABCMeta):
+    """
+    Interface for the creation of different output type handlers.
+    """
+    @classmethod
+    def __subclasshook__(cls, subclass):
+        return (hasattr(subclass, 'initialize') and
+                callable(subclass.initialize) and
+                hasattr(subclass, 'update') and
+                callable(subclass.update) and
+                hasattr(subclass, 'postprocess') and
+                callable(subclass.postprocess) and
+                hasattr(subclass, 'add_run_elements') and
+                callable(subclass.add_run_elements) or
+                NotImplemented)
+
+    @abc.abstractmethod
+    def initialize(self, model, capture_elements):
+        """
+        Create the results object and its elements based on capture_elemetns.
+        """
+        raise NotImplementedError
+
+    def update(self, model, capture_elements):
+        """
+        Update the results object at each iteration at which resutls are
+        stored.
+        """
+        raise NotImplementedError
+
+    def postprocess(self, **kwargs):
+        """
+        Perform different tasks at the time of returning the results object.
+        """
+        raise NotImplementedError
+
+    def add_run_elements(self, model, capture_elements):
+        """
+        Add elements with run cache to the results object.
+        """
+        raise NotImplementedError
+
+
+class DatasetHandler(OutputHandlerInterface):
+    """
+    Manages simulation results stored as netCDF4 Dataset.
+    """
+
+    def __init__(self, out_file):
+        self.out_file = out_file
+        self.step = 0
+        self.ds = None
+
+    def initialize(self, model, capture_elements):
+        """
[…]
