### 2022-09-05 6289dfb
move output classes to separate file
