+++ pysd/py_backend/model.py
@@ -8,14 +8,14 @@ a separate file).
 import warnings
 import inspect
 import pickle
-import time as t
 from pathlib import Path
 from typing import Union
 
 import numpy as np
 import xarray as xr
 import pandas as pd
-import netCDF4 as nc
+
+from pysd._version import __version__
 
 from . import utils
 from .statefuls import DynamicStateful, Stateful
@@ -24,8 +24,7 @@ from .cache import Cache, constant_cache
 from .data import TabData
 from .lookups import HardcodedLookups
 from .components import Components, Time
-
-from pysd._version import __version__
+from .output import ModelOutput
 
 
 class Macro(DynamicStateful):
@@ -1025,6 +1024,7 @@ class Model(Macro):
         self.time.set_control_vars(**self.components._control_vars)
         self.data_files = data_files
         self.missing_values = missing_values
+        self.progress = None
         if initialize:
             self.initialize()
 
@@ -1673,7 +1673,6 @@ class Model(Macro):
         """
         self.state = self.state + self.ddt() * dt
 
-
     def _integrate(self, out_obj):
         """
         Performs euler integration and writes results to the out_obj.
@@ -1713,326 +1712,3 @@ class Model(Macro):
             out_obj.update(self)
 
         progressbar.finish()
-
-
-class OutputHandler():
-    """
-    Interface for the different output handlers.
-    """
-
-    def initialize(self, model, capture_elements):
-        pass
-
-    def update(self, model, capture_elements):
-        pass
-
-    def postprocess(self, **kwargs):
-        pass
-
-    def add_run_elements(self, model, capture_elements):
-        pass
-
-
-class DatasetHandler(OutputHandler):
-    """
-    Manages simulation results stored as netCDF4 Dataset.
-    """
-
-    def __init__(self, out_file):
-        self.out_file = out_file
-        self.step = 0
-        self.ds = None
-
-    def initialize(self, model, capture_elements):
-        """
-        Creates a netCDF4 Dataset and adds model dimensions and variables
-        present in the capture elements to it.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
-
-        """
-        self.ds = nc.Dataset(self.out_file, "w")
-        # defining global attributes
-        self.ds.description = "Results for simulation run on" \
-        f"{t.ctime(t.time())} using PySD version {__version__}"
-        self.ds.model_file = model.py_model_file
-        self.ds.timestep = "{}".format(model.components.time_step())
-        self.ds.initial_time = "{}".format(model.components.initial_time())
-        self.ds.final_time = "{}".format(model.components.final_time())
-
-        # creating variables for all model dimensions
-        for dim_name,coords in model.subscripts.items():
-            coords = np.array(coords)
-            # create dimension
-            self.ds.createDimension(dim_name, len(coords))
-            # length of the longest string in the coords
-            max_str_len = len(max(coords, key=len))
-
-            # create variable
-            # TODO: check if the type could be defined otherwise)
-            var = self.ds.createVariable(dim_name,
-                                    f"S{max_str_len}",
-                                    (dim_name,))
-            # assigning values to variable
-            var[:] = coords
-
-        # creating the time dimension as unlimited
-        self.ds.createDimension("time", None)
-
-        # creating variables in capture_elements
-        self.__create_ds_vars(model, capture_elements)
-
-    def update(self, model, capture_elements):
-        """
-        Writes values of variables in capture_elements in the netCDF4 Dataset.
-
-        Parameters
-        ----------
-        model: pysd.Model
-            PySD Model object
-        capture_elements: set
-            Which model elements to capture - uses pysafe names.
-
-        Returns
-        -------
-        None
[…]
