### 2022-09-05 78480ca
allow .nc exports from cli
