### 2022-09-05 4ce0221
allow .nc exports from cli
