### 2022-09-05 #4329 @pdebuyl
Thank you @sixpearls and @enekomartinmartinez for the updates.

@sixpearls could you review the changes?

A few comments:
- Regarding the references to a zenodo archive: this is a common issue and it is usually fixed at the publication of the JOSS article.
- Regarding the structure of the sections in the documentation, the review here goes a bit beyond what we typically do in JOSS. As there is some subjective impression at play, I would recommend to let some flexibility to the authors.

(sorry for coming back here so late, I'll check regularly until we can get the article published).

As I am checking the reviewer list of @sixpearls I see "community ..." missing. On the GitHub README, the community contribution part is not clear. Are contributions welcome? Please check how some other JOSS-published codes have written this part to get an impression of what is expected.
