### 2022-09-05 daf724c
Set parsimonious version to avoid crash
