+++ requirements.txt
@@ -1,5 +1,5 @@
 pandas
-parsimonious
+parsimonious==0.9.0
 xarray
 xlrd
 lxml
