### 2022-09-05 47dd3a3
move make_flat_df and tests
