+++ tests/pytest_pysd/pytest_pysd.py
@@ -8,6 +8,8 @@ import xarray as xr
 import netCDF4 as nc
 
 from pysd.tools.benchmarking import assert_frames_close
+from pysd.py_backend.output import OutputHandlerInterface, DatasetHandler, \
+    DataFrameHandler, ModelOutput
 
 import pysd
 
@@ -24,12 +26,14 @@ test_model_look = _root.joinpath(
     + "test_get_lookups_subscripted_args.mdl")
 test_model_data = _root.joinpath(
     "test-models/tests/get_data_args_3d_xls/test_get_data_args_3d_xls.mdl")
-
+test_model_constants = _root.joinpath(
+    "test-models/tests/get_constants_subranges/"
+    "test_get_constants_subranges.mdl"
+)
 more_tests = _root.joinpath("more-tests/")
 
 test_model_constant_pipe = more_tests.joinpath(
     "constant_pipeline/test_constant_pipeline.mdl")
-
 
 class TestPySD():
 
@@ -1684,9 +1688,6 @@ class TestExportImport():
 class TestOutputs():
 
     def test_output_handler_interface(self):
-        from pysd.py_backend.output import OutputHandlerInterface, \
-         DatasetHandler, DataFrameHandler, ModelOutput
-
         # when the class does not inherit from OutputHandlerInterface, it must
         # implement all the interface to be a subclass of
         # OutputHandlerInterface.
@@ -1717,7 +1718,6 @@ class TestOutputs():
         # considered a subclass, because it follows the interface
         assert issubclass(ThatFollowsInterface, OutputHandlerInterface)
 
-
         class IncompleteHandler:
             """
             Class that does not follow the full interface
@@ -1734,8 +1734,7 @@ class TestOutputs():
 
         # It does not inherit from OutputHandlerInterface and does not fulfill
         # its interface
-        assert issubclass(IncompleteHandler, OutputHandlerInterface) == False
-
+        assert not issubclass(IncompleteHandler, OutputHandlerInterface)
 
         class EmptyHandler(OutputHandlerInterface):
             """
@@ -1769,7 +1768,6 @@ class TestOutputs():
             EmptyHandler.add_run_elements(
                 EmptyHandler, "model", "capture")
 
-
     def test_output_with_dimensions(self, shared_tmpdir):
         model = pysd.read_vensim(test_model_look)
         model.progress = False
@@ -1781,8 +1779,9 @@ class TestOutputs():
             model.run(output_file=out_file)
 
         with nc.Dataset(out_file, "r")as ds:
-            assert ds.ncattrs() == ['description', 'model_file', 'timestep',
-             'initial_time', 'final_time']
+            assert ds.ncattrs() == [
+                'description', 'model_file', 'timestep', 'initial_time',
+                'final_time']
             assert list(ds.dimensions.keys()) == ["Rows", "Dim", "time"]
             # dimensions are stored as variables
             assert ds["Rows"][:].size == 2
@@ -1797,3 +1796,207 @@ class TestOutputs():
             assert ds["d2d"].description == "Missing"
             assert ds["d2d"].units == "Missing"
 
+        # test cache run variables with dimensions
+        model2 = pysd.read_vensim(test_model_constants)
+        model2.progress = False
+
+        out_file2 = shared_tmpdir.joinpath("results2.nc")
+
+        with catch_warnings(record=True) as w:
+            simplefilter("always")
+            model2.run(output_file=out_file2)
+
+        with nc.Dataset(out_file2, "r")as ds:
[…]
