### 2022-09-06 8839b47
add section on output_file argument in the docs
