### 2022-09-06 19b0f0f
add section on output_file argument in the docs
