+++ pysd/cli/parser.py
@@ -30,7 +30,7 @@ def check_output(string):
 
     """
     if not string.endswith('.tab') and not string.endswith('.csv') and not \
-         string.endswith('.nc'):
+     string.endswith('.nc'):
         parser.error(
             f'when parsing {string}'
             '\nThe output file name must be .tab, .csv or .nc...')
