### 2022-09-06 53febea
Codegen cleanup and draws2data implementation
