+++ pysd/builders/stan/stan_model.py
@@ -7,7 +7,9 @@ from pysd.translators.structures.abstract_expressions import *
 import cmdstanpy
 
 class SamplingStatement:
-    lhs_name: str
+    lhs_expr: str
+    lhs_variables: Set[str]
+    rhs_variables: Set[str]
     distribution_type: str
     distribution_return_type: Type
     distribution_args: Tuple[str]
@@ -15,10 +17,14 @@ class SamplingStatement:
     upper: float
     init_state: bool
 
-    def __init__(self, lhs_name, distribution_type, *distribution_args, lower=float("-inf"), upper=float("inf"), init_state=False):
-        self.lhs_name = lhs_name
+    def __init__(self, lhs_expr, distribution_type, *distribution_args, lower=float("-inf"), upper=float("inf"), init_state=False):
+        self.lhs_expr = lhs_expr
+        self.lhs_variables = set()
+        self.lhs_variables.update([node.id for node in ast.walk(ast.parse(lhs_expr)) if isinstance(node, ast.Name)])
         self.distribution_type = distribution_type
         self.distribution_args = distribution_args
+        self.rhs_variables = set()
+        self.rhs_variables.update([node.id for arg in self.distribution_args for node in ast.walk(ast.parse(str(arg))) if isinstance(node, ast.Name)])
         self.lower = lower
         self.upper = upper
         self.init_state = init_state
@@ -33,6 +39,8 @@ class SamplingStatement:
 
 @dataclass
 class StanModelContext:
+    initial_time: float
+    integration_times: Iterable[float]
     sample_statements: List[SamplingStatement] = field(default_factory=list)
     exposed_parameters: Set[str] = field(default_factory=set)
 
@@ -81,7 +89,7 @@ class StanVensimModel:
         self.model_name = model_name
         self.initial_time = float(initial_time)
         self.integration_times = integration_times
-        self.stan_model_context = StanModelContext()
+        self.stan_model_context = StanModelContext(initial_time, integration_times)
         self.vensim_model_context = VensimModelContext(self.abstract_model)
         if initial_time in integration_times:
             raise Exception("initial_time shouldn't be present in integration_times")
@@ -90,9 +98,9 @@ class StanVensimModel:
         if not os.path.exists(self.stan_model_dir):
             os.mkdir(self.stan_model_dir)
 
-        self.init_variable_regex = re.compile(".+?(?=_init$)")
-        # This regex is to match all preceding characters that come before '_init' at the end of the string.
-        # So something like stock_var_init_init would match into stock_var_init.
+        self.init_variable_regex = re.compile(".+?(?=__init$)")
+        # This regex is to match all preceding characters that come before '__init' at the end of the string.
+        # So something like stock_var_init__init would match into stock_var__init.
         # This is used to parse out the corresponding stock names for init parameters.
 
     def print_info(self):
@@ -112,12 +120,12 @@ class StanVensimModel:
                 if isinstance(arg, str):
                     # If the distribution argument is an expression, parse the dependant variables
                     # We're using the python parser here, which might be problematic
-                    used_variable_names = [node.id for node in ast.walk(ast.parse(arg)) if isinstance(node, ast.Name)]
+                    used_variable_names = [node.id.strip() for node in ast.walk(ast.parse(arg)) if isinstance(node, ast.Name)]
                     for name in used_variable_names:
-                        if name in self.vensim_model_context.variable_names:
+                        if name in self.vensim_model_context.variable_names and name not in self.vensim_model_context.stock_variable_names:
                             self.stan_model_context.exposed_parameters.update(used_variable_names)
 
-            if variable_name in self.vensim_model_context.variable_names:
[…]
