+++ pysd/builders/stan/stan_block_builder.py
@@ -12,7 +12,7 @@ from pysd.translators.structures.abstract_model import (
 )
 from typing import TYPE_CHECKING
 if TYPE_CHECKING:
-    from .stan_model import SamplingStatement
+    from .stan_model import StanModelContext, VensimModelContext
 
 
 class StanTransformedParametersBuilder:
@@ -57,7 +57,7 @@ class StanTransformedParametersBuilder:
                     assert isinstance(
                         component.ast, IntegStructure
                     ), "Output variable component must be an INTEG."
-                    self.code += f"real {outcome_variable_name}_init = {InitialValueCodegenWalker(lookup_function_dict, variable_ast_dict).walk(component.ast)};\n"
+                    self.code += f"real {outcome_variable_name}__init = {InitialValueCodegenWalker(lookup_function_dict, variable_ast_dict).walk(component.ast)};\n"
                     break
 
         self.code += "\n"
@@ -67,11 +67,15 @@ class StanTransformedParametersBuilder:
             if name in stock_initial_values:
                 self.code += f"initial_outcome[{index}] = {stock_initial_values[name]};  // Defined within stan\n"
             else:
-                self.code += f"initial_outcome[{index}] = {name}_init;\n"
+                self.code += f"initial_outcome[{index}] = {name}__init;\n"
 
         self.code += "\n"
 
-        self.code += f"vector[{len(outcome_variable_names)}] integrated_result[n_t] = ode_rk45({function_name}, initial_outcome, initial_time, times, {', '.join(argument_variables)});\n"
+        ode_solver_code = f"vector[{len(outcome_variable_names)}] integrated_result[n_t] = ode_rk45({function_name}, initial_outcome, initial_time, times"
+        if len(argument_variables) > 0:
+            ode_solver_code += ", " + f"{', '.join(argument_variables)}"
+
+        self.code += ode_solver_code + ");\n"
 
         for index, name in enumerate(outcome_variable_names, 1):
             self.code += f"array[n_t] real {name} = integrated_result[:, {index}];\n"
@@ -83,26 +87,32 @@ class StanTransformedParametersBuilder:
 
 
 class StanParametersBuilder:
-    def __init__(self, sampling_statements: Iterable["SamplingStatement"]):
-        self.sampling_statements = sampling_statements
+    def __init__(self, stan_model_context: "StanModelContext"):
+        self.stan_model_context = stan_model_context
 
     def build_block(self, data_variable_names: Tuple[str]):
         code = IndentedString()
         code += "parameters{\n"
         code.indent_level += 1  # Enter parameters block
 
-        for statement in self.sampling_statements:
-            if statement.lhs_name in data_variable_names:
-                continue
+        added_parameters = set()
 
-            if statement.lower > float("-inf") and statement.upper < float("inf"):
-                code += f"real<lower={statement.lower}, upper={statement.upper}> {statement.lhs_name};\n"
-            elif statement.lower > float("-inf"):
-                code += f"real<lower={statement.lower}> {statement.lhs_name};\n"
-            elif statement.upper < float("inf"):
-                code += f"real<upper={statement.upper}> {statement.lhs_name};\n"
-            else:
-                code += f"real {statement.lhs_name};\n"
+        for statement in self.stan_model_context.sample_statements:
+            for lhs_variable in statement.lhs_variables:
+                if lhs_variable in data_variable_names:
+                    continue
+                if lhs_variable in added_parameters:
+                    continue
+                if statement.lower > float("-inf") and statement.upper < float("inf"):
+                    code += f"real<lower={statement.lower}, upper={statement.upper}> {lhs_variable};\n"
+                elif statement.lower > float("-inf"):
[…]
