### 2022-09-06 00efc5f
Make PySD work with parsimonius 0.10.0
