+++ requirements.txt
@@ -1,5 +1,5 @@
 pandas
-parsimonious==0.9.0
+parsimonious
 xarray
 xlrd
 lxml
