### 2022-09-06 f5aec83
Fix bug when a WITH LOOKUPS argument has subscripts.
