### 2022-09-06 78cb772
add all doc attrs to vars in nc
