### 2022-09-06 980dbe5
add all doc attrs to vars in nc
