### 2022-09-07 #4329 @enekomartinmartinez
I still have to include the link to the cookbook in getting started and review the contributions section in documentation and README. I will also add a "Contributions welcome" badge to make it clear.

We will migrate PySD to [SDXorg](https://github.com/SDXorg) next Thursday. After that migration, we potentially will add a new page for contributions with also a mention of all the historic contributors. I will let you know when it is ready for review.
