### 2022-09-07 JamesPHoughton/pysd#348 @enekomartinmartinez
It would be interesting to configure this GHAction for checking the links
https://github.com/lycheeverse/lychee-action
