### 2022-09-07 JamesPHoughton/pysd#348 @enekomartinmartinez
https://github.com/codespell-project/actions-codespell
