### 2022-09-07 #4329 @sixpearls
@pdebuyl I am happy to keep the documentation structure off the blocking list, but I do think reviewing the documentation and tests is an important "feature" of publishing with JOSS.

@enekomartinmartinez

I'll clear the check list from above and list the blocking issues here:

- [ ] Adding a reference to the cookbook on the "getting started" page and elsewhere as appropriate in the documentation.
- [ ] Adding a summary community guidelines section in the readme with a link to the appropriate development documentation pages
- [ ] Missing space after comma on line 14, page 1 of the paper (from new version)
- [ ] Please add tests that ensure the Euler integration is correct. I recommend identifying several models with known analytic solutions (e.g., linear systems with different stability properties -- unstable, stable, oscillatory). Then run the model with different time steps. I believe the error should be proportional to 1/h^2, where h is the step size.

Below are some non-blocking suggestions:

- Add documentation to how the tests are designed, particular how the test-models are run. In general I like the structure you've developed, but I think a new contributor may have difficulty adding new tests.
- Add documentation (especially to the relevant docstrings) about when a function is case sensitive or not.
- As far as the structure of the structure pages, the confusion of the current layout comes from the navigation panel showing the third level links as under a section of the first page. I don't mind very long pages as you've proposed since there will be links to the various sections. Another option would be many short pages, as in

    - Structure of the PySD library (Page 1) -- current "structure of the PySD library" section (3 paragraphs and figure) with new table of contents
        - Translation (Page 2) -- current "translation" section from page 1, table of contents
            - Vensim Translation (page 3), with sections
            - XMILE Translation (page 4), with sections
            - AMR Representation (page 5), with sections
        - Building the Model (page 6) -- since this is (currently) small, this could be a single page with the content from the current overview page and the current python builder page.
        - The Python Model (page 7)

    Since this is non-blocking, I leave the trade-off terms of possible confusion with some of the sub-pages, many short pages, or several long pages, to you.
