### 2022-09-08 #4329 @pdebuyl
Thanks @sixpearls for the reply. I fully agree that the documentation is within the scope of the review. At some point of detail, though, there has to be some freedom for the authors on the layout.

@enekomartinmartinez make sure to ping us here when the repository move and the corrections are done.
