### 2022-09-08 SDXorg/pysd#348 @enekomartinmartinez
@JamesPHoughton @alexprey @rogersamso @julienmalard @SimonStrong @ConnectedSystems @pbreach @DVRodri8 @ainar @keith-citrenbaum @JoonaTuovinen @rugue1 @mhy05 @marrobl @sepiatone
We have already migrated PySD!!!
It is recommendable that you update your URLs to point to `SDXorg` instead of `JamesPHoughton`
