+++ docs/requirements.txt
@@ -4,3 +4,4 @@ sphinx==4.2.0
 sphinx_rtd_theme==1.0.0
 readthedocs-sphinx-search==0.1.1
 jinja2==3.0.0
+mock
