+++ .github/workflows/spell-check.yml
@@ -1,0 +1,6 @@
+name: Codespell with annotations
+uses: codespell-project/actions-codespell@master
+
+on:
+  push:
+  pull_request:

[…]
