### 2022-09-08 SDXorg/pysd#348 @enekomartinmartinez
@JamesPHoughton RTD is not working in PR we need to check/update the link in RTD configuration and check the integration section
