### 2022-09-09 6d0a7d7
add test for property
