### 2022-09-09 0893709
add test for variable timestep and final_time
