### 2022-09-09 b2ef2cd
add test for variable timestep and final_time
