### 2022-09-09 #4329 @enekomartinmartinez
@sixpearls thanks for the updates in the TODOs and new suggestions.
I will let you know when I have completed them.
We do not test the Euler method itself as we test several models with the output of Vensim/Stella with the same method.
But I think it will be interesting having some convergence tests for the particular method as you suggest. This will probably take me some days as I have some other priorities but I hope that I can implement them before the end of the moth.
