+++ .github/workflows/ci.yml
@@ -22,6 +22,7 @@ jobs:
           python-version: ${{ matrix.python-version }}
       - name: Install dependencies
         run: |
+          pip install --upgrade pip setuptools
           pip install -r tests/requirements.txt
           pip install -e .
       - name: Test and coverage
