### 2022-09-09 #4329 @rogersamso
@sixpearls, as @enekomartinmartinez mentions, we benchmark the results of the PySD solver against those of Vensim and Stella for [all of these tests](https://github.com/SDXorg/test-models/tree/master/tests). So we should be relatively safe in that regard. Accordingly, even if @enekomartinmartinez has accepted to work on that in the coming weeks, I think this issue is currently pretty well handled in the library and should not be a blocking point.
