### 2022-09-09 #4329 @sixpearls
@rogersamso I asked about testing the mathematical properties of your implementation of the Euler solver or using an implementation that is already tested in [my first batch of comments](https://github.com/openjournals/joss-reviews/issues/4329#issuecomment-1213178569). Since this software package is claiming to implement this algorithm, the [JOSS review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#functionality) requires verifying its functionality.
