+++ requirements.txt
@@ -1,8 +1,7 @@
 pandas
 parsimonious
 xarray
-h5py
-netCDF4
+netCDF4<=1.5
 xlrd
 lxml
 regex
