### 2022-09-10 1495e80
conditional compression
