+++ pysd/builders/stan/stan_block_builder.py
@@ -1,4 +1,5 @@
 import os
+import itertools
 from pathlib import Path
 from typing import Union, List, Dict, Set, Sequence, Iterable
 from numbers import Number
@@ -16,10 +17,23 @@ if TYPE_CHECKING:
 
 
 class StanTransformedParametersBuilder:
-    def __init__(self, abstract_model: AbstractModel):
-        self.abstract_model = abstract_model
+    def __init__(self, stan_model_context, vensim_model_context):
+        self.stan_model_context = stan_model_context
+        self.vensim_model_context = vensim_model_context
+        self.abstract_model = self.vensim_model_context.abstract_model
 
-    def build_block(
+    def build_block(self, predictor_variable_names, outcome_variable_names, lookup_function_dict, function_name, stock_initial_values: Dict[str, str]):
+        self.code = IndentedString()
+        self.code += "transformed parameters {\n"
+        self.code.indent_level += 1
+        self.write_block(predictor_variable_names, outcome_variable_names, lookup_function_dict, function_name, stock_initial_values)
+        self.code.indent_level -= 1
+        self.code += "}\n"
+
+        return str(self.code)
+
+
+    def write_block(
         self,
         predictor_variable_names,
         outcome_variable_names,
@@ -27,10 +41,6 @@ class StanTransformedParametersBuilder:
         function_name,
         stock_initial_values: Dict[str, str]
     ):
-        self.code = IndentedString()
-        self.code += "transformed parameters {\n"
-        self.code.indent_level += 1
-
         argument_variables = []
         for var in predictor_variable_names:
             if isinstance(var, str):
@@ -43,6 +53,11 @@ class StanTransformedParametersBuilder:
         for element in self.abstract_model.sections[0].elements:
             stan_varname = vensim_name_to_identifier(element.name)
             variable_ast_dict[stan_varname] = element.components[0].ast
+
+        # Create variables defined through assignment
+        for statement in self.stan_model_context.sample_statements:
+            if statement.distribution_type == statement.assignment_dist:
+                self. code += f"real {statement.lhs_expr} = {''.join(statement.distribution_args)};\n"
 
         self.code += "// Initial ODE values\n"
         for outcome_variable_name in outcome_variable_names:
@@ -80,17 +95,14 @@ class StanTransformedParametersBuilder:
         for index, name in enumerate(outcome_variable_names, 1):
             self.code += f"array[n_t] real {name} = integrated_result[:, {index}];\n"
 
-        self.code.indent_level -= 1
-        self.code += "}\n"
-
-        return str(self.code)
-
 
 class StanParametersBuilder:
-    def __init__(self, stan_model_context: "StanModelContext"):
+    def __init__(self, stan_model_context: "StanModelContext", vensim_model_context: "VensimModelContext"):
         self.stan_model_context = stan_model_context
+        self.vensim_model_context = vensim_model_context
 
-    def build_block(self, data_variable_names: Tuple[str]):
+    def build_block(self):
+        data_variable_names = tuple(self.stan_model_context.stan_data.keys())
         code = IndentedString()
         code += "parameters{\n"
         code.indent_level += 1  # Enter parameters block
@@ -98,10 +110,12 @@ class StanParametersBuilder:
         added_parameters = set()
 
         for statement in self.stan_model_context.sample_statements:
-            for lhs_variable in statement.lhs_variables:
+            for lhs_variable in statement.lhs_variable:
                 if lhs_variable in data_variable_names:
                     continue
                 if lhs_variable in added_parameters:
+                    continue
+                if statement.distribution_type == statement.assignment_dist:
                     continue
[…]
