+++ pysd/builders/stan/stan_model.py
@@ -8,7 +8,7 @@ import cmdstanpy
 
 class SamplingStatement:
     lhs_expr: str
-    lhs_variables: Set[str]
+    lhs_variable: str
     rhs_variables: Set[str]
     distribution_type: str
     distribution_return_type: Type
@@ -19,15 +19,17 @@ class SamplingStatement:
 
     def __init__(self, lhs_expr, distribution_type, *distribution_args, lower=float("-inf"), upper=float("inf"), init_state=False):
         self.lhs_expr = lhs_expr
-        self.lhs_variables = set()
-        self.lhs_variables.update([node.id for node in ast.walk(ast.parse(lhs_expr)) if isinstance(node, ast.Name)])
+        lhs_variables = [node.id for node in ast.walk(ast.parse(lhs_expr)) if isinstance(node, ast.Name)]
+        assert len(lhs_variables) == 1, "The LHS expression for prior specification must contain just 1 parameter name"
+        self.lhs_variable = lhs_variables[0]
         self.distribution_type = distribution_type
-        self.distribution_args = distribution_args
+        self.distribution_args = tuple(str(x) for x in distribution_args)
         self.rhs_variables = set()
         self.rhs_variables.update([node.id for arg in self.distribution_args for node in ast.walk(ast.parse(str(arg))) if isinstance(node, ast.Name)])
         self.lower = lower
         self.upper = upper
         self.init_state = init_state
+        self.assignment_dist = "assignment"
 
     def __post_init__(self):
         if self.distribution_type in ("bernoulli", "binomial", "beta_binomial", "neg_binomial", "poisson"):
@@ -38,17 +40,53 @@ class SamplingStatement:
 
 
 @dataclass
+class StanDataEntry:
+    data_name: str
+    stan_type: str
+
+@dataclass
 class StanModelContext:
     initial_time: float
     integration_times: Iterable[float]
+    stan_data: Dict[str, StanDataEntry] = field(default_factory=dict)
     sample_statements: List[SamplingStatement] = field(default_factory=list)
     exposed_parameters: Set[str] = field(default_factory=set)
+
+    def identify_stan_data_types(self, data_dict):
+        def get_dims(obj):
+            try:
+                iter(obj)
+            except:
+                return None
+            else:
+                dim = len(obj)
+                inner_dim = get_dims(obj[0])
+                if inner_dim:
+                    return [dim] + inner_dim
+                else:
+                    return [dim]
+
+        for key, val in data_dict.items():
+            if isinstance(val, int):
+                self.stan_data[key] = StanDataEntry(key, "int")
+            elif isinstance(val, float):
+                self.stan_data[key] = StanDataEntry(key, "real")
+            else:
+                # Multidimensional data
+                dims = get_dims(val)
+                if not dims:
+                    raise Exception(f"Can't process data entry {key}.")
+                elif len(dims) == 1:
+                    self.stan_data[key] = StanDataEntry(key, f"vector[{dims[0]}]")
+                elif len(dims) == 2:
+                    self.stan_data[key] = StanDataEntry(key, f"array[{dims[0]}] vector[{dims[1]}]")
 
 
 class VensimModelContext:
     def __init__(self, abstract_model):
         self.variable_names = set()
         self.stock_variable_names = set()
+        self.abstract_model = abstract_model
 
         # Some basic checks to make sure the AM is compatible
         assert len(abstract_model.sections) == 1, "Number of sections in AbstractModel must be 1."
@@ -84,12 +122,14 @@ class VensimModelContext:
 
 
 class StanVensimModel:
-    def __init__(self, model_name: str, abstract_model, initial_time: float, integration_times: Iterable[Number]):
[…]
