### 2022-09-11 7e769f3
better handling of compression
