### 2022-09-12 d730341
add tests, fix export to csv
