+++ tests/pytest_pysd/pytest_pysd.py
@@ -8,8 +8,6 @@ import xarray as xr
 import netCDF4 as nc
 
 from pysd.tools.benchmarking import assert_frames_close
-from pysd.py_backend.output import OutputHandlerInterface, DatasetHandler, \
-    DataFrameHandler, ModelOutput
 
 import pysd
 
@@ -26,18 +24,7 @@ test_model_look = _root.joinpath(
     + "test_get_lookups_subscripted_args.mdl")
 test_model_data = _root.joinpath(
     "test-models/tests/get_data_args_3d_xls/test_get_data_args_3d_xls.mdl")
-test_model_constants = _root.joinpath(
-    "test-models/tests/get_constants_subranges/"
-    "test_get_constants_subranges.mdl"
-)
-test_variable_step = _root.joinpath(
-    "test-models/tests/control_vars/"
-    "test_control_vars.mdl"
-)
-test_model_numeric_coords = _root.joinpath(
-    "test-models/tests/subscript_1d_arrays/"
-    "test_subscript_1d_arrays.mdl"
-)
+
 
 more_tests = _root.joinpath("more-tests/")
 
@@ -1232,7 +1219,9 @@ class TestPySD():
         res = out.handler.ds
         assert isinstance(res, nc.Dataset)
         assert 'teacup_temperature' in res.variables
-        assert np.array_equal(res["time"][:].data, np.arange(0, 5, 2))
+        with catch_warnings(record=True) as w:
+            simplefilter("always")
+            assert np.array_equal(res["time"][:].data, np.arange(0, 5, 2))
         res.close()
 
     def test_default_returns_with_construction_functions(self):
@@ -1693,368 +1682,3 @@ class TestExportImport():
             Path('initial7.pic').unlink()
 
             assert_frames_close(stocks2, stocks)
-
-
-class TestOutputs():
-
-    def test_output_handler_interface(self):
-        # when the class does not inherit from OutputHandlerInterface, it must
-        # implement all the interface to be a subclass of
-        # OutputHandlerInterface.
-        # Add any additional Handler here.
-        assert issubclass(DatasetHandler, OutputHandlerInterface)
-        assert issubclass(DataFrameHandler, OutputHandlerInterface)
-
-        class ThatFollowsInterface:
-            """
-            This class does not inherit from OutputHandlerInterface, but it
-            overrides all its methods (it follows the interface).
-            """
-            def process_output(self, out_file):
-                pass
-
-            def initialize(self, model, capture_elements):
-                pass
-
-            def update(self, model, capture_elements):
-                pass
-
-            def postprocess(self, **kwargs):
-                pass
-
-            def add_run_elements(self, capture_elemetns):
-                pass
-
-        # eventhough it does not inherit from OutputHandlerInterface, it is
-        # considered a subclass, because it follows the interface
-        assert issubclass(ThatFollowsInterface, OutputHandlerInterface)
-
-        class IncompleteHandler:
-            """
-            Class that does not follow the full interface
-            (add_run_elements is missing).
-            """
-            def initialize(self, model, capture_elements):
-                pass
-
-            def update(self, model, capture_elements):
-                pass
-
-            def postprocess(self, **kwargs):
-                pass
-
-        # It does not inherit from OutputHandlerInterface and does not fulfill
-        # its interface
-        assert not issubclass(IncompleteHandler, OutputHandlerInterface)
-
-        class EmptyHandler(OutputHandlerInterface):
-            """
[…]
