### 2022-09-12 edf9617
add tests, fix export to csv
