+++ tests/pytest_pysd/pytest_output.py
@@ -1,0 +1,467 @@
+from pathlib import Path
+from warnings import simplefilter, catch_warnings
+
+import pytest
+import numpy as np
+import pandas as pd
+import xarray as xr
+import netCDF4 as nc
+
+from pysd.tools.benchmarking import assert_frames_close
+from pysd.py_backend.output import OutputHandlerInterface, DatasetHandler, \
+    DataFrameHandler, ModelOutput
+
+import pysd
+
+
+_root = Path(__file__).parent.parent
+
+test_model_look = _root.joinpath(
+    "test-models/tests/get_lookups_subscripted_args/"
+    + "test_get_lookups_subscripted_args.mdl")
+test_model_constants = _root.joinpath(
+    "test-models/tests/get_constants_subranges/"
+    "test_get_constants_subranges.mdl"
+)
+test_model_numeric_coords = _root.joinpath(
+    "test-models/tests/subscript_1d_arrays/"
+    "test_subscript_1d_arrays.mdl"
+)
+test_variable_step = _root.joinpath(
+    "test-models/tests/control_vars/"
+    "test_control_vars.mdl"
+)
+
+
+class TestOutput():
+
+    def test_output_handler_interface(self):
+        # when the class does not inherit from OutputHandlerInterface, it must
+        # implement all the interface to be a subclass of
+        # OutputHandlerInterface.
+        # Add any additional Handler here.
+        assert issubclass(DatasetHandler, OutputHandlerInterface)
+        assert issubclass(DataFrameHandler, OutputHandlerInterface)
+
+        class ThatFollowsInterface:
+            """
+            This class does not inherit from OutputHandlerInterface, but it
+            overrides all its methods (it follows the interface).
+            """
+            def process_output(self, out_file):
+                pass
+
+            def initialize(self, model, capture_elements):
+                pass
+
+            def update(self, model, capture_elements):
+                pass
+
+            def postprocess(self, **kwargs):
+                pass
+
+            def add_run_elements(self, capture_elemetns):
+                pass
+
+        # eventhough it does not inherit from OutputHandlerInterface, it is
+        # considered a subclass, because it follows the interface
+        assert issubclass(ThatFollowsInterface, OutputHandlerInterface)
+
+        class IncompleteHandler:
+            """
+            Class that does not follow the full interface
+            (add_run_elements is missing).
+            """
+            def initialize(self, model, capture_elements):
+                pass
+
+            def update(self, model, capture_elements):
+                pass
+
+            def postprocess(self, **kwargs):
+                pass
+
+        # It does not inherit from OutputHandlerInterface and does not fulfill
+        # its interface
+        assert not issubclass(IncompleteHandler, OutputHandlerInterface)
+
+        class EmptyHandler(OutputHandlerInterface):
+            """
+            When the class DOES inherit from OutputHandlerInterface, but does
+            not override all its abstract methods, then it cannot be
+            instantiated
+            """
+            pass
+
+        # it is a subclass because it inherits from it
+        assert issubclass(EmptyHandler, OutputHandlerInterface)
+
+        # it cannot be instantiated because it does not override all abstract
+        # methods
+        with pytest.raises(TypeError):
+            empty = EmptyHandler()
+
+        # calling methods that are not overriden returns NotImplementedError
[…]
