### 2022-09-12 SDXorg/pysd#348 @julienmalard
@enekomartinmartinez Wonderful! Regarding my contributions, I think they were mostly limited to adding Unicode support some time back. https://github.com/SDXorg/pysd/pull/179
