### 2022-09-12 #4329 @pdebuyl
Hi @sixpearls is this the kind of verification that you have in mind?

![teacup](https://user-images.githubusercontent.com/1121690/189694197-1a2f52c9-dddc-4716-8282-c94b9435c108.png)

it is the teacup example with an exponential of "70 + (T0-70) * exp(t / t0)"
