### 2022-09-12 #4329 @sixpearls
@pdebuyl  Yes, this a stable, linear, first-order system with a known analytic solution. The verification should show error as the step-size of the Euler method changes. Best,
