### 2022-09-13 edd6c9a
Remove netcdf4 from dependencies
