### 2022-09-13 53b8b97
Update getting_started.rst
