### 2022-09-14 e05be9f
Update whats_new.rst
