+++ tests/pytest_pysd/pytest_output.py
@@ -1,5 +1,5 @@
 from pathlib import Path
-from warnings import simplefilter, catch_warnings
+import shutil
 
 import pytest
 import numpy as np
@@ -14,27 +14,38 @@ from pysd.py_backend.output import OutputHandlerInterface, DatasetHandler, \
 import pysd
 
 
-_root = Path(__file__).parent.parent
-
-test_model_look = _root.joinpath(
+test_model_look = Path(
     "test-models/tests/get_lookups_subscripted_args/"
-    + "test_get_lookups_subscripted_args.mdl")
-test_model_constants = _root.joinpath(
+    "test_get_lookups_subscripted_args.mdl"
+)
+test_model_constants = Path(
     "test-models/tests/get_constants_subranges/"
     "test_get_constants_subranges.mdl"
 )
-test_model_numeric_coords = _root.joinpath(
+test_model_numeric_coords = Path(
     "test-models/tests/subscript_1d_arrays/"
     "test_subscript_1d_arrays.mdl"
 )
-test_variable_step = _root.joinpath(
+test_variable_step = Path(
     "test-models/tests/control_vars/"
     "test_control_vars.mdl"
 )
-test_partial_definitions = _root.joinpath(
+test_partial_definitions = Path(
     "test-models/tests/partial_range_definitions/"
     "test_partial_range_definitions.mdl"
 )
+
+
+@pytest.fixture
+@pytest.mark.filterwarnings("ignore")
+def model(_root, tmp_path, model_path):
+    """
+    Copy model to the tmp_path and translate it
+    """
+
+    target = tmp_path / model_path.parent
+    shutil.copytree(_root / model_path.parent, target)
+    return pysd.read_vensim(target / model_path.name)
 
 
 class TestOutput():
@@ -124,151 +135,167 @@ class TestOutput():
             EmptyHandler.add_run_elements(
                 EmptyHandler, "model", "capture")
 
-    def test_output_nc(self, shared_tmpdir):
-        model = pysd.read_vensim(test_model_look)
-        model.progress = False
+    @pytest.mark.parametrize(
+        "model_path,dims,values",
+        [
+            (
+                test_model_look,
+                {
+                    "Rows": 2,
+                    "Dim": 2,
+                    "time": 61
+                },
+                {
+                    "lookup_1d_time": (("time",), None),
+                    "d2d": (("time", "Rows", "Dim"), None),
+                    "initial_time": (tuple(), 0),
+                    "final_time": (tuple(), 30),
+                    "saveper": (tuple(), 0.5),
+                    "time_step": (tuple(), 0.5)
+                }
 
-        out_file = shared_tmpdir.joinpath("results.nc")
+            ),
+            (
+                test_model_constants,
+                {
+                    "dim1": 5,
+                    "dim1a": 2,
+                    "dim1c": 3,
+                    'time': 2
+                },
+                {
+                    "constant": (
+                        ("dim1",),
+                        np.array([0.,  0.,  1., 15., 50.])
+                    )
+                }
+            ),
+            (
+                test_model_numeric_coords,
+                {
+                    "One Dimensional Subscript": 3,
+                    'time': 101
+                },
+                {
+                    "rate_a": (
+                        ("One Dimensional Subscript",),
+                        np.array([0.01, 0.02, 0.03])),
+                    "stock_a": (
+                        ("time", "One Dimensional Subscript"),
+                        np.array([
+                            np.arange(0, 1.0001, 0.01),
+                            np.arange(0, 2.0001, 0.02),
+                            np.arange(0, 3.0001, 0.03)],
+                            dtype=float).transpose()),
+                    "time": (("time",), np.arange(0.0, 101.0, 1.0))
+                }
+            ),
+            (
[…]
