### 2022-09-16 f638ac5
Three checks with draws2data, data2draws function
