+++ tests/pytest_pysd/pytest_pysd.py
@@ -1,5 +1,4 @@
 from pathlib import Path
-from warnings import simplefilter, catch_warnings
 
 import pytest
 import pandas as pd
@@ -14,19 +13,17 @@ import pysd
 # TODO replace test paths by fixtures and translate and run the models
 # in temporal directories
 
-_root = Path(__file__).parent.parent
-
-test_model = _root.joinpath("test-models/samples/teacup/teacup.mdl")
-test_model_subs = _root.joinpath(
+test_model = Path("test-models/samples/teacup/teacup.mdl")
+test_model_subs = Path(
     "test-models/tests/subscript_2d_arrays/test_subscript_2d_arrays.mdl")
-test_model_look = _root.joinpath(
+test_model_look = Path(
     "test-models/tests/get_lookups_subscripted_args/"
     + "test_get_lookups_subscripted_args.mdl")
-test_model_data = _root.joinpath(
+test_model_data = Path(
     "test-models/tests/get_data_args_3d_xls/test_get_data_args_3d_xls.mdl")
 
 
-more_tests = _root.joinpath("more-tests/")
+more_tests = Path("more-tests")
 
 test_model_constant_pipe = more_tests.joinpath(
     "constant_pipeline/test_constant_pipeline.mdl")
@@ -34,8 +31,8 @@ test_model_constant_pipe = more_tests.joinpath(
 
 class TestPySD():
 
-    def test_run(self):
-        model = pysd.read_vensim(test_model)
+    @pytest.mark.parametrize("model_path", [test_model])
+    def test_run(self, model):
         stocks = model.run()
         # return a dataframe
         assert isinstance(stocks, pd.DataFrame)
@@ -46,7 +43,7 @@ class TestPySD():
         # there are no null values in the set
         assert stocks.notnull().all().all()
 
-    def test_run_ignore_missing(self):
+    def test_run_ignore_missing(self, _root):
         model_mdl = _root.joinpath(
             'test-models/tests/get_with_missing_values_xlsx/'
             + 'test_get_with_missing_values_xlsx.mdl')
@@ -73,32 +70,46 @@ class TestPySD():
             # errors for missing values
             pysd.load(model_py, missing_values="raise")
 
-    def test_run_includes_last_value(self):
-        model = pysd.read_vensim(test_model)
+    @pytest.mark.parametrize("model_path", [test_model])
+    def test_run_includes_last_value(self, model):
         res = model.run()
         assert res.index[-1] == model.components.final_time()
 
-    def test_run_build_timeseries(self):
-        model = pysd.read_vensim(test_model)
+    @pytest.mark.parametrize("model_path", [test_model])
+    def test_run_build_timeseries(self, model):
         res = model.run(final_time=7, time_step=2, initial_condition=(3, {}))
 
         actual = list(res.index)
         expected = [3.0, 5.0, 7.0]
         assert actual == expected
 
-    def test_run_progress(self):
+    @pytest.mark.parametrize("model_path", [test_model])
+    def test_run_progress(self, model):
         # same as test_run but with progressbar
-        model = pysd.read_vensim(test_model)
         stocks = model.run(progress=True)
         assert isinstance(stocks, pd.DataFrame)
         assert "Teacup Temperature" in stocks.columns.values
         assert len(stocks) > 3
         assert stocks.notnull().all().all()
 
-    def test_run_return_timestamps(self):
-        """Addresses https://github.com/JamesPHoughton/pysd/issues/17"""
+    @pytest.mark.parametrize(
+        "model_path",
+        [Path("test-models/tests/control_vars/test_control_vars.mdl")])
+    def test_run_progress_dynamic(self, model):
+        # same as test_run but with progressbar
+        warn_message = r"The progressbar is not compatible with dynamic "\
[…]
