### 2022-09-16 4cb3f1e
preliminar implementation of nc_to_csv
