+++ pysd/tools/ncfiles.py
@@ -1,43 +1,248 @@
+"""
+Tools for importing and converting netCDF files generated from simulations run
+using PySD.
+"""
+
 from csv import QUOTE_NONE
 import xarray as xr
 import pandas as pd
-import itertools
 from pysd.py_backend.utils import xrsplit
+from pathlib import Path
+import warnings
 
-def nc_to_csv(ncfile, outfile="result.csv", subset=None):
+def load_nc_file(ncfile):
+    """
+    Loads netCDF file into xarray Dataset.
 
-    ds = xr.open_dataset(ncfile)
+    Parameters
+    ----------
+    ncfile: str
+        Path to the netCDF file to process.
 
-    savedict = {}
+    Returns
+    -------
+    xarray.Dataset
 
-    for name, da in ds.data_vars.items():
-        # TODO only if in subset
-        #if name in subset:
-        dims = da.dims
+    """
+    if isinstance(ncfile, (str, Path)):
+        if isinstance(ncfile, str):
+            ncfile = Path(ncfile)
+        if not ncfile.is_file():
+            raise FileNotFoundError(f"{ncfile} could not be found.")
+        if not ncfile.suffix == ".nc":
+            raise TypeError("Input file must have nc extension.")
+    else:
+        raise TypeError(f"Invalid file path type: {type(ncfile)}.\n"
+                        "Please provide string or pathlib Path")
 
-        if not da.dims:
-            savedict[name] = da.values.tolist()
+    # open netCDF file
+    return xr.open_dataset(ncfile)
+
+def _validate_ds_subset(ds, subset):
+    """
+    If subset = None, it returns a list with all variables in the ds Dataset.
+    Else, if vars in subset are present in ds, it returns them, else it warns
+    the user.
+
+    Parameters
+    ----------
+    ds: xarray.Dataset
+
+    subset: list
+        Subset of variable names in the dataset.
+
+    Returns
+    -------
+    new_subset:
+    """
+    # process subset list
+    if not subset: # use all names
+        new_subset = [name for name  in ds.data_vars.keys()]
+    else:
+        if not isinstance(subset, list):
+            raise TypeError("Subset argument must be a list of strings.")
         else:
-            # TODO make a function that returns coordinates of all elements
-            # instead of using xrsplit, because in the end we just use it to
-            # index the DataArray directly
-            split_array = xrsplit(da)
-            for elem in split_array:
-                coords = {dim: str(elem.coords[dim].values) for dim in dims if dim != "time"}
-                subs = "[" + ",".join(coords.values()) + "]"
-                savedict[name + subs] = da.loc[coords].values
+            if not all(map(lambda x: isinstance(x, str), subset)):
+                raise TypeError("subset argument must be a list of strings.")
+
+        new_subset = [name if name in ds.data_vars.keys()
+                      else warnings.warn(f"{name} not in Dataset.")
+                      for name in subset]
+    return new_subset
+
+def _ds_var_to_dict_item(savedict, ds, name):
+    """
+    Convert xarray Dataset variable to dict items for each combination of
+    dimensions.
+
+    Parameters
+    ----------
+    savedict: dict
+        Dictionary to which keys and values are added.
+    ds: xr.Dataset
+        Dataset containing the variable to convert into a dictionary.
+    name: str
+        Name of the variable to extract from the Dataset.
+
+    Returns
+    -------
+    None
+
+    """
+    da = ds[name]
+    dims = da.dims
+
+    if not dims:
+        savedict[name] = da.values.tolist()
+    else:
+        for elem in xrsplit(da):
[…]
