+++ pysd/tools/ncfiles.py
@@ -1,43 +1,277 @@
+"""
+Tools for importing and converting netCDF files generated from simulations run
+using PySD.
+"""
+
 from csv import QUOTE_NONE
 import xarray as xr
 import pandas as pd
-import itertools
 from pysd.py_backend.utils import xrsplit
+from pathlib import Path
+import warnings
+from typing import Union, Optional
 
-def nc_to_csv(ncfile, outfile="result.csv", subset=None):
+ChunkTypes = Union[str, int, dict]
 
-    ds = xr.open_dataset(ncfile)
+def load_nc_file(ncfile: Union[str, Path],
+                 chunks: Optional[ChunkTypes]={}) -> xr.Dataset:
+    """
+    Loads netCDF file into xarray Dataset.
 
-    savedict = {}
+    Parameters
+    ----------
+    ncfile: str or pathlib.Path
+        Path to the netCDF file to process.
+    chunks: int, dict, auto or None (optional)
+        chunks argument from the xarray.load_dataset method.
+        Used to load the dataset into dask arrays. The default
+        value of this argument will load the dataset with dask
+        using engine preferred chunks if exposed by the backend,
+        otherwise with a single chunk for all arrays. See dask
+        chunking for more details.
 
-    for name, da in ds.data_vars.items():
-        # TODO only if in subset
-        #if name in subset:
-        dims = da.dims
+    Returns
+    -------
+    xarray.Dataset
 
-        if not da.dims:
-            savedict[name] = da.values.tolist()
-        else:
-            # TODO make a function that returns coordinates of all elements
-            # instead of using xrsplit, because in the end we just use it to
-            # index the DataArray directly
-            split_array = xrsplit(da)
-            for elem in split_array:
-                coords = {dim: str(elem.coords[dim].values) for dim in dims if dim != "time"}
-                subs = "[" + ",".join(coords.values()) + "]"
-                savedict[name + subs] = da.loc[coords].values
+    """
+    if not isinstance(ncfile, (str, Path)):
+        raise TypeError(f"Invalid file path type: {type(ncfile)}.\n"
+                        "Please provide string or pathlib Path")
+
+    if isinstance(ncfile, str):
+        ncfile = Path(ncfile)
+
+    if not ncfile.is_file():
+        raise FileNotFoundError(f"{ncfile} could not be found.")
+
+    if not ncfile.suffix == ".nc":
+        raise TypeError("Input file must have nc extension.")
+
+    # open netCDF file
+    return xr.open_dataset(ncfile, engine="netcdf4", chunks=chunks)
+
+def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
+    """
+    If subset = None, it returns a list with all variables in the ds Dataset.
+    Else, if vars in subset are present in ds, it returns them, else it warns
+    the user.
+
+    Parameters
+    ----------
+    subset: list
+        Subset of variable names in the xarray Dataset.
+
+    """
+    # process subset list
+    if not subset: # use all names
+        new_subset = [name for name  in ds.data_vars.keys()]
+    else:
+        if not isinstance(subset, list) or \
+             not all(map(lambda x: isinstance(x, str), subset)):
+            raise TypeError("Subset argument must be a list of strings.")
+
+        new_subset = [name if name in ds.data_vars.keys()
+                      else warnings.warn(f"{name} not in Dataset.")
+                      for name in subset]
+    return new_subset
+
+def _ds_var_to_dict_item(savedict: dict, ds: xr.Dataset, name: str) -> None:
+    """
+    Convert xarray Dataset variable to dict items for each combination of
+    dimensions.
+
+    Parameters
+    ----------
[…]
