+++ nc_export.py
@@ -1,0 +1,3 @@
+from pysd.tools.ncfiles import nc_to_csv
+
+nc_to_csv("../netcdf_tests/results.nc", "results.tab")
