### 2022-09-18 53a2576
Bug fix and analysis notebook for preypredator
