### 2022-09-18 a002f5d
Add convergence tests
