+++ tests/pytest_integration/pytest_integration_euler.py
@@ -1,0 +1,213 @@
+from pathlib import Path
+
+import pytest
+import numpy as np
+import pandas as pd
+
+
+def harmonic_position(t, x0, k, m):
+    """
+    Position for the simple harmonic oscillator
+    'test-models/samples/simple_harmonic_oscillator
+     /simple_harmonic_oscillator.mdl'
+    """
+    return x0*np.cos(np.sqrt(k/m)*t)
+
+
+def harmonic_speed(t, x0, k, m):
+    """
+    Speed for the simple harmonic oscillator
+    'test-models/samples/simple_harmonic_oscillator
+     /simple_harmonic_oscillator.mdl'
+    """
+    return - x0*np.sqrt(k/m)*np.sin(np.sqrt(k/m)*t)
+
+
+@pytest.mark.parametrize(
+    "model_path,f,f2,stocks,arguments,integration_frame,ts_log,ts_rmse",
+    [
+        # model_path: model path: pathlib.Path object
+        # f: stocks analytical solutions: tuple of funcs
+        # f2: stocks analytical solutions second derivative: tuple of funcs
+        # stocks: stock names in the model: tuple of strings
+        # arguments: arguments of fs and f2s in the model: tuple of strings
+        # integration_frame: minimum and maximum time to find solutions: tuple
+        # ts_log: logarithmic range of time steps in base 10: tuple
+        # ts_rmse: sorted time step for RMSE test: iterable
+        (
+            Path("test-models/samples/teacup/teacup.mdl"),
+            (lambda t, T0, TR, ct: TR + (T0-TR)*np.exp(-t/ct),),
+            (lambda t, T0, TR, ct: (T0-TR)*np.exp(-t/ct)/ct**2,),
+            ("Teacup Temperature",),
+            ("Teacup Temperature", "Room Temperature", "Characteristic Time"),
+            (0, 40),
+            (1, -5),
+            [10, 5, 1, 0.5, 0.1, 0.05, 0.01]
+        ),
+        (
+            Path("test-models/samples/simple_harmonic_oscillator/"
+                 "simple_harmonic_oscillator.mdl"),
+            (
+                harmonic_position,
+                harmonic_speed
+            ),
+            (
+                lambda t, x0, k, m: -k/m*harmonic_position(t, x0, k, m),
+                lambda t, x0, k, m: -k/m*harmonic_speed(t, x0, k, m)
+            ),
+            ("position", "speed"),
+            ("initial position", "elastic constant", "mass"),
+            (0, 40),
+            (-1, -5),
+            [10, 5, 1, 0.5, 0.1, 0.05, 0.01]
+        )
+    ],
+    ids=["teacup", "harmonic"]
+)
+class TestEulerConvergence:
+    """
+    Test for splitting Vensim views in modules and submodules
+    """
+    # Number of points to compute the tests
+    n_points_lte = 30
+
+    def test_local_truncation_error(self, model, f, f2, stocks, arguments,
+                                    integration_frame, ts_log, ts_rmse):
+        """
+        Test the locat truncation error (LTE).
+        LTE = y_1 - y(t_0+h) = 0.5*h**2*y''(x) for x in [t_0, t_0+h]
+
+        where y_1 = y(t_0) + h*f(t_0, y(t_0)) and
+
+        Generates n_points_lte in the given integration frame and test the
+        convergence with logarithmically uniform split time_steps.
+
+        Parameter
+        ---------
+        model: pysd.py_backend.model.Model
+            The model to integrate.
+        f: tuple of functions
+            The functions of the analytical solution of each stock.
+        f2: tuple of functions
+            The second derivative of the functions of the analytical
+            solution of each stock.
+        stocks: tuple of strings
+            The name of the stocks.
+        arguments: tuple of strings
[…]
