### 2022-09-18 #4329 @enekomartinmartinez
Hi @pdebuyl and @sixpearls !

I have added some tests for Euler method convergence in the following file (still not merged to master and thus not merged to paper branch):
https://github.com/SDXorg/pysd/blob/migration/tests/pytest_integration/pytest_integration_euler.py

I have included the following tests:

- Local truncation error test in some random points with several time steps, which is expected to be equal to 1/2 h² y''(x) for x in [t0, t0+h]
- Convergence test which checks that RMSE decreases when the time step decreases.

I have used the teacup model and a simple harmonic oscillator model.

Let me know if this is what you expected.
