### 2022-09-18 3a5f985
Remove contributing
