+++ CONTRIBUTING.md
@@ -1,7 +1,0 @@
-PySD is currently a community-maintained project, any contribution is welcome.
-
-The code for this package is available at: https://github.com/SDXorg/pysd
-
-If you find a bug, or are interested in a particular feature, see [reporting bugs](https://pysd.readthedocs.io/en/latest/reporting_bugs.html).
-
-If you are interested in contributing to the development of PySD, see the [developer documentation](https://pysd.readthedocs.io/en/latest/development/development_index.html) listed above.
