+++ MANIFEST.in
@@ -1,6 +1,5 @@
 include requirements.txt
 include README.md
-include CONTRIBUTING.md
 include LICENSE
 include docs/images/PySD_Logo*
 graft pysd/translators/*/parsing_grammars
