+++ docs/generate_tables.py
@@ -1,6 +1,6 @@
-import pandas as pd
 from pathlib import Path
 
+import pandas as pd
 
 def generate(table, columns, output):
     """Generate markdown table."""
