+++ docs/requirements.txt
@@ -5,3 +5,4 @@ sphinx_rtd_theme==1.0.0
 readthedocs-sphinx-search==0.1.1
 jinja2==3.0.0
 mock
+pandas
