+++ .github/workflows/ci.yml
@@ -25,7 +25,7 @@ jobs:
           pip install -r tests/requirements.txt
           pip install -e .
       - name: Test and coverage
-        run: pytest tests/ --cov=pysd
+        run: python -m pytest tests/ --cov=pysd
       - name: Coveralls
         if: ${{ matrix.python-version == 3.7 && matrix.os == 'ubuntu-latest' }}
         env:
