+++ docs/whats_new.rst
@@ -1,6 +1,6 @@
 What's New
 ==========
-v3.7.0 (2022/09/xx)
+v3.7.0 (2022/09/19)
 -------------------
 
 New Features
