### 2022-09-19 2e37be7
Remove old deprecation warning
