+++ docs/whats_new.rst
@@ -41,7 +41,7 @@ Internal Changes
 - Update License. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 - Include `Maintained? Yes` and `Contributions welcome` badges. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 - Update links to the new repository location. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
-- Reduce relative precission from 1e-10 to 1e-5 to compute the saving times and final time. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Reduce relative precision from 1e-10 to 1e-5 to compute the saving times and final time. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 - Add convergence tests for euler integration method. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 - Include build docs check in the GHA workflow to avoid warnings with sphinx. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 
