+++ tests/pytest.ini
@@ -3,5 +3,4 @@ python_files = pytest_*/**/*.py pytest_*/*.py
 filterwarnings =
     error
     ignore:Creating an ndarray from ragged nested sequences
-    ignore:distutils Version classes are deprecated. Use packaging.version instead
     ignore:`np.bool` is a deprecated alias for the builtin `bool`. To silence this warning, use `bool` by itself. Doing this will not modify any behavior and is safe. If you specifically wanted the numpy scalar type, use `np.bool_` here.
