### 2022-09-19 afc8702
migrate to SDXorg (#362)
* migrate to SDXorg

* Update  crashing links

* Add cookbook note in getting_started

* Update whats_new.rst

* reduce relative precision

* Add convergence tests

* Remove contributing

* Add building docs workflow

* Update about

* Fix docstring

* minor fixes

* Remove old deprecation warning
