+++ docs/whats_new.rst
@@ -1,11 +1,11 @@
 What's New
 ==========
-v3.7.0 (to be released)
+v3.7.0 (2022/09/19)
 -------------------
 
 New Features
 ~~~~~~~~~~~~
-- Simulation results can now be stored as netCDF4 files (:issue:`355`). (`@rogersamso <https://github.com/rogersamso>`_)
+- Simulation results can now be stored as netCDF4 files. (`@rogersamso <https://github.com/rogersamso>`_)
 - The CLI also accepts netCDF4 file paths after the -o argument. (`@rogersamso <https://github.com/rogersamso>`_)
 
 Breaking changes
@@ -23,7 +23,9 @@ Bug fixes
 
 Documentation
 ~~~~~~~~~~~~~
-- Adds Storing simulation results on a file section in the getting started page. (`@rogersamso <https://github.com/rogersamso>`_)
+- Add `Storing simulation results on a file` section in the :doc:`getting_started` page. (`@rogersamso <https://github.com/rogersamso>`_)
+- Include cookbook information in the :doc:`getting_started` page. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Include an introduction of main historical changes in the :doc:`about` page. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 
 Performance
 ~~~~~~~~~~~
@@ -35,7 +37,13 @@ Internal Changes
 - Add netCDF4 dependency for tests. (`@rogersamso <https://github.com/rogersamso>`_)
 - Improve warning message when replacing a stock with a parameter.  (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 - Include more pytest parametrizations in some test and make them translate the models in temporary directories.  (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
-
+- Include lychee-action in the GHA workflow to check the links. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Update License. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Include `Maintained? Yes` and `Contributions welcome` badges. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Update links to the new repository location. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Reduce relative precision from 1e-10 to 1e-5 to compute the saving times and final time. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Add convergence tests for euler integration method. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
+- Include build docs check in the GHA workflow to avoid warnings with sphinx. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 
 v3.6.1 (2022/09/05)
 -------------------
@@ -60,15 +68,15 @@ Performance
 
 Internal Changes
 ~~~~~~~~~~~~~~~~
-- Set :py:mod:`parsimonius` requirement to 0.9.0 to avoid a breaking-change in the newest version. Pending to update PySD to run it with :py:mod:`parsimonious` 0.10.0.
+- Set :py:mod:`parsimonius` requirement to 0.9.0 to avoid a breaking-change in the newest version. Pending to update PySD to run it with :py:mod:`parsimonious` 0.10.0. (`@enekomartinmartinez <https://github.com/enekomartinmartinez>`_)
 
 v3.6.0 (2022/08/31)
 -------------------
 
 New Features
 ~~~~~~~~~~~~
[…]
