+++ CONTRIBUTING.md
@@ -1,3 +1,0 @@
-..under construction..
-
-We are super welcoming of contributions, here's how:
