+++ MANIFEST.in
@@ -1,5 +1,5 @@
 include requirements.txt
 include README.md
-include docs/images/PySD_Logo*
 include LICENSE
+include docs/images/PySD_Logo*
 graft pysd/translators/*/parsing_grammars

[…]
