### 2022-09-19 f30e36b
Fix bugs with pandas 1.5.0
