### 2022-09-19 49785ac
Fix bugs with pandas 1.5.0
