### 2022-09-19 0837f68
avoids changing the state of the argument
