### 2022-09-20 #4329 @enekomartinmartinez
@pdebuyl @sixpearls required changes and new tests have already included in the paper branch.

About the non-blocking suggestion:
- There is already a short description of the tests in https://pysd.readthedocs.io/en/master/development/guidelines.html#test-suite I will try to improve it in the future but, currently, I don't have time.
- About the structure section, I am thinking about reordering the documentation as some other RTD/sphinx using libraries have done (e.g. [xarray](https://docs.xarray.dev/en/stable/)). Therefore, I will wait until I have time to include that structure to organize the section better.
