### 2022-09-20 #4329 @enekomartinmartinez
@pdebuyl I have recently changed my affiliation. I will keep the CREAF affiliation in the paper as it is the place where my work in PySD was done. However, @rogersamso suggested that I could add my new affiliation as a footnote saying "Currently at...". My current superior also said that it was not necessary but up to me to include my new affiliation in the same way.

Is there any standard of doing that in JOSS? I couldn't find any example in the documentation.
