### 2022-09-20 14bbd40
prototype with parallel processing of DataArrays
