+++ pysd/tools/ncfiles.py
@@ -4,17 +4,18 @@ using PySD.
 """
 
 from csv import QUOTE_NONE
+from lib2to3.pgen2.token import OP
 import xarray as xr
 import pandas as pd
-from pysd.py_backend.utils import xrsplit
 from pathlib import Path
 import warnings
 from typing import Union, Optional
+import itertools
 
-ChunkTypes = Union[str, int, dict]
+TypeChunks = Union[int, dict, str]
 
 def load_nc_file(ncfile: Union[str, Path],
-                 chunks: Optional[ChunkTypes]={}) -> xr.Dataset:
+                 chunks: Optional[TypeChunks]=None) -> xr.Dataset:
     """
     Loads netCDF file into xarray Dataset.
 
@@ -22,13 +23,10 @@ def load_nc_file(ncfile: Union[str, Path],
     ----------
     ncfile: str or pathlib.Path
         Path to the netCDF file to process.
-    chunks: int, dict, auto or None (optional)
-        chunks argument from the xarray.load_dataset method.
-        Used to load the dataset into dask arrays. The default
-        value of this argument will load the dataset with dask
-        using engine preferred chunks if exposed by the backend,
-        otherwise with a single chunk for all arrays. See dask
-        chunking for more details.
+    chunks: int, dict, str, None (optional)
+        Check xarray.open_dataset for potential values for this argument.
+        Frequent values are -1, to load each DataArray in a single chunk, or
+        None, to not use dask array. Using chunks requires installing dask.
 
     Returns
     -------
@@ -48,8 +46,10 @@ def load_nc_file(ncfile: Union[str, Path],
     if not ncfile.suffix == ".nc":
         raise TypeError("Input file must have nc extension.")
 
-    # open netCDF file
-    return xr.open_dataset(ncfile, engine="netcdf4", chunks=chunks)
+    if chunks:
+        return xr.open_dataset(ncfile, engine="netcdf4", chunks=chunks)
+
+    return xr.open_dataset(ncfile, engine="netcdf4")
 
 def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
     """
@@ -76,37 +76,74 @@ def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
                       for name in subset]
     return new_subset
 
-def _ds_var_to_dict_item(savedict: dict, ds: xr.Dataset, name: str) -> None:
+def _time_dim_from_da(da: xr.DataArray,
+                      dims: list,
+                      coords: tuple
+                      ) -> tuple:
     """
-    Convert xarray Dataset variable to dict items for each combination of
-    dimensions.
+    Function to extract all values along the time dimension with fixed
+    coordinates for the other dimensions as defined in coords.
+
 
     Parameters
     ----------
-    savedict: dict
-        Dictionary to which keys and values are added.
-    ds: xr.Dataset
-        Dataset containing the variable to convert into a dictionary.
-    name: str
-        Name of the variable to extract from the Dataset.
+    da: xr.Dataset
+        Dataset to be indexed.
+    dims: list
+        Dimensions along which the DataArray will be indexed.
+    coords: tuple
+        Coordinate names for each of the dimensons in the dims list.
 
     Returns
     -------
-    None
+    A tuple consisting of the string
+    var_name[dim1_coord, dim2_coord, ....] in the first index, and the
+    indexed data as the second index.
 
     """
-    da = ds[name]
-    dims = da.dims
+    name = da.attrs["Py Name"]
+    idx = dict(zip(dims, coords))
+    subs = "[" + ",".join(coords) + "]"
+    return name + subs, da.loc[idx].values
 
-    if not dims:
+def split_data_array(da):
+
+    if not da.dims:
+        savedict = {}
+        name = da.attrs["Py Name"]
[…]
