+++ pysd/tools/ncfiles.py
@@ -5,6 +5,7 @@ using PySD.
 
 from csv import QUOTE_NONE
 from lib2to3.pgen2.token import OP
+from tkinter.ttk import Progressbar
 import xarray as xr
 import pandas as pd
 from pathlib import Path
@@ -12,27 +13,29 @@ import warnings
 from typing import Union, Optional
 import itertools
 
-TypeChunks = Union[int, dict, str]
 
-def load_nc_file(ncfile: Union[str, Path],
-                 chunks: Optional[TypeChunks]=None) -> xr.Dataset:
+def open_nc_file(ncfile: Union[str, Path],
+                 parallel: Optional[bool]=False) -> xr.Dataset:
     """
-    Loads netCDF file into xarray Dataset.
+    Loads netCDF file into xarray Dataset. It's basically a wrapper to
+    xr.open_dataset to simplify the interface for pysd use case (loading
+    simulation results).
 
     Parameters
     ----------
     ncfile: str or pathlib.Path
         Path to the netCDF file to process.
-    chunks: int, dict, str, None (optional)
-        Check xarray.open_dataset for potential values for this argument.
-        Frequent values are -1, to load each DataArray in a single chunk, or
-        None, to not use dask array. Using chunks requires installing dask.
+    parallel: bool
+        When True, the Dataset is opened using chunks=-1 (see xarray
+        documentation for details), and when it's False, chunks=None.
 
     Returns
     -------
     xarray.Dataset
 
     """
+
+    # we run all these checks because xarrays Exceptions are not very clear.
     if not isinstance(ncfile, (str, Path)):
         raise TypeError(f"Invalid file path type: {type(ncfile)}.\n"
                         "Please provide string or pathlib Path")
@@ -46,15 +49,15 @@ def load_nc_file(ncfile: Union[str, Path],
     if not ncfile.suffix == ".nc":
         raise TypeError("Input file must have nc extension.")
 
-    if chunks:
-        return xr.open_dataset(ncfile, engine="netcdf4", chunks=chunks)
+    if parallel:
+        return xr.open_dataset(ncfile, engine="netcdf4", chunks=-1)
 
     return xr.open_dataset(ncfile, engine="netcdf4")
 
 def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
     """
-    If subset = None, it returns a list with all variables in the ds Dataset.
-    Else, if vars in subset are present in ds, it returns them, else it warns
+    If subset=None, it returns a list with all variable names in the ds.
+    If var names in subset are present in ds, it returns them, else it warns
     the user.
 
     Parameters
@@ -65,25 +68,33 @@ def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
     """
     # process subset list
     if not subset: # use all names
-        new_subset = [name for name  in ds.data_vars.keys()]
+        new_subset = [name for name in ds.data_vars.keys()]
     else:
         if not isinstance(subset, list) or \
              not all(map(lambda x: isinstance(x, str), subset)):
             raise TypeError("Subset argument must be a list of strings.")
 
-        new_subset = [name if name in ds.data_vars.keys()
-                      else warnings.warn(f"{name} not in Dataset.")
-                      for name in subset]
+        new_subset =[]
+        for name in subset:
+            if name in ds.data_vars.keys():
+                new_subset.append(name)
+            else:
+                warnings.warn(f"{name} not in Dataset.")
+
+        if not new_subset:
+            raise ValueError("None of the elements of the subset are present"
+                             "in the Dataset.")
     return new_subset
 
-def _time_dim_from_da(da: xr.DataArray,
-                      dims: list,
-                      coords: tuple
-                      ) -> tuple:
+def _index_da_by_coord_labels(da: xr.DataArray,
[…]
