### 2022-09-21 2cf356a
turn it into a class
