+++ pysd/tools/ncfiles.py
@@ -4,8 +4,6 @@ using PySD.
 """
 
 from csv import QUOTE_NONE
-from lib2to3.pgen2.token import OP
-from tkinter.ttk import Progressbar
 import xarray as xr
 import pandas as pd
 from pathlib import Path
@@ -13,13 +11,9 @@ import warnings
 from typing import Union, Optional
 import itertools
 
-
-def open_nc_file(ncfile: Union[str, Path],
-                 parallel: Optional[bool]=False) -> xr.Dataset:
+class NCFile():
     """
-    Loads netCDF file into xarray Dataset. It's basically a wrapper to
-    xr.open_dataset to simplify the interface for pysd use case (loading
-    simulation results).
+    Helper class to extract data from netCDF files.
 
     Parameters
     ----------
@@ -27,352 +21,389 @@ def open_nc_file(ncfile: Union[str, Path],
         Path to the netCDF file to process.
     parallel: bool
         When True, the Dataset is opened using chunks=-1 (see xarray
-        documentation for details), and when it's False, chunks=None.
-
-    Returns
-    -------
-    xarray.Dataset
-
-    """
-
-    # we run all these checks because xarrays Exceptions are not very clear.
-    if not isinstance(ncfile, (str, Path)):
-        raise TypeError(f"Invalid file path type: {type(ncfile)}.\n"
-                        "Please provide string or pathlib Path")
-
-    if isinstance(ncfile, str):
-        ncfile = Path(ncfile)
-
-    if not ncfile.is_file():
-        raise FileNotFoundError(f"{ncfile} could not be found.")
-
-    if not ncfile.suffix == ".nc":
-        raise TypeError("Input file must have nc extension.")
-
-    if parallel:
-        return xr.open_dataset(ncfile, engine="netcdf4", chunks=-1)
-
-    return xr.open_dataset(ncfile, engine="netcdf4")
-
-def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
-    """
-    If subset=None, it returns a list with all variable names in the ds.
-    If var names in subset are present in ds, it returns them, else it warns
-    the user.
-
-    Parameters
-    ----------
-    subset: list
-        Subset of variable names in the xarray Dataset.
+        documentation for details) and DataArrays are processed in parallel
+        using dask delayed. Dask is not included as a requirement for pysd,
+        hence it must be installed separately. Setting parallel=True is
+        highly recommended when the Dataset contains large multidimensional
+        DataArrays.
 
     """
-    # process subset list
-    if not subset: # use all names
-        new_subset = [name for name in ds.data_vars.keys()]
-    else:
-        if not isinstance(subset, list) or \
-             not all(map(lambda x: isinstance(x, str), subset)):
-            raise TypeError("Subset argument must be a list of strings.")
 
-        new_subset =[]
-        for name in subset:
-            if name in ds.data_vars.keys():
-                new_subset.append(name)
-            else:
-                warnings.warn(f"{name} not in Dataset.")
+    valid_export_file_types = [".csv", ".tab"]
 
-        if not new_subset:
-            raise ValueError("None of the elements of the subset are present"
-                             "in the Dataset.")
-    return new_subset
+    def __init__(self, filename: Union[str, Path],
+                 parallel: bool=False) -> None:
 
-def _index_da_by_coord_labels(da: xr.DataArray,
-                              dims: list,
-                              coords: tuple
-                              ) -> tuple:
-    """
-    Function to generate variable names, combining the actual name
-    with the coordinate names between brackets and separated by commas,
-    and to index the DataArray by the coordinate names specified in
-    the coords argument.
+        self.ncfile = NCFile._validate_nc_path(filename)
+        self.parallel = parallel
+        self.ds = self.open_nc()
 
[…]
