### 2022-09-21 #4329 @pdebuyl
Thanks @enekomartinmartinez for the updates.

@sixpearls could you assess the changes? From a first look, it seems that all of your requirements are met.

@enekomartinmartinez regarding the affiliations, I could not find specific instructions either. The footnote that you suggest is fine. If you wish that the affilication is included in the metadata, you can add it as a second affiliation though.
