+++ pysd/tools/ncfiles.py
@@ -2,14 +2,16 @@
 Tools for importing and converting netCDF files generated from simulations run
 using PySD.
 """
+import itertools
+import warnings
 
+from pathlib import Path
 from csv import QUOTE_NONE
+from typing import Union, Optional
+
 import xarray as xr
 import pandas as pd
-from pathlib import Path
-import warnings
-from typing import Union, Optional
-import itertools
+
 
 class NCFile():
     """
@@ -56,33 +58,20 @@ class NCFile():
             Whether time increases along row.
             Default is False.
 
-
         Returns
         -------
         df: pandas.DataFrame
             Dataframe with all colums specified in subset.
 
         """
-        if isinstance(outfile, str):
-            outfile = Path(outfile)
+        df = self.to_df(subset=subset)
 
-        out_fmt = outfile.suffix
-
-        if out_fmt not in NCFile.valid_export_file_types:
-            raise TypeError("Invalid output file format {out_fmt}\n"
-                            "Supported formats are csv and tab.")
-
-        outfile.parent. mkdir(parents=True, exist_ok=True)
-
-        df = self.to_df(subset=subset, time_in_row=time_in_row)
-
-        NCFile.df_to_csv(df, outfile)
+        NCFile.df_to_csv(df, outfile, time_in_row)
 
         return df
 
     def to_df(self,
               subset: Optional[list]=None,
-              time_in_row: Optional[bool]=False,
               ) -> pd.DataFrame:
         """
         Wrapper to ds_to_df static method. Convert xarray.Dataset into a
@@ -92,8 +81,6 @@ class NCFile():
         ----------
         subset: list
             List of variables to export from the Dataset.
-        time_in_row: bool
-            Whether time increases along row. Default is False.
 
         Returns
         -------
@@ -101,7 +88,7 @@ class NCFile():
             Dataframe with all colums specified in subset.
         """
 
-        return NCFile.ds_to_df(self.ds, subset, time_in_row, self.parallel)
+        return NCFile.ds_to_df(self.ds, subset, self.parallel)
 
     def open_nc(self) -> xr.Dataset:
         """
@@ -114,7 +101,6 @@ class NCFile():
         xarray.Dataset
 
         """
-
         if self.parallel:
             return xr.open_dataset(self.ncfile, engine="netcdf4", chunks=-1)
 
@@ -138,7 +124,7 @@ class NCFile():
             raise FileNotFoundError(f"{nc_path} could not be found.")
 
         if not nc_path.suffix == ".nc":
-            raise TypeError("Input file must have nc extension.")
+            raise ValueError("Input file must have nc extension.")
 
         return nc_path
 
@@ -146,8 +132,8 @@ class NCFile():
     def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
         """
         If subset=None, it returns a list with all variable names in the ds.
-        If var names in subset are present in ds, it returns them, else it warns
-        the user.
+        If var names in subset are present in ds, it returns them, else it
+        warns the user.
 
         Parameters
         ----------
@@ -155,8 +141,9 @@ class NCFile():
             Subset of variable names in the xarray Dataset.
 
         """
-        # process subset list
-        if not subset: # use all names
+        # use all variable names except for those of dimensions
+        # this excludes time, which is a variable too.
+        if not subset:
             new_subset = [name for name in ds.data_vars.keys()]
         else:
             if not isinstance(subset, list) or \
@@ -171,8 +158,13 @@ class NCFile():
                     warnings.warn(f"{name} not in Dataset.")
 
             if not new_subset:
-                raise ValueError("None of the elements of the subset are present"
-                                 "in the Dataset.")
[…]
