### 2022-09-24 7fb2345
add documentation for NCFile class
