### 2022-09-24 1cc7592
move ignore warnings to pytest.init
