### 2022-09-24 912f219
refractor to reduce code repetition
