+++ pysd/tools/ncfiles.py
@@ -107,124 +107,104 @@ class NCFile():
         return xr.open_dataset(self.ncfile, engine="netcdf4")
 
     @staticmethod
-    def _validate_nc_path(nc_path: Union[str, Path]) -> Path:
+    def ds_to_df(ds: xr.Dataset,
+                 subset: Optional[list]=None,
+                 parallel: Optional[bool]=False,
+                 index_dim: Optional[str]="time"
+                 ) -> pd.DataFrame:
         """
-        Checks validity of the nc_path passed by the user. We run these
-        checks because xarray Exceptions are not very explicit.
-        """
-
-        if not isinstance(nc_path, (str, Path)):
-            raise TypeError(f"Invalid file path type: {type(nc_path)}.\n"
-                            "Please provide string or pathlib Path")
-
-        if isinstance(nc_path, str):
-            nc_path = Path(nc_path)
-
-        if not nc_path.is_file():
-            raise FileNotFoundError(f"{nc_path} could not be found.")
-
-        if not nc_path.suffix == ".nc":
-            raise ValueError("Input file must have nc extension.")
-
-        return nc_path
-
-    @staticmethod
-    def _validate_ds_subset(ds: xr.Dataset, subset: list) -> list:
-        """
-        If subset=None, it returns a list with all variable names in the ds.
-        If var names in subset are present in ds, it returns them, else it
-        warns the user.
+        Convert xarray.Dataset into a pandas DataFrame.
 
         Parameters
         ----------
+        ds: xarray.Dataset
+            Dataset object.
         subset: list
-            Subset of variable names in the xarray Dataset.
+            List of variables to export from the Dataset.
+        parallel: bool
+            When True, DataArrays are processed in parallel using dask delayed.
+            Setting parallel=True is highly recommended when DataArrays are
+            large and multidimensional.
+        index_dim: str
+            Name of dimensions to use as index of the resulting DataFrame
+            (usually "time").
+
+        Returns
+        -------
+        df: pandas.DataFrame
+            Dataframe with all colums specified in subset.
 
         """
-        # use all variable names
-        if not subset:
-            new_subset = [name for name in ds.data_vars.keys()]
+        subset = NCFile._validate_ds_subset(ds, subset)
+
+        if parallel:
+            processing_func = NCFile.da_to_dict_delayed
         else:
-            if not isinstance(subset, list) or \
-                 not all(map(lambda x: isinstance(x, str), subset)):
-                raise TypeError("Subset argument must be a list of strings.")
+            processing_func = NCFile.da_to_dict
 
-            new_subset =[]
-            for name in subset:
-                if name in ds.data_vars.keys():
-                    new_subset.append(name)
-                else:
-                    warnings.warn(f"{name} not in Dataset.")
+        savedict = {}
 
-            if not new_subset:
-                raise ValueError("None of the elements of the subset are "
-                                 "present in the Dataset.")
+        for name in subset:
+            print(f"\nProcessing variable {name}.")
+            da = ds[name]
+            dims = da.dims
 
-        # adding time in the final subset
-        if "time" not in new_subset:
-            new_subset.append("time")
+            if not dims or dims == (index_dim,):
+                savedict.update({name: da.values.tolist()})
+            else:
+                savedict.update(processing_func(da, index_dim))
 
-        return new_subset
+        return NCFile.dict_to_df(savedict)
 
     @staticmethod
-    def _index_da_by_coord_labels(da: xr.DataArray, dims: list, coords: tuple
-                                  ) -> tuple:
+    def df_to_csv(df: pd.DataFrame, outfile: Path,
+                  time_in_row: Optional[bool]=False
+                  ) -> None:
         """
[…]
