### 2022-09-25 11a77c3
preload DataArray into memory for better perf
