### 2022-10-02 0f96d3d
Utilities to export netCDF data to csv or tab files (#369)
* preliminar implementation of nc_to_csv

* add new args

* prototype with parallel processing of DataArrays

* working functions

* turn it into a class

* unify df_to_csv

* adding tests

* add dask requirements

* capture Deprecation warnings

* small performance improvement

* add documentation for NCFile class

* move ignore warnings to pytest.init

* refractor to reduce code repetition

* preload DataArray into memory for better perf

* code revision

Co-authored-by: Roger Samso <roger.samso@bioentech.eu>
