+++ pysd/tools/ncfiles.py
@@ -1,0 +1,411 @@
+"""
+Tools for importing and converting netCDF files generated from simulations run
+using PySD.
+"""
+import itertools
+import warnings
+
+from pathlib import Path
+from csv import QUOTE_NONE
+from typing import Union, Optional
+
+import xarray as xr
+import pandas as pd
+
+
+class NCFile():
+    """
+    Helper class to extract data from netCDF files.
+
+    Parameters
+    ----------
+    ncfile: str or pathlib.Path
+        Path to the netCDF file to process.
+    parallel: bool (optional)
+        When True, the Dataset is opened using chunks=-1 (see xarray
+        documentation for details) and DataArrays are processed in parallel
+        using dask delayed. Dask is not included as a requirement for pysd,
+        hence it must be installed separately. Setting parallel=True is
+        highly recommended when the Dataset contains large multidimensional
+        DataArrays.
+
+    """
+
+    valid_export_file_types = [".csv", ".tab"]
+
+    def __init__(self,
+                 filename: Union[str, Path],
+                 parallel: Optional[bool]=False) -> None:
+
+        self.ncfile = NCFile._validate_nc_path(filename)
+        self.parallel = parallel
+        self.ds = self.open_nc()
+
+    def to_text_file(self,
+                     outfile: Optional[Union[str, Path]]="result.tab",
+                     subset: Optional[list]=None,
+                     time_in_row: Optional[bool]=False,
+                     ) -> pd.DataFrame:
+        """
+        Convert netCDF file contents into comma separated or tab delimited
+        file.
+
+        Parameters
+        ----------
+        outfile: str or pathlib.Path (optional)
+            Path to the output file.
+        subset: list (optional)
+            List of variables to export from the netCDF.
+        time_in_row: bool (optional)
+            Whether time increases along row.
+            Default is False.
+
+        Returns
+        -------
+        df: pandas.DataFrame
+            Dataframe with all colums specified in subset.
+
+        """
+        df = self.to_df(subset=subset)
+
+        NCFile.df_to_text_file(df, outfile, time_in_row)
+
+        return df
+
+    def to_df(self,
+              subset: Optional[list]=None,
+              ) -> pd.DataFrame:
+        """
+        Wrapper to ds_to_df static method. Convert xarray.Dataset into a
+        pandas DataFrame.
+
+        Parameters
+        ----------
+        subset: list (optional)
+            List of variables to export from the Dataset.
+
+        Returns
+        -------
+        df: pandas.DataFrame
+            Dataframe with all colums specified in subset.
+        """
+
+        return NCFile.ds_to_df(self.ds, subset, self.parallel)
+
+    def open_nc(self) -> xr.Dataset:
+        """
+        Loads netCDF file into xarray Dataset. It's basically a wrapper to
+        xr.open_dataset to simplify the interface for pysd use case (loading
+        simulation results).
+
+        Returns
+        -------
+        xarray.Dataset
+
+        """
+        if self.parallel:
+            return xr.open_dataset(self.ncfile, engine="netcdf4", chunks=-1)
+
+        return xr.open_dataset(self.ncfile, engine="netcdf4")
+
+    @staticmethod
+    def ds_to_df(ds: xr.Dataset,
+                 subset: Optional[list]=None,
+                 parallel: Optional[bool]=False,
+                 index_dim: Optional[str]="time"
+                 ) -> pd.DataFrame:
+        """
+        Convert xarray.Dataset into a pandas DataFrame.
+
+        Parameters
+        ----------
+        ds: xarray.Dataset
+            Dataset object.
+        subset: list (optional)
+            List of variables to export from the Dataset.
+        parallel: bool (optional)
+            When True, DataArrays are processed in parallel using dask delayed.
[…]
