+++ tests/pytest_tools/pytest_ncfile.py
@@ -1,0 +1,212 @@
+import pytest
+
+from pathlib import Path
+
+import xarray as xr
+import numpy as np
+import pandas as pd
+
+from pysd.tools.ncfiles import NCFile
+
+@pytest.fixture(scope="session")
+def sample_dataset():
+    # generation of data (slight modification from xarray.Dataset example)
+    np.random.seed(0)
+    temperature = 15 + 8 * np.random.randn(3, 2, 2)
+    precipitation = 10 * np.random.rand(3, 2, 2)
+    altitude = 1000 * np.random.rand(2, 2)
+    lon = [[-99.83, -99.32], [-99.79, -99.23]]
+    lat = [[42.25, 42.21], [42.63, 42.59]]
+    time = pd.date_range("2014-09-06", periods=3)
+
+    ds = xr.Dataset(
+        data_vars=dict(
+            temperature=(
+                ["time", "x", "y"], temperature),
+            precipitation=(
+                ["time", "x", "y"], precipitation),
+            altitude=(["x", "y"], altitude),
+            non_dimensional_var=([], np.array(5)),
+            time=(["time"], time)
+        ),
+        coords=dict(
+            lon=(["x", "y"], lon),
+            lat=(["x", "y"], lat),
+        ),
+        attrs=dict(description="Weather related data."),
+    )
+
+    return ds
+
+@pytest.fixture(scope="session")
+def nc_file(sample_dataset, tmp_path_factory):
+
+    path = tmp_path_factory.mktemp("data") / "dataset.nc"
+
+    sample_dataset.to_netcdf(
+        path, mode="w", format="NETCDF4", engine="netcdf4")
+
+    return path
+
+
+class TestNCFile():
+
+    @pytest.mark.parametrize(
+        "file_path,raises,expect",
+        [(1, True, TypeError),
+        ("unexisting_file.nc", True, FileNotFoundError),
+        (Path(__file__), True, ValueError),
+        ("nc_file", False, "nc_file")
+        ],
+        ids=["Wrong type", "Not found", "Wrong extension", "Correct path"]
+        )
+    def test__validate_nc_path_errors(
+         self, request, file_path, raises, expect):
+        if not raises:
+            nc_file = request.getfixturevalue(file_path)
+            assert NCFile._validate_nc_path(nc_file) == nc_file
+        else:
+            with pytest.raises(expect):
+                NCFile._validate_nc_path(file_path)
+
+    @pytest.mark.parametrize(
+        "subset,raises,expected",
+        [
+            ([], False, [
+                            "temperature",
+                            "precipitation",
+                            "altitude",
+                            "non_dimensional_var",
+                            "time"
+                        ]
+            ),
+            (["temperature"], False, ["temperature", "time"]),
+            (["non_existing"], True, ValueError),
+            ("I am not a list", True, TypeError)
+        ],
+        ids=["empty", "one var", "non existing single var", "wrong type"]
+        )
+    @pytest.mark.filterwarnings("ignore")
+    def test__validate_ds_subset(
+         self, sample_dataset, subset, raises, expected):
+
+        if raises:
+            with pytest.raises(expected):
+                NCFile._validate_ds_subset(sample_dataset, subset)
+        else:
+            assert NCFile._validate_ds_subset(
+                sample_dataset, subset) == expected
+
+    @pytest.mark.parametrize(
+        "data_var,dims,coords,expected",
+        [
+            ("temperature", ["x", "y"], (0,0), ("temperature[0,0]", np.array(
+                [29.11241876774131 ,29.94046392119974, 14.174249185651536]))
+            ),
[…]
