### 2022-10-02 SDXorg/pysd#348 @JamesPHoughton
Weird, I set it up from scratch in the new repo. It’s still broken? I’ll
look again…

On Fri, Sep 30, 2022 at 4:07 PM Eneko Martin-Martinez <
***@***.***> wrote:

> @JamesPHoughton <https://github.com/JamesPHoughton> we still need to
> correct the RTD webhook. I do not have access to the RTD project!
>
