### 2022-10-03 SDXorg/pysd#348 @enekomartinmartinez
Seems that in the repo is set correctly as GitHub it says so. But probably we are missing something in the RTD project as RTD is not building the docs in PRs.

Note that the RTD for the cookbook was not updated after the last changes.
