### 2022-10-03 #4329 @pdebuyl
Hello,

Can you confirm that v3.7.1 is the suitable release? If so, can you archive it on zenodo and mention the link to the zenodo archive here? You can use a rolling DOI with zenodo if you wish but the one given here should point to the reviewed version.

I had a look at the pdf and have no further editorial comments on the paper itself.
