### 2022-10-03 #4329 @enekomartinmartinez
Hi,

Thank you!

Yes, v3.7.1 is the suitable release. I think the zenodo account permissions belong to @JamesPHoughton. However, DOIs are automatically generated for all releases. Is it enough or do we need to do something else?

10.5281/zenodo.7094483
