### 2022-10-05 a9ef60d
implementing externals serialization for faster model initialization
