+++ pysd/py_backend/model.py
@@ -5,6 +5,7 @@ Several methods and propierties are inherited from Macro class, which
 allows integrating a model or a Macro expression (set of functions in
 a separate file).
 """
+import time
 import warnings
 import inspect
 import pickle
@@ -19,7 +20,8 @@ from pysd._version import __version__
 
 from . import utils
 from .statefuls import DynamicStateful, Stateful
-from .external import External, Excels
+from .external import External, Excels, ExtLookup, ExtData
+
 from .cache import Cache, constant_cache
 from .data import TabData
 from .lookups import HardcodedLookups
@@ -57,8 +59,8 @@ class Macro(DynamicStateful):
         Time to set at the begginning of the Macro. Default is None.
     data_files: dict or list or str or None
         The dictionary with keys the name of file and variables to
-        load the data from there. Or the list of names or name of the
-        file to search the data in. Only works for TabData type object
+        load the data from. Or the list of names or name of the file
+        to search the data in. Only works for TabData type object
         and it is neccessary to provide it. Default is None.
     py_name: str or None
         The name of the Macro object. Default is None.
@@ -453,13 +455,7 @@ class Macro(DynamicStateful):
 
         if not self.external_loaded:
             # Initialize external elements
-            for element in self._external_elements:
-                element.initialize()
-
-            # Remove Excel data from memory
-            Excels.clean()
-
-            self.external_loaded = True
+            self.initialize_external_data()
 
         # Initialize stateful objects
         for element_name in self.initialize_order:
@@ -531,6 +527,297 @@ class Macro(DynamicStateful):
 
         self.set_stateful(stateful_dict)
         self.time.set_control_vars(initial_time=time)
+
+    def initialize_external_data(self, externals=None):
+
+        """
+        Initializes external data.
+
+        If a path to a netCDF file containing serialized values of some or
+        all of the model external data is passed in the external argument,
+        those will be loaded from the file.
+
+        To get the full performance gain of loading the externals from a netCDF
+        file, the model should be loaded with initialize=False first.
+
+        Parameters
+        ----------
+        externals: str or pathlib.Path (optional)
+            Path to the netCDF file that contains the model external objects.
+
+        Returns
+        -------
+        None
+
+        """
+
+        if not externals:
+            for ext in self._external_elements:
+                ext.initialize()
+
+            # Remove Excel data from memory
+            Excels.clean()
+
+            self.external_loaded = True
+
+            return
+
+        externals = Path(externals)
+
+        if not externals.is_file():
+            raise FileNotFoundError(f"Invalid file path ({str(externals)})")
+
+        ds = xr.open_dataset(externals)
+
+        for ext in self._external_elements:
+            py_name = ext.py_name
+
+            if py_name in ds.data_vars.keys():
+                da = ds.data_vars[py_name]
+                dimensions = ext.final_coords
+
+                if isinstance(ext, ExtData):
+                    time_dim = [dim for dim in da.dims if dim.startswith(
+                        "time_#")][0]
+                    da = da.rename({time_dim: "time"})
+                    da = da.loc[dimensions]
+                elif isinstance(ext, ExtLookup):
+                    lookup_dim = [dim for dim in da.dims if dim.startswith(
+                        "lookup_dim_#")][0]
+                    da = da.rename({lookup_dim: "lookup_dim"})
+                    da = da.loc[dimensions]
+                else: #  ExtConstant
+                    if dimensions:
+                        da = da.loc[dimensions]
+
+                if da.dims:
+                    ext.data = da
+                else:
+                    ext.data = float(da.data)
+
+            else:
+                ext.initialize()
+
+        Excels.clean()
+
+        self.external_loaded = True
+
+    def serialize_externals(self, export_path="externals.nc",
+                            include_externals="all", exclude_externals=None):
+        """
[…]
