### 2022-10-05 5f224ae
implementing externals serialization for faster model initialization
