### 2022-10-05 eec7e25
implementing externals serialization for faster model initialization
