+++ tests/pytest_pysd/pytest_serialize_externals.py
@@ -1,0 +1,273 @@
+import re
+import pytest
+
+from pathlib import Path
+import xarray as xr
+import numpy as np
+
+import pysd
+
+data_ext = Path(
+    "test-models/tests/get_data_args_3d_xls/test_get_data_args_3d_xls.mdl")
+constant_ext = Path("test-models/tests/get_constants_subranges/test_get_constants_subranges.mdl")
+lookup_ext = Path("test-models/tests/get_lookups_data_3d_xls/test_get_lookups_data_3d_xls.mdl")
+lookup_subranges = Path("test-models/tests/get_lookups_subset/test_get_lookups_subset.mdl")
+mixed_definitions =Path("test-models/tests/get_mixed_definitions/test_get_mixed_definitions.mdl")
+subscript_ext = Path("test-models/tests/get_subscript_3d_arrays_xls/test_get_subscript_3d_arrays_xls.mdl")
+multiple_excels = Path("more-tests/externals_multiple_files/test_externals_multiple_files.mdl")
+
+class TestSerialization():
+
+    @pytest.mark.parametrize("model_path", [data_ext])
+    def test_serialize_externals_file_content(self, tmp_path, model):
+
+        result_path = tmp_path / "externals.nc"
+
+        model.serialize_externals(export_path=result_path,
+                                  include_externals=["input.xls"],
+                                  exclude_externals=["data_backward"])
+
+        ds = xr.open_dataset(result_path)
+
+        assert "_ext_data_data_backward" not in ds.data_vars.keys()
+        assert "_ext_data_data_forward" in ds.data_vars.keys()
+        assert np.allclose(
+            ds.data_vars["_ext_data_data_forward"].loc[
+                {"time_#1": 10.0}].data,
+                np.array([[10., 5.], [-5., 1.]]))
+        assert np.allclose(ds.coords["time_#1"].values,
+                           np.array([ 0.,  5., 10., 15., 20., 25., 30.]))
+        assert "description" in ds.attrs.keys()
+        assert all(map(
+            lambda x: x in ds.data_vars["_ext_data_data_forward"].attrs.keys(),
+            ["Py Name", "sheets", "files"]))
+
+    @pytest.mark.parametrize(
+        "model_path,include,exclude",
+        [(data_ext, ["data_forward"], None),
+         (data_ext, "all", ["data_backward"]),
+         (lookup_ext, ["lookup_function"], None),
+         (lookup_subranges, "all", None),
+         (constant_ext,["constant"], None),
+         ],
+        ids=["data", "data_exclude", "lookup", "lookup_subranges", "constant"])
+    def test_serialize_externals_different_types(
+        self, model, include, exclude):
+
+        # external subscripts are not included because they are hardcoded in
+        # the Python model.
+        model_path = Path(model.py_model_file)
+        externals_path = model_path.with_suffix(".nc")
+
+        if include == "all":
+            if exclude:
+                varnames_included = {
+                    ext.py_name: "_".join(ext.py_name.split("_")[3:]) for \
+                        ext in model._external_elements if \
+                            not "_".join(ext.py_name.split("_")[3:]) in exclude
+                            }
+            else:
+                varnames_included = {
+                    ext.py_name: "_".join(ext.py_name.split("_")[3:]) for \
+                        ext in model._external_elements}
+        else:
+            if exclude:
+                varnames_included = {
+                    model._dependencies[var]["__external__"]: var for \
+                        var in include if var not in exclude}
+            else:
+                varnames_included = {
+                    model._dependencies[var]["__external__"]: var for \
+                         var in include}
+
+
+        model.serialize_externals(export_path=externals_path,
+                                  include_externals=include,
+                                  exclude_externals=exclude)
+
+        ds = xr.open_dataset(externals_path)
+
+        # all expected variables are in the dataset
[…]
