### 2022-10-05 8bfbad0
update optional dependencies in installation.rst
