### 2022-10-08 #4329 @Kevin-Mattheus-Moerman
@rogersamso

In relation to the archive:
- [ ] Please manually edit the ZENODO archive listing so that the meta-data matches that of the paper. In particular fix the title and the author list. Furthermore you can ensure the correct license is listed and may opt to add the author ORCID fields.

On the paper
- [ ] Please add city and country information to all affiliations (do not use acronyms for countries)
