### 2022-10-11 c67345e
adding documentation and updating whats_new
