### 2022-10-11 a53e70b
adding documentation and updating whats_new
