+++ pysd/py_backend/utils.py
@@ -21,16 +21,20 @@ class UniqueDims():
     same dimension name but different coords in xarray Datasets.
     """
     def __init__(self, original_dim_name):
-        self.dim_prefix = original_dim_name + "_#"
+        self.dim_name = original_dim_name
+        self.dim_prefix = self.dim_name + "_#"
         self.unique_dims = []
         self.num = 1
 
-    def name_new_dim(self, coords):
+    def name_new_dim(self, dim_name, coords):
         """
         Returns either a new name (original_dim_name + _# + num) if the coords
         list is not in unique_dims, or the preexisting dimension name if it is.
         Parameters
         ----------
+        dim_name: str
+            This argument is used to verify that we are passing the right
+            dimension name to the class.
         coords: list
             List of coordinates of a dimension.
 
@@ -38,6 +42,10 @@ class UniqueDims():
         -------
         Updated name of the original dimension.
         """
+        if dim_name != self.dim_name:
+            raise ValueError("This object is configured to process dimension "\
+                + f"{self.dim_name} and it's being passed a dimension named "\
+                + f"{dim_name}")
         if self.is_new(coords):
             new_dim_name = self.dim_prefix + str(self.num)
             self.unique_dims.append((new_dim_name, coords))
