+++ tests/pytest_pysd/pytest_serialize_externals.py
@@ -1,4 +1,5 @@
 import re
+from xml.dom import UserDataHandler
 import pytest
 
 from pathlib import Path
@@ -9,12 +10,20 @@ import pysd
 
 data_ext = Path(
     "test-models/tests/get_data_args_3d_xls/test_get_data_args_3d_xls.mdl")
-constant_ext = Path("test-models/tests/get_constants_subranges/test_get_constants_subranges.mdl")
-lookup_ext = Path("test-models/tests/get_lookups_data_3d_xls/test_get_lookups_data_3d_xls.mdl")
-lookup_subranges = Path("test-models/tests/get_lookups_subset/test_get_lookups_subset.mdl")
-mixed_definitions =Path("test-models/tests/get_mixed_definitions/test_get_mixed_definitions.mdl")
-subscript_ext = Path("test-models/tests/get_subscript_3d_arrays_xls/test_get_subscript_3d_arrays_xls.mdl")
-multiple_excels = Path("more-tests/externals_multiple_files/test_externals_multiple_files.mdl")
+constant_ext = Path("test-models/tests/get_constants_subranges/"
+                    "test_get_constants_subranges.mdl")
+lookup_ext = Path("test-models/tests/get_lookups_data_3d_xls/"
+                  "test_get_lookups_data_3d_xls.mdl")
+lookup_subranges = Path("test-models/tests/get_lookups_subset/"
+                        "test_get_lookups_subset.mdl")
+mixed_definitions =Path("test-models/tests/get_mixed_definitions/"
+                        "test_get_mixed_definitions.mdl")
+subscript_ext = Path("test-models/tests/get_subscript_3d_arrays_xls/"
+                     "test_get_subscript_3d_arrays_xls.mdl")
+multiple_excels = Path("more-tests/externals_multiple_files/"
+                       "test_externals_multiple_files.mdl")
+incomplete_constant_def = Path("more-tests/incomplete_dims/" \
+                               "get_incomplete_dims.mdl")
 
 class TestSerialization():
 
@@ -34,9 +43,10 @@ class TestSerialization():
         assert np.allclose(
             ds.data_vars["_ext_data_data_forward"].loc[
                 {"time_#1": 10.0}].data,
-                np.array([[10., 5.], [-5., 1.]]))
+                np.array([[10., 5.], [-5., 1.]]), equal_nan=True)
         assert np.allclose(ds.coords["time_#1"].values,
-                           np.array([ 0.,  5., 10., 15., 20., 25., 30.]))
+                           np.array([ 0.,  5., 10., 15., 20., 25., 30.]),
+                           equal_nan=True)
         assert "description" in ds.attrs.keys()
         assert all(map(
             lambda x: x in ds.data_vars["_ext_data_data_forward"].attrs.keys(),
@@ -80,7 +90,6 @@ class TestSerialization():
                     model._dependencies[var]["__external__"]: var for \
                          var in include}
 
-
         model.serialize_externals(export_path=externals_path,
                                   include_externals=include,
                                   exclude_externals=exclude)
@@ -96,8 +105,8 @@ class TestSerialization():
         # the values in the dataset are the same than in the model
         assert all(np.allclose(
             ds.data_vars[py_name].values,
-            getattr(model.components, py_name).data.values) for \
-                py_name in varnames_included.keys())
+            getattr(model.components, py_name).data.values, equal_nan=True) \
+                for py_name in varnames_included.keys())
 
         model = pysd.load(model_path, initialize=False)
         model.initialize_external_data(externals=externals_path)
@@ -106,8 +115,8 @@ class TestSerialization():
         # reinitialized
         assert all(np.allclose(
             ds.data_vars[py_name].values,
[…]
