+++ tests/pytest_pysd/pytest_serialize_externals.py
@@ -1,6 +1,3 @@
-import re
-from xml.dom import UserDataHandler
-from xml.etree.ElementInclude import include
 import pytest
 
 from pathlib import Path
@@ -17,14 +14,15 @@ lookup_ext = Path("test-models/tests/get_lookups_data_3d_xls/"
                   "test_get_lookups_data_3d_xls.mdl")
 lookup_subranges = Path("test-models/tests/get_lookups_subset/"
                         "test_get_lookups_subset.mdl")
-mixed_definitions =Path("test-models/tests/get_mixed_definitions/"
-                        "test_get_mixed_definitions.mdl")
+mixed_definitions = Path("test-models/tests/get_mixed_definitions/"
+                         "test_get_mixed_definitions.mdl")
 subscript_ext = Path("test-models/tests/get_subscript_3d_arrays_xls/"
                      "test_get_subscript_3d_arrays_xls.mdl")
 multiple_excels = Path("more-tests/externals_multiple_files/"
                        "test_externals_multiple_files.mdl")
-incomplete_constant_def = Path("more-tests/incomplete_dims/" \
+incomplete_constant_def = Path("more-tests/incomplete_dims/"
                                "get_incomplete_dims.mdl")
+
 
 class TestSerialization():
 
@@ -44,9 +42,9 @@ class TestSerialization():
         assert np.allclose(
             ds.data_vars["_ext_data_data_forward"].loc[
                 {"time_#1": 10.0}].data,
-                np.array([[10., 5.], [-5., 1.]]), equal_nan=True)
+            np.array([[10., 5.], [-5., 1.]]), equal_nan=True)
         assert np.allclose(ds.coords["time_#1"].values,
-                           np.array([ 0.,  5., 10., 15., 20., 25., 30.]),
+                           np.array([0.,  5., 10., 15., 20., 25., 30.]),
                            equal_nan=True)
         assert "description" in ds.attrs.keys()
         assert all(map(
@@ -55,15 +53,16 @@ class TestSerialization():
 
     @pytest.mark.parametrize(
         "model_path,include,exclude",
-        [(data_ext, ["data_forward"], None),
-         (data_ext, "all", ["data_backward"]),
-         (lookup_ext, ["lookup_function"], None),
-         (lookup_subranges, "all", None),
-         (constant_ext,["constant"], None),
-         ],
+        [
+            (data_ext, ["data_forward"], None),
+            (data_ext, "all", ["data_backward"]),
+            (lookup_ext, ["lookup_function"], None),
+            (lookup_subranges, "all", None),
+            (constant_ext, ["constant"], None),
+        ],
         ids=["data", "data_exclude", "lookup", "lookup_subranges", "constant"])
-    def test_serialize_externals_different_types(
-        self, model, include, exclude):
+    def test_serialize_externals_different_types(self, model,
+                                                 include, exclude):
 
         # external subscripts are not included because they are hardcoded in
         # the Python model.
@@ -73,23 +72,26 @@ class TestSerialization():
         if include == "all":
             if exclude:
                 varnames_included = {
-                    ext.py_name: "_".join(ext.py_name.split("_")[3:]) for \
-                        ext in model._external_elements if \
-                            not "_".join(ext.py_name.split("_")[3:]) in exclude
-                            }
+                    ext.py_name: "_".join(ext.py_name.split("_")[3:])
+                    for ext in model._external_elements
+                    if not "_".join(ext.py_name.split("_")[3:]) in exclude
+                }
             else:
                 varnames_included = {
-                    ext.py_name: "_".join(ext.py_name.split("_")[3:]) for \
-                        ext in model._external_elements}
+                    ext.py_name: "_".join(ext.py_name.split("_")[3:])
+                    for ext in model._external_elements
+                }
         else:
             if exclude:
                 varnames_included = {
[…]
