### 2022-10-15 6b9e2e5
Simplify the externals serialization
