+++ pysd/py_backend/model.py
@@ -586,7 +586,7 @@ class Macro(DynamicStateful):
                         "lookup_dim_#")][0]
                     da = da.rename({lookup_dim: "lookup_dim"})
                     da = da.loc[dimensions]
-                else: #  ExtConstant
+                else:  # ExtConstant
                     if dimensions:
                         da = da.loc[dimensions]
 
@@ -652,60 +652,63 @@ class Macro(DynamicStateful):
         if not include_externals:
             raise ValueError("include_externals argument must not be None.")
 
+        # TODO include also checking the original name
+        externals_dict = {
+            "py_name": [], "py_var_name": [], "file": [], "ext": []}
+        for ext in self._external_elements:
+            externals_dict["py_name"].append(ext.py_name)
+            externals_dict["py_var_name"].append(
+                self.__get_varname_from_ext_name(ext.py_name))
+            externals_dict["file"].append(set(ext.files))
+            externals_dict["ext"].append(ext)
+        exts_df = pd.DataFrame(data=externals_dict)
+
         if include_externals != "all":
-            if not isinstance(include_externals, list):
-                raise TypeError("include_externals must be 'all', or a list.")
+            if not isinstance(include_externals, (list, set)):
+                raise TypeError(
+                    "include_externals must be 'all', or a list, or a set.")
+
+            # subset only to the externals to include
+            exts_df = exts_df[[
+                name in include_externals
+                or var_name in include_externals
+                or bool(file.intersection(include_externals))
+                for name, var_name, file
+                in zip(
+                    externals_dict["py_name"],
+                    externals_dict["py_var_name"],
+                    externals_dict["file"]
+                )
+            ]]
 
         if exclude_externals:
-            if not isinstance(exclude_externals, list):
-                raise TypeError("exclude_externals must be a list")
+            if not isinstance(exclude_externals, (list, set)):
+                raise TypeError("exclude_externals must be a list or a set.")
 
-        if include_externals == "all" and not exclude_externals:
-            for ext in self._external_elements:
-                real_py_name = self.__get_varname_from_ext_name(ext.py_name)
+            # subset only to the externals to include
+            exts_df = exts_df[[
+                name not in exclude_externals
+                and var_name not in exclude_externals
+                and not bool(file.intersection(exclude_externals))
+                for name, var_name, file
+                in zip(
+                    externals_dict["py_name"],
+                    externals_dict["py_var_name"],
+                    externals_dict["file"]
+                )
+            ]]
 
-                self.__include_for_serialization(
-                    ext, real_py_name, data, metadata, lookup_dims, data_dims)
-
-        elif include_externals == "all" and exclude_externals:
-            for ext in self._external_elements:
-                real_py_name = self.__get_varname_from_ext_name(ext.py_name)
-                if real_py_name not in exclude_externals and \
-                     not any(
-                        map(lambda x: str(x) in exclude_externals, ext.files)):
-
-                    self.__include_for_serialization(ext, real_py_name, data,
-                                                     metadata, lookup_dims,
-                                                     data_dims)
-
-        elif include_externals != "all" and not exclude_externals:
-            for ext in self._external_elements:
-                real_py_name = self.__get_varname_from_ext_name(ext.py_name)
-                if real_py_name in include_externals or \
-                     any(
-                        map(lambda x: str(x) in include_externals, ext.files)):
-
-                    self.__include_for_serialization(ext, real_py_name, data,
-                                                     metadata, lookup_dims,
-                                                     data_dims)
-
-        else:
-            for ext in self._external_elements:
[…]
