### 2022-10-15 a319f75
Clean model and utils
