### 2022-10-15 5347a51
Clean model and utils
