### 2022-10-15 db95b6b
Remove warning and unnecessary assertions
