### 2022-10-15 21d8aa9
Include all dimensions for all externals
