### 2022-10-15 38c9604
Include all dimensions for all externals
