### 2022-10-17 #4329 @kyleniemeyer
Congratulations @rogersamso @enekomartinmartinez and @JamesPHoughton on your article's publication in JOSS!
