### 2022-10-17 #4329 @rogersamso
Thanks @kyleniemeyer @pdebuyl @blsqr and @sixpearls ! Publishing in JOSS was such an awesome experience! Thank you all for helping us improve PySD!
